### 2021-02-11 #3031 @whedon
[PRE REVIEW]: The MOOSE geochemistry module
**Submitting author:** @WilkAndy (<a href="http://orcid.org/0000-0001-6472-9560">Andy Wilkins</a>)
**Repository:** <a href="https://github.com/idaholab/moose" target ="_blank">https://github.com/idaholab/moose</a>
**Version:** snapshot-20-10-27
**Reviewers:** smolins
