### 2021-02-11 ece8075
Set subdomain ID in LineMaterialSamplerBase
Closes #16959
