### 2021-02-11 edbd75a
Clean up existing hierarchy
