+++ modules/tensor_mechanics/doc/content/source/materials/ADTransverselyIsotropicPlasticityStressUpdate.md
@@ -10,10 +10,10 @@ The Hill yield function can be defined as
 \begin{equation}
 f = \frac{1}{2} \boldsymbol{s}^{T} \boldsymbol{A} \boldsymbol{s}^{T} - s_{y}(\alpha)
 \end{equation}
-where $\boldsymbol{s}$ is the stress tensor, $\boldsymbol{A}$ is the anisotropy (Hill) tensor, and $\alpha$ is an internal parameter that can be used, for example, to prescribe strain hardening through a plasticity modulus. Note that the Hill tensor is defined as a six by six matrix using the following unitless constants: $F$, $G$, $H$, $L$, $M$ and $N$.
+where $\boldsymbol{s}$ is the deviatoric stress tensor in vector form, $\boldsymbol{A}$ is the anisotropy (Hill) tensor, and $\alpha$ is an internal parameter that can be used, for example, to prescribe strain hardening through a plasticity modulus. Note that the Hill tensor is defined as a six by six matrix using the following unitless constants: $F$, $G$, $H$, $L$, $M$ and $N$.
 
 !alert warning
-The combination of elastic anisotropy and plastic anisotropy requires further tensor projections and is scheduled to be implemented soon.
+The combination of elastic anisotropy and plastic anisotropy requires further tensor projections. Use `ADTransverselyIsotropicElastoPlasticityStressUpdate` for that application.
 
 The effective plastic strain increment is obtained within the framework of a generalized (Hill plasticity) radial return mapping, see
 [ADGeneralizedRadialReturnStressUpdate](/ADGeneralizedRadialReturnStressUpdate.md).
