### 2021-02-11 e4711a4
Initial implementation/verification of elastoplastic anisotropy
