+++ modules/tensor_mechanics/src/materials/ADTransverselyIsotropicElastoPlasticityStressUpdate.C
@@ -1,0 +1,470 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#include "ADTransverselyIsotropicElastoPlasticityStressUpdate.h"
+
+registerMooseObject("TensorMechanicsApp", ADTransverselyIsotropicElastoPlasticityStressUpdate);
+
+InputParameters
+ADTransverselyIsotropicElastoPlasticityStressUpdate::validParams()
+{
+  InputParameters params = ADAnisotropicReturnPlasticityStressUpdateBase::validParams();
+  params.addClassDescription(
+      "This class uses the stress update material in a radial return isotropic power law creep "
+      "model.  This class can be used in conjunction with other creep and plasticity materials for "
+      "more complex simulations.");
+
+  params.addRequiredParam<Real>("hardening_constant",
+                                "Hardening constant (H) for anisotropic plasticity");
+  params.addRequiredParam<Real>("yield_stress",
+                                "Yield stress (constant value) for anisotropic plasticity");
+
+  params.addRequiredRangeCheckedParam<std::vector<Real>>("hill_constants",
+                                                         "hill_constants_size = 6",
+                                                         "Hill material constants in order: F, "
+                                                         "G, H, L, M, N");
+
+  return params;
+}
+
+ADTransverselyIsotropicElastoPlasticityStressUpdate::
+    ADTransverselyIsotropicElastoPlasticityStressUpdate(const InputParameters & parameters)
+  : ADAnisotropicReturnPlasticityStressUpdateBase(parameters),
+    _hill_constants(6),
+    _qsigma(0.0),
+    _eigenvalues_hill(6),
+    _eigenvectors_hill(6, 6),
+    _anisotropic_elastic_tensor(6, 6),
+    _elasticity_tensor_name(_base_name + "elasticity_tensor"),
+    _elasticity_tensor(getADMaterialProperty<RankFourTensor>(_elasticity_tensor_name)),
+    _hardening_constant(getParam<Real>("hardening_constant")),
+    _hardening_function(isParamValid("hardening_function") ? &getFunction("hardening_function")
+                                                           : nullptr),
+    _hardening_variable(declareADProperty<Real>(_base_name + "hardening_variable")),
+    _hardening_variable_old(getMaterialPropertyOld<Real>(_base_name + "hardening_variable")),
+    _elasticity_eigenvectors(
+        declareADProperty<DenseMatrix<Real>>(_base_name + "elasticity_eigenvectors")),
+    _elasticity_eigenvalues(
[…]
