+++ modules/tensor_mechanics/include/materials/ADTransverselyIsotropicElastoPlasticityStressUpdate.h
@@ -1,0 +1,129 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#pragma once
+
+#include "ADAnisotropicReturnPlasticityStressUpdateBase.h"
+
+/**
+ * This class uses the stress update material in an anisotropic return isotropic creep
+ * model.  This class is one of the basic radial return constitutive models; more complex
+ * constitutive models combine creep and plasticity.
+ *
+ */
+
+class ADTransverselyIsotropicElastoPlasticityStressUpdate
+  : public ADAnisotropicReturnPlasticityStressUpdateBase
+{
+public:
+  static InputParameters validParams();
+
+  ADTransverselyIsotropicElastoPlasticityStressUpdate(const InputParameters & parameters);
+
+  virtual Real
+  computeStrainEnergyRateDensity(const ADMaterialProperty<RankTwoTensor> & stress,
+                                 const ADMaterialProperty<RankTwoTensor> & strain_rate) override;
+
+protected:
+  virtual void computeStressInitialize(const ADDenseVector & stress_dev,
+                                       const ADDenseVector & stress,
+                                       const ADRankFourTensor & elasticity_tensor) override;
+
+  /**
+   * Compute eigendecomposition of element-wise elasticity tensor
+   */
+  void computeElasticityTensorEigenDecomposition();
+
+  virtual ADReal computeResidual(const ADDenseVector & effective_trial_stress,
+                                 const ADDenseVector & stress_new,
+                                 const ADReal & scalar) override;
+  virtual ADReal computeDerivative(const ADDenseVector & effective_trial_stress,
+                                   const ADDenseVector & stress_new,
+                                   const ADReal & scalar) override;
+
+  virtual Real computeReferenceResidual(const ADDenseVector & effective_trial_stress,
+                                        const ADDenseVector & stress_new,
+                                        const ADReal & residual,
+                                        const ADReal & scalar_effective_inelastic_strain) override;
+  virtual void propagateQpStatefulProperties() override;
+  /**
+   * Does the model require the elasticity tensor to be isotropic? No, this class does
+   * anisotropic *elasto-plasticity*
+   */
+  bool requiresIsotropicTensor() override { return false; }
+
+  ADReal computeHardeningDerivative();
+  ADReal computeHardeningValue(const ADReal & scalar, const ADReal & omega);
+  /**
+   * Compute eigendecomposition of Hill's tensor for anisotropic plasticity
+   * @param hill_tensor 6x6 matrix representing fourth order Hill's tensor describing anisotropy
+   */
+  void computeHillTensorEigenDecomposition(const ADDenseMatrix & hill_tensor);
+
+  /**
+   * Perform any necessary steps to finalize strain increment after return mapping iterations
+   * @param inelasticStrainIncrement Inelastic strain increment
+   */
+  virtual void computeStrainFinalize(ADRankTwoTensor & /*inelasticStrainIncrement*/,
+                                     const ADRankTwoTensor & /*stress*/,
+                                     const ADDenseVector & /*stress_dev*/,
+                                     const ADReal & /*delta_gamma*/) override;
+
+  /**
+   * Perform any necessary steps to finalize state after return mapping iterations
[…]
