### 2021-02-11 d719de3
Minor changes to a documentation online page, minor fix on doc build script
