### 2021-02-11 77b3ddd
Register the new kernels to a NS app rather than a moose App
