### 2021-02-11 #3031 @WilkAndy
Hi @arfon .   Yes, the MOOSE framework contains a number of modules designed to solve different types of physical problems.  #2176 described the module that allows users to simulate fluid and heat flow through porous materials.  This submission describes the module that allows users to model geochemistry.
