+++ modules/navier_stokes/src/fvkernels/INSFVMomentumAdvection.C
@@ -326,71 +326,70 @@ INSFVMomentumAdvection::coeffCalculator(const Elem & elem, const ADReal & mu) co
 
     if (onBoundary(*fi))
     {
-      // In my mind there should only be about one bc_id per FaceInfo
-      mooseAssert(fi->boundaryIDs().size() == 1,
-                  "I think some of my logic might depend on my implicit assumption that we have "
-                  "one boundary ID at most per face");
-      const auto bc_id = *fi->boundaryIDs().begin();
-
-      if (_no_slip_wall_boundaries.find(bc_id) != _no_slip_wall_boundaries.end())
+      // Find the boundary id that has an associated INSFV boundary condition
+      // if a face has more than one bc_id
+      for (const auto bc_id : fi->boundaryIDs())
       {
-        // Need to account for viscous shear stress from wall
-        for (const auto i : make_range(_dim))
-          coeff(i) += mu * surface_vector.norm() /
-                      std::abs((fi->faceCentroid() - rc_centroid) * normal) *
-                      (1 - normal(i) * normal(i));
+        if (_no_slip_wall_boundaries.find(bc_id) != _no_slip_wall_boundaries.end())
+        {
+          // Need to account for viscous shear stress from wall
+          for (const auto i : make_range(_dim))
+            coeff(i) += mu * surface_vector.norm() /
+                        std::abs((fi->faceCentroid() - rc_centroid) * normal) *
+                        (1 - normal(i) * normal(i));
 
-        // No flow normal to wall, so no contribution to coefficient from the advection term
-        return;
-      }
+          // No flow normal to wall, so no contribution to coefficient from the advection term
+          return;
+        }
 
-      if (_slip_wall_boundaries.find(bc_id) != _slip_wall_boundaries.end())
-        // In the case of a slip wall we neither have viscous shear stress from the wall nor
-        // normal outflow, so our contribution to the coefficient is zero
-        return;
+        if (_slip_wall_boundaries.find(bc_id) != _slip_wall_boundaries.end())
+          // In the case of a slip wall we neither have viscous shear stress from the wall nor
+          // normal outflow, so our contribution to the coefficient is zero
+          return;
 
-      if (_flow_boundaries.find(bc_id) != _flow_boundaries.end())
-      {
-        ADRealVectorValue face_velocity(_u_var->getBoundaryFaceValue(*fi));
-        if (_v_var)
-          face_velocity(1) = _v_var->getBoundaryFaceValue(*fi);
-        if (_w_var)
-          face_velocity(2) = _w_var->getBoundaryFaceValue(*fi);
+        if (_flow_boundaries.find(bc_id) != _flow_boundaries.end())
+        {
+          ADRealVectorValue face_velocity(_u_var->getBoundaryFaceValue(*fi));
+          if (_v_var)
+            face_velocity(1) = _v_var->getBoundaryFaceValue(*fi);
+          if (_w_var)
+            face_velocity(2) = _w_var->getBoundaryFaceValue(*fi);
 
-        const auto advection_coeffs =
-            Moose::FV::interpCoeffs(_advected_interp_method, *fi, elem_has_info, face_velocity);
-        ADReal temp_coeff = _rho * face_velocity * surface_vector * advection_coeffs.first;
+          const auto advection_coeffs =
+              Moose::FV::interpCoeffs(_advected_interp_method, *fi, elem_has_info, face_velocity);
+          ADReal temp_coeff = _rho * face_velocity * surface_vector * advection_coeffs.first;
 
-        if (_fully_developed_flow_boundaries.find(bc_id) == _fully_developed_flow_boundaries.end())
-          // We are not on a fully developed flow boundary, so we have a viscous term contribution.
-          // This term is slightly modified relative to the internal face term. Instead of the
[…]
