### 2021-02-11 adb33c6
Test compressible flow kernels with artificial diffusion
Displays first order convergence for first order upwind and HLLC and second order
convergence for central difference, so that is what we want.

Refs #16758
