+++ modules/navier_stokes/test/include/userobjects/PressureEqualsDensityFluidProperties.h
@@ -1,0 +1,28 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#pragma once
+
+#include "IdealGasFluidProperties.h"
+
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Woverloaded-virtual"
+
+class PressureEqualsDensityFluidProperties : public IdealGasFluidProperties
+{
+public:
+  static InputParameters validParams();
+
+  PressureEqualsDensityFluidProperties(const InputParameters & parameters);
+
+  ADReal p_from_v_e(const ADReal & v, const ADReal & /*e*/) const override;
+  ADReal T_from_v_e(const ADReal & /*v*/, const ADReal & /*e*/) const override;
+};
+
+#pragma GCC diagnostic pop
