### 2021-02-12 e331c50
Style changes Remove test with a different inteprolation method in the kernel and the postprocessor
