### 2021-02-12 ca4501b
Remove extraneous meshChanged calls
These should be extraneous because there appears to be no mesh operation
immediately before the calls *other* than possible uniform refinement
during restart...and within that uniform refine call, e.g.
`Adaptivity::uniformRefineWithProjection`, there is already a call to
`SubProblem::meshChanged` so we don't need to repeat ourselves
