### 2021-02-12 #3031 @jedbrown
Hi, @WilkAndy. The paper in #2176 mentions chemistry/geochemistry as features several times. I don't know a great way to write criteria to distinguish what is the contribution of a single publication and what other to-be-submitted features does it leverage. Without addressing this specific submission, I think we have to be careful with the precedent and extent to which it may encourage minimum publishable units.
