### 2021-02-12 #3031 @WilkAndy
Hi @arfon , I understand your problem: people might just add a few lines of code to an existing project and publish a paper - "look at our exciting new code!"  I think the current article does not fall into this category.  The chem/geochem mentioned in #2176 is quite basic, and part of the reason for developing the current capability is that geochemists found it difficult to use and lacking features they usually use.

The current addition is a full geochemistry solver, and contains the standard features that geochemists will be familiar with from packages such as [GWB](https://en.wikipedia.org/wiki/The_Geochemist%27s_Workbench) or [EQ3/6](https://www-gs.llnl.gov/energy-homeland-security/geochemistry) or [ToughReact](https://tough.lbl.gov/software/toughreact/)  (I'm not sure whether you've heard of these other packages, but just mention them to demonstrate this is not a small addition to MOOSE).  The current addition can be used as a standalone geochem solver, or can be coupled to the rest of MOOSE to model complicated scenarios.

To give you a feel for its size:
```
/Users/wil04q/projects/moose/modules/geochemistry/> find . -name "*.C" | xargs wc -l
    5025 ./unit/src/GeochemicalSystemTest.C
     771 ./unit/src/GeochemicalDatabaseReaderTest.C
     359 ./unit/src/EquilibriumConstantInterpolatorTest.C
      58 ./unit/src/GeochemistryFormattedOutputTest.C
      92 ./unit/src/GeochemistryActivityCalculatorsTest.C
      34 ./unit/src/main.C
    1874 ./unit/src/MinimalGeochemicalSystemTest.C
    2755 ./unit/src/GeochemicalSolverTest.C
     873 ./unit/src/GeochemistryKineticRateCalculatorTest.C
     306 ./unit/src/GeochemistryIonicStrengthTest.C
      27 ./unit/src/GeochemistrySortedIndicesTest.C
     115 ./unit/src/GeochemicalDatabaseValidatorTest.C
    1509 ./unit/src/GeochemistrySpeciesSwapperTest.C
     159 ./unit/src/GeochemistryActivityCoefficientsDebyeHuckelTest.C
      61 ./test/src/base/GeochemistryTestApp.C
       3 ./build/unity_src/auxkernels_Unity.C
       9 ./build/unity_src/userobjects_Unity.C
      14 ./build/unity_src/utils_Unity.C
       3 ./build/unity_src/kernels_Unity.C
       3 ./build/unity_src/outputs_Unity.C
       6 ./build/unity_src/actions_Unity.C
       2 ./build/unity_src/postprocessors_Unity.C
     212 ./src/auxkernels/GeochemistryQuantityAux.C
      36 ./src/auxkernels/NodalVoidVolumeAux.C
      41 ./src/kernels/GeochemistryDispersion.C
      47 ./src/kernels/GeochemistryTimeDerivative.C
      95 ./src/userobjects/GeochemistryKineticRate.C
     107 ./src/userobjects/NearestNodeNumberUO.C
     222 ./src/userobjects/NodalVoidVolume.C
     126 ./src/userobjects/GeochemistryTimeIndependentReactor.C
     619 ./src/userobjects/GeochemistryTimeDependentReactor.C
     149 ./src/userobjects/GeochemicalModelDefinition.C
     549 ./src/userobjects/GeochemistrySpatialReactor.C
     183 ./src/userobjects/GeochemistryReactorBase.C
      35 ./src/postprocessors/NearestNodeNumber.C
      58 ./src/utils/GeochemistryFormattedOutput.C
     168 ./src/utils/GeochemistryIonicStrength.C
     887 ./src/utils/PertinentGeochemicalSystem.C
      26 ./src/utils/GeochemistrySortedIndices.C
     408 ./src/utils/GeochemicalDatabaseValidator.C
     375 ./src/utils/GeochemistrySpeciesSwapper.C
     457 ./src/utils/GeochemicalSolver.C
     227 ./src/utils/GeochemistryActivityCoefficientsDebyeHuckel.C
      70 ./src/utils/GeochemistryActivityCalculators.C
    2310 ./src/utils/GeochemicalSystem.C
     829 ./src/utils/GeochemicalDatabaseReader.C
     145 ./src/utils/GeochemistryKineticRateCalculator.C
     189 ./src/utils/EquilibriumConstantInterpolator.C
      36 ./src/main.C
     155 ./src/actions/AddTimeDependentReactionSolverAction.C
     268 ./src/actions/AddGeochemistrySolverAction.C
     147 ./src/actions/AddSpatialReactionSolverAction.C
     100 ./src/actions/AddGeochemicalModelInterrogatorAction.C
     139 ./src/actions/AddTimeIndependentReactionSolverAction.C
     330 ./src/outputs/GeochemicalModelInterrogator.C
     343 ./src/outputs/GeochemistryConsoleOutput.C
      84 ./src/base/GeochemistryApp.C
   24230 total
```
