### 2021-02-12 c72d08d
Delete unnecessary usage of individual displacement variables in Density and Contact. Closes #16982.
