+++ modules/tensor_mechanics/src/materials/ADHillElastoPlasticityStressUpdate.C
@@ -7,12 +7,12 @@
 //* Licensed under LGPL 2.1, please see LICENSE for details
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 
-#include "ADTransverselyIsotropicElastoPlasticityStressUpdate.h"
+#include "ADHillElastoPlasticityStressUpdate.h"
 
-registerMooseObject("TensorMechanicsApp", ADTransverselyIsotropicElastoPlasticityStressUpdate);
+registerMooseObject("TensorMechanicsApp", ADHillElastoPlasticityStressUpdate);
 
 InputParameters
-ADTransverselyIsotropicElastoPlasticityStressUpdate::validParams()
+ADHillElastoPlasticityStressUpdate::validParams()
 {
   InputParameters params = ADAnisotropicReturnPlasticityStressUpdateBase::validParams();
   params.addClassDescription(
@@ -33,8 +33,8 @@ ADTransverselyIsotropicElastoPlasticityStressUpdate::validParams()
   return params;
 }
 
-ADTransverselyIsotropicElastoPlasticityStressUpdate::
-    ADTransverselyIsotropicElastoPlasticityStressUpdate(const InputParameters & parameters)
+ADHillElastoPlasticityStressUpdate::ADHillElastoPlasticityStressUpdate(
+    const InputParameters & parameters)
   : ADAnisotropicReturnPlasticityStressUpdateBase(parameters),
     _hill_constants(6),
     _qsigma(0.0),
@@ -86,7 +86,7 @@ ADTransverselyIsotropicElastoPlasticityStressUpdate::
 }
 
 void
-ADTransverselyIsotropicElastoPlasticityStressUpdate::propagateQpStatefulProperties()
+ADHillElastoPlasticityStressUpdate::propagateQpStatefulProperties()
 {
   _hardening_variable[_qp] = _hardening_variable_old[_qp];
   _plasticity_strain[_qp] = _plasticity_strain_old[_qp];
@@ -94,7 +94,7 @@ ADTransverselyIsotropicElastoPlasticityStressUpdate::propagateQpStatefulProperti
 }
 
 void
-ADTransverselyIsotropicElastoPlasticityStressUpdate::computeStressInitialize(
+ADHillElastoPlasticityStressUpdate::computeStressInitialize(
     const ADDenseVector & stress_dev,
     const ADDenseVector & stress,
     const ADRankFourTensor & /*elasticity_tensor*/)
@@ -120,7 +120,7 @@ ADTransverselyIsotropicElastoPlasticityStressUpdate::computeStressInitialize(
 }
 
 void
-ADTransverselyIsotropicElastoPlasticityStressUpdate::computeElasticityTensorEigenDecomposition()
+ADHillElastoPlasticityStressUpdate::computeElasticityTensorEigenDecomposition()
 {
   // Helper method to compute qp matrices required for the elasto-plasticity algorithm.
   _anisotropic_elastic_tensor(0, 0) = (_elasticity_tensor)[_qp](0, 0, 0, 0);
@@ -240,8 +240,8 @@ ADTransverselyIsotropicElastoPlasticityStressUpdate::computeElasticityTensorEige
 }
 
 ADReal
-ADTransverselyIsotropicElastoPlasticityStressUpdate::computeOmega(
-    const ADReal & delta_gamma, const ADDenseVector & /*stress_trial*/)
+ADHillElastoPlasticityStressUpdate::computeOmega(const ADReal & delta_gamma,
+                                                 const ADDenseVector & /*stress_trial*/)
 {
   ADDenseVector K(6);
   ADReal omega = 0.0;
[…]
