+++ modules/tensor_mechanics/doc/content/source/materials/ADGeneralizedRadialReturnStressUpdate.md
@@ -5,8 +5,8 @@ The generalization of the isotropic radial return mapping to Hill plasticity all
 computing the return to the yield surface through a scalar variable when computed
 in a proper subspace.
 
-Anisotropic (Hill) creep and plasticity classes such as `ADTransverselyIsotropicCreepStressUpdate` and 
-`ADTransverselyIsotropicPlasticityStressUpdate` inherit from the generalized radial return mapping. For
+Anisotropic (Hill) creep and plasticity classes such as `ADHillCreepStressUpdate` and 
+`ADHillPlasticityStressUpdate` inherit from the generalized radial return mapping. For
 details on the anisotropic plasticity and anisotropic elasto-plasticity algorithms, consult
 [!cite](versino2018generalized).
 
