+++ modules/tensor_mechanics/src/materials/ADHillPlasticityStressUpdate.C
@@ -7,12 +7,12 @@
 //* Licensed under LGPL 2.1, please see LICENSE for details
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 
-#include "ADTransverselyIsotropicPlasticityStressUpdate.h"
+#include "ADHillPlasticityStressUpdate.h"
 
-registerMooseObject("TensorMechanicsApp", ADTransverselyIsotropicPlasticityStressUpdate);
+registerMooseObject("TensorMechanicsApp", ADHillPlasticityStressUpdate);
 
 InputParameters
-ADTransverselyIsotropicPlasticityStressUpdate::validParams()
+ADHillPlasticityStressUpdate::validParams()
 {
   InputParameters params = ADAnisotropicReturnPlasticityStressUpdateBase::validParams();
   params.addClassDescription(
@@ -33,8 +33,7 @@ ADTransverselyIsotropicPlasticityStressUpdate::validParams()
   return params;
 }
 
-ADTransverselyIsotropicPlasticityStressUpdate::ADTransverselyIsotropicPlasticityStressUpdate(
-    const InputParameters & parameters)
+ADHillPlasticityStressUpdate::ADHillPlasticityStressUpdate(const InputParameters & parameters)
   : ADAnisotropicReturnPlasticityStressUpdateBase(parameters),
     _hill_constants(6),
     _qsigma(0.0),
@@ -79,7 +78,7 @@ ADTransverselyIsotropicPlasticityStressUpdate::ADTransverselyIsotropicPlasticity
 }
 
 void
-ADTransverselyIsotropicPlasticityStressUpdate::propagateQpStatefulProperties()
+ADHillPlasticityStressUpdate::propagateQpStatefulProperties()
 {
   _hardening_variable[_qp] = _hardening_variable_old[_qp];
   _plasticity_strain[_qp] = _plasticity_strain_old[_qp];
@@ -87,10 +86,9 @@ ADTransverselyIsotropicPlasticityStressUpdate::propagateQpStatefulProperties()
 }
 
 void
-ADTransverselyIsotropicPlasticityStressUpdate::computeStressInitialize(
-    const ADDenseVector & stress_dev,
-    const ADDenseVector & /*stress*/,
-    const ADRankFourTensor & elasticity_tensor)
+ADHillPlasticityStressUpdate::computeStressInitialize(const ADDenseVector & stress_dev,
+                                                      const ADDenseVector & /*stress*/,
+                                                      const ADRankFourTensor & elasticity_tensor)
 {
   _hardening_variable[_qp] = _hardening_variable_old[_qp];
   _plasticity_strain[_qp] = _plasticity_strain_old[_qp];
@@ -103,8 +101,8 @@ ADTransverselyIsotropicPlasticityStressUpdate::computeStressInitialize(
 }
 
 ADReal
-ADTransverselyIsotropicPlasticityStressUpdate::computeOmega(const ADReal & delta_gamma,
-                                                            const ADDenseVector & stress_trial)
+ADHillPlasticityStressUpdate::computeOmega(const ADReal & delta_gamma,
+                                           const ADDenseVector & stress_trial)
 {
   ADDenseVector K(6);
   ADReal omega = 0.0;
@@ -124,13 +122,12 @@ ADTransverselyIsotropicPlasticityStressUpdate::computeOmega(const ADReal & delta
 }
 
 void
-ADTransverselyIsotropicPlasticityStressUpdate::computeDeltaDerivatives(
-    const ADReal & delta_gamma,
-    const ADDenseVector & stress_trial,
-    const ADReal & sy_alpha,
-    ADReal & omega,
-    ADReal & omega_gamma,
-    ADReal & sy_gamma)
+ADHillPlasticityStressUpdate::computeDeltaDerivatives(const ADReal & delta_gamma,
[…]

[…]
