+++ modules/navier_stokes/src/fvkernels/PINSFVMomentumAdvection.C
@@ -32,10 +32,347 @@ PINSFVMomentumAdvection::PINSFVMomentumAdvection(const InputParameters & params)
         "of variable type PINSFVVelocityVariable.");
 }
 
+const VectorValue<ADReal> &
+PINSFVMomentumAdvection::rcCoeff(const Elem & elem, const ADReal & mu) const
+{
+  auto it = _rc_a_coeffs.find(&_app);
+  mooseAssert(it != _rc_a_coeffs.end(),
+              "No RC coeffs structure exists for the given MooseApp pointer");
+  mooseAssert(_tid < it->second.size(),
+              "The RC coeffs structure size "
+                  << it->second.size() << " is greater than or equal to the provided thread ID "
+                  << _tid);
+  auto & my_map = it->second[_tid];
+
+  auto rc_map_it = my_map.find(&elem);
+
+  if (rc_map_it != my_map.end())
+    return rc_map_it->second;
+
+  // Returns a pair with the first being an iterator pointing to the key-value pair and the second a
+  // boolean denoting whether a new insertion took place
+  auto emplace_ret = my_map.emplace(&elem, coeffCalculator(elem, mu));
+
+  mooseAssert(emplace_ret.second, "We should have inserted a new key-value pair");
+
+  return emplace_ret.first->second;
+}
+
+#ifdef MOOSE_GLOBAL_AD_INDEXING
+VectorValue<ADReal>
+PINSFVMomentumAdvection::coeffCalculator(const Elem & elem, const ADReal & mu) const
+{
+  // these coefficients arise from simple control volume balances of advection and diffusion. These
+  // coefficients are the linear coefficients associated with the centroid of the control volume.
+  // Note that diffusion coefficients should always be positive, e.g. elliptic operators always
+  // yield positive definite matrices
+  //
+  // Example 1D discretization of diffusion, e.g. the sum of the fluxes around a control volume:
+  //
+  // \sum_f -D \nabla \phi * \hat{n} =
+  //   -D_e * (phi_E - \phi_C) / d_{CE} * 1 - D_w * (\phi_C - \phi_W) / d_{WC} * -1 =
+  //   D_e / d_{CE} * (\phi_C - \phi_E) + D_w / d_{WC} * (\phi_C - \phi_W)
+  //
+  // Note the positive coefficients for \phi_C !!
+  //
+  // Now an example 1D discretization for advection using central differences, e.g. an average
+  // interpolation
+  //
+  // \sum_f \vec{u} \phi \hat{n} =
+  //   u_w * (\phi_W + \phi_C) / 2 * -1 + u_e * (\phi_C + \phi_E) / 2 * 1 =
+  //   -u_w / 2 * \phi_W + u_e / 2 * \phi_E + (u_e - u_w) / 2 * \phi_C
+  //
+  // Note that the coefficient for \phi_C may or may not be positive depending on the values of u_e
+  // and u_w
+
+  VectorValue<ADReal> coeff = 0;
+
+  ADRealVectorValue elem_velocity(_u_var->getElemValue(&elem));
+
+  if (_v_var)
+    elem_velocity(1) = _v_var->getElemValue(&elem);
+  if (_w_var)
+    elem_velocity(2) = _w_var->getElemValue(&elem);
+
+  auto action_functor = [&coeff,
+                         &elem_velocity,
+                         &mu,
+#ifndef NDEBUG
+                         &elem,
+#endif
+                         this](const Elem & libmesh_dbg_var(functor_elem),
+                               const Elem * const neighbor,
+                               const FaceInfo * const fi,
+                               const Point & surface_vector,
+                               Real libmesh_dbg_var(coord),
+                               const bool elem_has_info) {
[…]
