+++ modules/navier_stokes/include/fvbcs/PINSFVMomentumAdvectionOutflowBC.h
@@ -1,0 +1,39 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#pragma once
+
+#include "FVMatAdvectionOutflowBC.h"
+#include "INSFVFullyDevelopedFlowBC.h"
+
+class INSFVVelocityVariable;
+
+class PINSFVMomentumAdvectionOutflowBC : public FVMatAdvectionOutflowBC,
+                                        public INSFVFullyDevelopedFlowBC
+{
+public:
+  static InputParameters validParams();
+  PINSFVMomentumAdvectionOutflowBC(const InputParameters & params);
+
+protected:
+  virtual ADReal computeQpResidual() override;
+
+  /// x-velocity
+  const INSFVVelocityVariable * const _u_var;
+  /// y-velocity
+  const INSFVVelocityVariable * const _v_var;
+  /// z-velocity
+  const INSFVVelocityVariable * const _w_var;
+  /// porosity
+  const VariableValue & _eps;
+  /// porosity in the neighbor cell (ghost?)
+  const VariableValue & _eps_neighbor;
+  /// the dimension of the simulation
+  const unsigned int _dim;
+};
