### 2021-02-15 9983ad1
Add ray_tracing dependency to heat_conduction
refs #16954
