+++ modules/heat_conduction/doc/config.yml
@@ -1,21 +1,23 @@
 Content:
     - framework/doc/content
     - modules/heat_conduction/doc/content
+    - modules/ray_tracing/doc/content
 
 Renderer:
     type: MooseDocs.base.MaterializeRenderer
 
 Extensions:
     MooseDocs.extensions.navigation:
-        name: MOOSE Functional Expansion Tools Module
+        name: MOOSE Heat Conduction Module
         repo: https://github.com/idaholab/moose
-        home: /modules/functional_expansion_tools
+        home: /modules/heat_conduction
     MooseDocs.extensions.appsyntax:
         executable: ${MOOSE_DIR}/modules/heat_conduction
         remove: !include framework/doc/remove.yml
         includes:
             - framework/include
             - modules/heat_conduction/include
+            - modules/ray_tracing/include
     MooseDocs.extensions.common:
         shortcuts: !include framework/doc/globals.yml
     MooseDocs.extensions.acronym:
