+++ modules/heat_conduction/src/actions/RadiationTransferAction.C
@@ -12,11 +12,15 @@
 #include "MooseMesh.h"
 #include "MeshGeneratorMesh.h"
 #include "FEProblemBase.h"
+#include "PatchSidesetGenerator.h"
+#include "ViewFactorRayBC.h"
+#include "ViewFactorRayStudy.h"
 
-registerMooseAction("HeatConductionApp", RadiationTransferAction, "add_mesh_generator");
+registerMooseAction("HeatConductionApp", RadiationTransferAction, "append_mesh_generator");
 registerMooseAction("HeatConductionApp", RadiationTransferAction, "setup_mesh_complete");
 registerMooseAction("HeatConductionApp", RadiationTransferAction, "add_user_object");
 registerMooseAction("HeatConductionApp", RadiationTransferAction, "add_bc");
+registerMooseAction("HeatConductionApp", RadiationTransferAction, "add_ray_boundary_condition");
 
 template <>
 InputParameters
@@ -43,13 +47,11 @@ validParams<RadiationTransferAction>()
                                                      "Number of radiation patches per sideset.");
   MultiMooseEnum partitioning(
       "default=-3 metis=-2 parmetis=-1 linear=0 centroid hilbert_sfc morton_sfc", "default");
+  partitioning += "grid";
   params.addParam<MultiMooseEnum>(
       "partitioners",
       partitioning,
       "Specifies a mesh partitioner to use when preparing the radiation patches.");
-
-  params.addRequiredParam<MeshGeneratorName>("final_mesh_generator",
-                                             "Name of the final mesh generator.");
 
   MultiMooseEnum direction("x y z radial");
   params.addParam<MultiMooseEnum>("centroid_partitioner_directions",
@@ -59,30 +61,104 @@ validParams<RadiationTransferAction>()
 
   params.addRequiredParam<VariableName>("temperature", "The coupled temperature variable.");
   params.addRequiredParam<std::vector<Real>>("emissivity", "Emissivities for each boundary.");
+
+  MooseEnum view_factor_calculator("analytical ray_tracing", "ray_tracing");
+  params.addParam<MooseEnum>(
+      "view_factor_calculator", view_factor_calculator, "The view factor calculator being used.");
+
+  params.addParam<bool>(
+      "print_view_factor_info", false, "Flag to print information about computed view factors.");
+  params.addParam<bool>("normalize_view_factor",
+                        true,
+                        "Determines if view factors are normalized to sum to one (consistent with "
+                        "their definition).");
+
+  std::vector<BoundaryName> empty = {};
+  params.addParam<std::vector<BoundaryName>>(
+      "symmetry_boundary",
+      empty,
+      "The sidesets that represent symmetry lines/planes for the problem. These sidesets do not "
+      "participate in the radiative exchange"
+      "so they should not be listed in the sidesets parameter.");
+
+  MooseEnum qtypes("GAUSS GRID", "GRID");
+  params.addParam<MooseEnum>(
+      "ray_tracing_face_type", qtypes, "The face quadrature rule type used for ray tracing.");
+
+  MooseEnum qorders("CONSTANT FIRST SECOND THIRD FOURTH FIFTH SIXTH SEVENTH EIGHTH NINTH TENTH "
+                    "ELEVENTH TWELFTH THIRTEENTH FOURTEENTH FIFTEENTH SIXTEENTH SEVENTEENTH "
+                    "EIGHTTEENTH NINTEENTH TWENTIETH",
+                    "CONSTANT");
+  params.addParam<MooseEnum>(
+      "ray_tracing_face_order", qorders, "The face quadrature rule order used for ray tracing.");
+
[…]

[…]
