### 2021-02-15 06f06ac
Initial commit for view factor ray tracing objects
refs #16954
