+++ modules/heat_conduction/src/userobjects/ViewFactorRayStudy.C
@@ -1,0 +1,549 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#include "ViewFactorRayStudy.h"
+
+// MOOSE includes
+#include "TimedPrint.h"
+
+// Local includes
+#include "ViewFactorRayBC.h"
+
+// libMesh includes
+#include "libmesh/parallel_algebra.h"
+#include "libmesh/parallel_sync.h"
+#include "libmesh/enum_quadrature_type.h"
+#include "libmesh/fe_base.h"
+#include "libmesh/quadrature.h"
+
+// Ray tracing includes
+#include "ReflectRayBC.h"
+#include "RayTracingPackingUtils.h"
+
+registerMooseObject("HeatConductionApp", ViewFactorRayStudy);
+
+InputParameters
+ViewFactorRayStudy::validParams()
+{
+  auto params = RayTracingStudy::validParams();
+
+  params.addRequiredParam<std::vector<BoundaryName>>(
+      "boundary", "The list of boundaries where view factors are desired");
+
+  MooseEnum qorders("CONSTANT FIRST SECOND THIRD FOURTH FIFTH SIXTH SEVENTH EIGHTH NINTH TENTH "
+                    "ELEVENTH TWELFTH THIRTEENTH FOURTEENTH FIFTEENTH SIXTEENTH SEVENTEENTH "
+                    "EIGHTTEENTH NINTEENTH TWENTIETH",
+                    "CONSTANT");
+  params.addParam<MooseEnum>("face_order", qorders, "The face quadrature rule order");
+
+  MooseEnum qtypes("GAUSS GRID", "GRID");
+  params.addParam<MooseEnum>("face_type", qtypes, "The face quadrature type");
+
+  MooseEnum convention("positive=0 negative=1", "positive");
+  params.addParam<MooseEnum>(
+      "internal_convention",
+      convention,
+      "The convention for spawning rays from internal sidesets; denotes the sign of the dot "
+      "product between a ray and the internal sideset side normal");
+
+  params.addParam<unsigned int>(
+      "polar_quad_order",
+      16,
+      "Order of the polar quadrature [polar angle is between ray and normal]. Must be even.");
+  params.addParam<unsigned int>(
+      "azimuthal_quad_order",
+      8,
+      "Order of the azimuthal quadrature per quadrant [azimuthal angle is measured in "
+      "a plane perpendicular to the normal].");
+
+  // Shouldn't ever need RayKernels for view factors
+  params.set<bool>("ray_kernel_coverage_check") = false;
+  params.suppressParameter<bool>("ray_kernel_coverage_check");
+
+  // So that the study executes before the RayTracingViewFactor
+  params.set<bool>("force_preaux") = true;
+  params.suppressParameter<bool>("force_preaux");
+
+  // Need to use internal sidesets
+  params.set<bool>("use_internal_sidesets") = true;
+  params.suppressParameter<bool>("use_internal_sidesets");
+
+  // Don't verify Rays in opt mode by default - it's expensive
+  params.set<bool>("verify_rays") = false;
+
+  // No need to use Ray registration
+  params.set<bool>("_use_ray_registration") = false;
+  // Do not need to bank Rays on completion
+  params.set<bool>("_bank_rays_on_completion") = false;
+
+  return params;
+}
+
[…]
