### 2021-02-15 0ca00fd
Tests and documentation for ray tracing view factors
refs #16954
