### 2021-02-15 3d450e9
Allow ray-dependent definition of data and aux data
refs #16999
