### 2021-02-15 4d35d4e
Do not require 'rays' parameter
refs #17001
