+++ modules/ray_tracing/src/base/RayTracingObject.C
@@ -24,8 +24,8 @@ RayTracingObject::validParams()
                                   "provided, this will default to the one study that exists.");
   params.addParam<std::vector<std::string>>(
       "rays",
-      "The name of the Rays associated with this object (only used if Ray registration is enabled "
-      "within the study)");
+      "The name of the Rays associated with this object; only used if Ray registration is enabled "
+      "within the study. If no Rays are supplied, this object will be applied to all Rays.");
 
   params.addPrivateParam<RayTracingStudy *>("_ray_tracing_study");
 
