### 2021-02-15 4fd85f9
Unify element activation, element death and moving interface
Generalize the existing ActivateElementUserObjects, add ElementSubdomainModifiers, and fix some parallel bugs.

close #17100
