+++ test/src/materials/StatefulMaterial.C
@@ -16,6 +16,7 @@ StatefulMaterial::validParams()
 {
   InputParameters params = Material::validParams();
   params.addParam<Real>("initial_diffusivity", 0.5, "The Initial Diffusivity");
+  params.addParam<Real>("multiplier", 2, "The factor to multiply at each step");
   return params;
 }
 
@@ -31,7 +32,11 @@ StatefulMaterial::StatefulMaterial(const InputParameters & parameters)
 
     // Retrieve/use an old value of diffusivity.
     // Note: this is _expensive_ - only do this if you REALLY need it!
-    _diffusivity_old(getMaterialPropertyOld<Real>("diffusivity"))
+    _diffusivity_old(getMaterialPropertyOld<Real>("diffusivity")),
+
+    // Get the multiplier.
+    // The diffusivity is multiplied by this multiplier at each step.
+    _multiplier(getParam<Real>("multiplier"))
 {
 }
 
@@ -44,5 +49,5 @@ StatefulMaterial::initQpStatefulProperties()
 void
 StatefulMaterial::computeQpProperties()
 {
-  _diffusivity[_qp] = _diffusivity_old[_qp] * 2.0;
+  _diffusivity[_qp] = _diffusivity_old[_qp] * _multiplier;
 }
