### 2021-02-15 d5757eb
Unify element activation, element death and moving interface
Generalize the existing ActivateElementUserObjects, add ElementSubdomainModifiers, and fix some parallel bugs.

close #17100
