+++ framework/src/userobject/ElementSubdomainModifier.C
@@ -1,0 +1,360 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#include "ElementSubdomainModifier.h"
+#include "DisplacedProblem.h"
+
+#include "libmesh/parallel_algebra.h"
+#include "libmesh/parallel.h"
+#include "libmesh/dof_map.h"
+
+#include "libmesh/parallel_ghost_sync.h"
+
+InputParameters
+ElementSubdomainModifier::validParams()
+{
+  InputParameters params = ElementUserObject::validParams();
+  params.addClassDescription(
+      "Modify element subdomain ID. This userobject only runs on the undisplaced mesh, and it will "
+      "modify both the undisplaced and the displaced mesh.");
+  params.addParam<bool>("apply_initial_conditions",
+                        true,
+                        "Whether to apply initial conditions on the moved nodes and elements");
+  params.addParam<BoundaryName>(
+      "moving_boundary_name",
+      "Boundary to modify when an element is moved. A boundary with the provided name will be "
+      "created if not already exists on the mesh.");
+  params.set<bool>("use_displaced_mesh") = false;
+  params.suppressParameter<bool>("use_displaced_mesh");
+  return params;
+}
+
+ElementSubdomainModifier::ElementSubdomainModifier(const InputParameters & parameters)
+  : ElementUserObject(parameters),
+    _displaced_problem(_fe_problem.getDisplacedProblem().get()),
+    _apply_ic(getParam<bool>("apply_initial_conditions")),
+    _moving_boundary_specified(isParamValid("moving_boundary_name"))
+{
+}
+
+void
+ElementSubdomainModifier::initialSetup()
+{
+  if (_moving_boundary_specified)
+  {
+    _moving_boundary_name = getParam<BoundaryName>("moving_boundary_name");
+    setMovingBoundaryName(_mesh);
+    if (_displaced_problem)
+      setMovingBoundaryName(_displaced_problem->mesh());
+  }
+}
+
+void
+ElementSubdomainModifier::setMovingBoundaryName(MooseMesh & mesh)
+{
+  // We only need one boundary to modify. Create a dummy vector just to use the API.
+  const std::vector<BoundaryID> boundary_ids = mesh.getBoundaryIDs({{_moving_boundary_name}}, true);
+  mooseAssert(boundary_ids.size() == 1, "Expect exactly one boundary ID.");
+  _moving_boundary_id = boundary_ids[0];
+  mesh.setBoundaryName(_moving_boundary_id, _moving_boundary_name);
+  mesh.getMesh().get_boundary_info().sideset_name(_moving_boundary_id) = _moving_boundary_name;
+  mesh.getMesh().get_boundary_info().nodeset_name(_moving_boundary_id) = _moving_boundary_name;
+}
+
+void
+ElementSubdomainModifier::initialize()
+{
+  _moved_elems.clear();
+  _moved_displaced_elems.clear();
+  _moved_nodes.clear();
+}
+
+void
+ElementSubdomainModifier::execute()
+{
+  // First, compute the desired subdomain ID for the current element.
+  SubdomainID subdomain_id = computeSubdomainID();
+
+  // If the current element's subdomain ID isn't what we want
+  if (subdomain_id != std::numeric_limits<SubdomainID>::max() &&
+      _current_elem->subdomain_id() != subdomain_id)
+  {
[…]
