+++ modules/electromagnetics/doc/content/source/kernels/FuncDiffusion.md
@@ -1,1 +1,23 @@
-!template load file=stubs/moose_object.md.template name=FuncDiffusion syntax=/Kernels/FuncDiffusion
+# FuncDiffusion
+
+!alert construction title=Undocumented Class
+The FuncDiffusion has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /Kernels/FuncDiffusion
+
+## Overview
+
+!! Replace these lines with information regarding the FuncDiffusion object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the FuncDiffusion object.
+
+!syntax parameters /Kernels/FuncDiffusion
+
+!syntax inputs /Kernels/FuncDiffusion
+
+!syntax children /Kernels/FuncDiffusion
