+++ modules/electromagnetics/doc/content/source/interfacekernels/ParallelElectricFieldInterface.md
@@ -1,1 +1,23 @@
-!template load file=stubs/moose_object.md.template name=ParallelElectricFieldInterface syntax=/InterfaceKernels/ParallelElectricFieldInterface
+# ParallelElectricFieldInterface
+
+!alert construction title=Undocumented Class
+The ParallelElectricFieldInterface has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /InterfaceKernels/ParallelElectricFieldInterface
+
+## Overview
+
+!! Replace these lines with information regarding the ParallelElectricFieldInterface object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the ParallelElectricFieldInterface object.
+
+!syntax parameters /InterfaceKernels/ParallelElectricFieldInterface
+
+!syntax inputs /InterfaceKernels/ParallelElectricFieldInterface
+
+!syntax children /InterfaceKernels/ParallelElectricFieldInterface
