+++ modules/electromagnetics/doc/content/source/kernels/CoeffField.md
@@ -1,1 +1,23 @@
-!template load file=stubs/moose_object.md.template name=CoeffField syntax=/Kernels/CoeffField
+# CoeffField
+
+!alert construction title=Undocumented Class
+The CoeffField has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /Kernels/CoeffField
+
+## Overview
+
+!! Replace these lines with information regarding the CoeffField object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the CoeffField object.
+
+!syntax parameters /Kernels/CoeffField
+
+!syntax inputs /Kernels/CoeffField
+
+!syntax children /Kernels/CoeffField
