+++ modules/electromagnetics/doc/content/source/kernels/CurlCurlField.md
@@ -1,1 +1,23 @@
-!template load file=stubs/moose_object.md.template name=CurlCurlField syntax=/Kernels/CurlCurlField
+# CurlCurlField
+
+!alert construction title=Undocumented Class
+The CurlCurlField has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /Kernels/CurlCurlField
+
+## Overview
+
+!! Replace these lines with information regarding the CurlCurlField object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the CurlCurlField object.
+
+!syntax parameters /Kernels/CurlCurlField
+
+!syntax inputs /Kernels/CurlCurlField
+
+!syntax children /Kernels/CurlCurlField
