+++ modules/electromagnetics/doc/content/source/kernels/CoupledCoeffField.md
@@ -1,1 +1,23 @@
-!template load file=stubs/moose_object.md.template name=CoupledCoeffField syntax=/Kernels/CoupledCoeffField
+# CoupledCoeffField
+
+!alert construction title=Undocumented Class
+The CoupledCoeffField has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /Kernels/CoupledCoeffField
+
+## Overview
+
+!! Replace these lines with information regarding the CoupledCoeffField object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the CoupledCoeffField object.
+
+!syntax parameters /Kernels/CoupledCoeffField
+
+!syntax inputs /Kernels/CoupledCoeffField
+
+!syntax children /Kernels/CoupledCoeffField
