+++ modules/electromagnetics/doc/content/source/functions/JinSlabCoeffFunc.md
@@ -1,1 +1,23 @@
-!template load file=stubs/moose_object.md.template name=JinSlabCoeffFunc syntax=/Functions/JinSlabCoeffFunc
+# JinSlabCoeffFunc
+
+!alert construction title=Undocumented Class
+The JinSlabCoeffFunc has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /Functions/JinSlabCoeffFunc
+
+## Overview
+
+!! Replace these lines with information regarding the JinSlabCoeffFunc object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the JinSlabCoeffFunc object.
+
+!syntax parameters /Functions/JinSlabCoeffFunc
+
+!syntax inputs /Functions/JinSlabCoeffFunc
+
+!syntax children /Functions/JinSlabCoeffFunc
