+++ modules/electromagnetics/doc/content/source/functions/RobinTestFunc.md
@@ -1,1 +1,23 @@
-!template load file=stubs/moose_object.md.template name=RobinTestFunc syntax=/Functions/RobinTestFunc
+# RobinTestFunc
+
+!alert construction title=Undocumented Class
+The RobinTestFunc has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /Functions/RobinTestFunc
+
+## Overview
+
+!! Replace these lines with information regarding the RobinTestFunc object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the RobinTestFunc object.
+
+!syntax parameters /Functions/RobinTestFunc
+
+!syntax inputs /Functions/RobinTestFunc
+
+!syntax children /Functions/RobinTestFunc
