+++ modules/electromagnetics/doc/content/source/kernels/VectorCurrentSource.md
@@ -1,1 +1,23 @@
-!template load file=stubs/moose_object.md.template name=VectorCurrentSource syntax=/Kernels/VectorCurrentSource
+# VectorCurrentSource
+
+!alert construction title=Undocumented Class
+The VectorCurrentSource has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /Kernels/VectorCurrentSource
+
+## Overview
+
+!! Replace these lines with information regarding the VectorCurrentSource object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the VectorCurrentSource object.
+
+!syntax parameters /Kernels/VectorCurrentSource
+
+!syntax inputs /Kernels/VectorCurrentSource
+
+!syntax children /Kernels/VectorCurrentSource
