+++ modules/electromagnetics/doc/content/source/bcs/VectorCurlPenaltyDirichletBC.md
@@ -1,1 +1,23 @@
-!template load file=stubs/moose_object.md.template name=VectorCurlPenaltyDirichletBC syntax=/BCs/VectorCurlPenaltyDirichletBC
+# VectorCurlPenaltyDirichletBC
+
+!alert construction title=Undocumented Class
+The VectorCurlPenaltyDirichletBC has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /BCs/VectorCurlPenaltyDirichletBC
+
+## Overview
+
+!! Replace these lines with information regarding the VectorCurlPenaltyDirichletBC object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the VectorCurlPenaltyDirichletBC object.
+
+!syntax parameters /BCs/VectorCurlPenaltyDirichletBC
+
+!syntax inputs /BCs/VectorCurlPenaltyDirichletBC
+
+!syntax children /BCs/VectorCurlPenaltyDirichletBC

[…]
