+++ modules/electromagnetics/doc/content/source/kernels/VectorSecondTimeDerivative.md
@@ -1,1 +1,23 @@
-!template load file=stubs/moose_object.md.template name=VectorSecondTimeDerivative syntax=/Kernels/VectorSecondTimeDerivative
+# VectorSecondTimeDerivative
+
+!alert construction title=Undocumented Class
+The VectorSecondTimeDerivative has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /Kernels/VectorSecondTimeDerivative
+
+## Overview
+
+!! Replace these lines with information regarding the VectorSecondTimeDerivative object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the VectorSecondTimeDerivative object.
+
+!syntax parameters /Kernels/VectorSecondTimeDerivative
+
+!syntax inputs /Kernels/VectorSecondTimeDerivative
+
+!syntax children /Kernels/VectorSecondTimeDerivative
