+++ modules/electromagnetics/doc/content/source/kernels/CoupledCoeffTimeDerivative.md
@@ -1,1 +1,23 @@
-!template load file=stubs/moose_object.md.template name=CoupledCoeffTimeDerivative syntax=/Kernels/CoupledCoeffTimeDerivative
+# CoupledCoeffTimeDerivative
+
+!alert construction title=Undocumented Class
+The CoupledCoeffTimeDerivative has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /Kernels/CoupledCoeffTimeDerivative
+
+## Overview
+
+!! Replace these lines with information regarding the CoupledCoeffTimeDerivative object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the CoupledCoeffTimeDerivative object.
+
+!syntax parameters /Kernels/CoupledCoeffTimeDerivative
+
+!syntax inputs /Kernels/CoupledCoeffTimeDerivative
+
+!syntax children /Kernels/CoupledCoeffTimeDerivative
