+++ modules/electromagnetics/doc/content/source/functions/ZPolarizedWave.md
@@ -1,1 +1,23 @@
-!template load file=stubs/moose_object.md.template name=ZPolarizedWave syntax=/Functions/ZPolarizedWave
+# ZPolarizedWave
+
+!alert construction title=Undocumented Class
+The ZPolarizedWave has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /Functions/ZPolarizedWave
+
+## Overview
+
+!! Replace these lines with information regarding the ZPolarizedWave object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the ZPolarizedWave object.
+
+!syntax parameters /Functions/ZPolarizedWave
+
+!syntax inputs /Functions/ZPolarizedWave
+
+!syntax children /Functions/ZPolarizedWave
