+++ modules/electromagnetics/doc/content/source/kernels/CoupledCurlCurlField.md
@@ -1,1 +1,23 @@
-!template load file=stubs/moose_object.md.template name=CoupledCurlCurlField syntax=/Kernels/CoupledCurlCurlField
+# CoupledCurlCurlField
+
+!alert construction title=Undocumented Class
+The CoupledCurlCurlField has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /Kernels/CoupledCurlCurlField
+
+## Overview
+
+!! Replace these lines with information regarding the CoupledCurlCurlField object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the CoupledCurlCurlField object.
+
+!syntax parameters /Kernels/CoupledCurlCurlField
+
+!syntax inputs /Kernels/CoupledCurlCurlField
+
+!syntax children /Kernels/CoupledCurlCurlField
