+++ modules/electromagnetics/doc/content/source/bcs/ReflectionBC.md
@@ -1,1 +1,23 @@
-!template load file=stubs/moose_object.md.template name=ReflectionBC syntax=/BCs/ReflectionBC
+# ReflectionBC
+
+!alert construction title=Undocumented Class
+The ReflectionBC has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /BCs/ReflectionBC
+
+## Overview
+
+!! Replace these lines with information regarding the ReflectionBC object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the ReflectionBC object.
+
+!syntax parameters /BCs/ReflectionBC
+
+!syntax inputs /BCs/ReflectionBC
+
+!syntax children /BCs/ReflectionBC
