+++ modules/electromagnetics/doc/content/source/functions/WaveCoeff.md
@@ -1,1 +1,23 @@
-!template load file=stubs/moose_object.md.template name=WaveCoeff syntax=/Functions/WaveCoeff
+# WaveCoeff
+
+!alert construction title=Undocumented Class
+The WaveCoeff has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /Functions/WaveCoeff
+
+## Overview
+
+!! Replace these lines with information regarding the WaveCoeff object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the WaveCoeff object.
+
+!syntax parameters /Functions/WaveCoeff
+
+!syntax inputs /Functions/WaveCoeff
+
+!syntax children /Functions/WaveCoeff
