+++ modules/electromagnetics/doc/content/source/postprocessors/ReflectionCoefficient.md
@@ -1,1 +1,23 @@
-!template load file=stubs/moose_object.md.template name=ReflectionCoefficient syntax=/Postprocessors/ReflectionCoefficient
+# ReflectionCoefficient
+
+!alert construction title=Undocumented Class
+The ReflectionCoefficient has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /UserObjects/ReflectionCoefficient
+
+## Overview
+
+!! Replace these lines with information regarding the ReflectionCoefficient object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the ReflectionCoefficient object.
+
+!syntax parameters /UserObjects/ReflectionCoefficient
+
+!syntax inputs /UserObjects/ReflectionCoefficient
+
+!syntax children /UserObjects/ReflectionCoefficient
