+++ modules/electromagnetics/doc/content/source/bcs/CoupledGradBC.md
@@ -1,1 +1,23 @@
-!template load file=stubs/moose_object.md.template name=CoupledGradBC syntax=/BCs/CoupledGradBC
+# CoupledGradBC
+
+!alert construction title=Undocumented Class
+The CoupledGradBC has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /BCs/CoupledGradBC
+
+## Overview
+
+!! Replace these lines with information regarding the CoupledGradBC object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the CoupledGradBC object.
+
+!syntax parameters /BCs/CoupledGradBC
+
+!syntax inputs /BCs/CoupledGradBC
+
+!syntax children /BCs/CoupledGradBC
