+++ modules/electromagnetics/doc/content/source/auxkernels/PotentialToFieldAux.md
@@ -1,1 +1,23 @@
-!template load file=stubs/moose_object.md.template name=PotentialToFieldAux syntax=/AuxKernels/PotentialToFieldAux
+# PotentialToFieldAux
+
+!alert construction title=Undocumented Class
+The PotentialToFieldAux has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /AuxKernels/PotentialToFieldAux
+
+## Overview
+
+!! Replace these lines with information regarding the PotentialToFieldAux object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the PotentialToFieldAux object.
+
+!syntax parameters /AuxKernels/PotentialToFieldAux
+
+!syntax inputs /AuxKernels/PotentialToFieldAux
+
+!syntax children /AuxKernels/PotentialToFieldAux
