+++ modules/electromagnetics/doc/content/source/bcs/AbsorbingBC.md
@@ -1,1 +1,23 @@
-!template load file=stubs/moose_object.md.template name=AbsorbingBC syntax=/BCs/AbsorbingBC
+# AbsorbingBC
+
+!alert construction title=Undocumented Class
+The AbsorbingBC has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /BCs/AbsorbingBC
+
+## Overview
+
+!! Replace these lines with information regarding the AbsorbingBC object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the AbsorbingBC object.
+
+!syntax parameters /BCs/AbsorbingBC
+
+!syntax inputs /BCs/AbsorbingBC
+
+!syntax children /BCs/AbsorbingBC

[…]
