### 2021-02-15 2e8a05d
Correctly handle application names with dashes (i.e., 'relap-7')
(refs #16868)
