### 2021-02-15 538b405
Fix creep plastic multiplier bug and update tests
