+++ modules/phase_field/src/action/PolycrystalDislocationKernelAction.C
@@ -13,6 +13,8 @@
 /****************************************************************/
 
 #include "PolycrystalDislocationKernelAction.h"
+#include "Factory.h"
+#include "FEProblemBase.h"
 
 registerMooseAction("MarmotApp", PolycrystalDislocationKernelAction, "add_kernel");
 
