### 2021-02-16 4d507cc
Add Hill creep tests from analytical verification exercise
Avoid Eigen's ambiguous operator error.

- Eigen matrices will be non-AD. For now, they can be considered as user input.
- Some clean up. Apply hit format to some inputs.
