+++ modules/tensor_mechanics/src/materials/ADHillCreepStressUpdate.C
@@ -8,6 +8,7 @@
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 
 #include "ADHillCreepStressUpdate.h"
+#include "ElasticityTensorTools.h"
 
 registerMooseObject("TensorMechanicsApp", ADHillCreepStressUpdate);
 
