### 2021-02-16 73c153a
Add Hill creep tests from analytical verification exercise
