+++ modules/tensor_mechanics/src/materials/ADHillPlasticityStressUpdate.C
@@ -8,6 +8,7 @@
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 
 #include "ADHillPlasticityStressUpdate.h"
+#include "ElasticityTensorTools.h"
 
 registerMooseObject("TensorMechanicsApp", ADHillPlasticityStressUpdate);
 
