### 2021-02-16 17da164
Add more options to PINSFV friction models
