### 2021-02-16 78a032a
Add more options to PINSFV friction models
