### 2021-02-17 ada278f
Test compressible flow kernels without diffusion
