+++ modules/navier_stokes/test/tests/finite_volume/cns/mms/1d/test.py
@@ -41,8 +41,36 @@ class Test1DHLLC(unittest.TestCase):
             print("%s, %f" % (key, value))
             self.assertTrue(fuzzyAbsoluteEqual(value, 1., .05))
 
+class Test1DUpwindNoDiffusion(unittest.TestCase):
+# class Test1DUpwindNoDiffusion():
+    def test(self):
+        labels = ['L2rho', 'L2rho_u', 'L2rho_et']
+        df1 = mms.run_spatial('no-diffusion-mms.i', list(range(6,12)), "--allow-test-objects", y_pp=labels, mpi=1)
+
+        fig = mms.ConvergencePlot(xlabel='Element Size ($h$)', ylabel='$L_2$ Error')
+        fig.plot(df1, label=labels, marker='o', markersize=8, num_fitted_points=3, slope_precision=1)
+        fig.save('no-diffusion-upwind.png')
+        for key,value in fig.label_to_slope.items():
+            print("%s, %f" % (key, value))
+            self.assertTrue(fuzzyAbsoluteEqual(value, 1., .05))
+
+class Test1DHLLCNoDiffusion(unittest.TestCase):
+# class Test1DHLLCNoDiffusion():
+    def test(self):
+        labels = ['L2rho', 'L2rho_u', 'L2rho_et']
+        df1 = mms.run_spatial('no-diffusion-hllc-mms.i', list(range(6,12)), "--allow-test-objects", y_pp=labels, mpi=1)
+
+        fig = mms.ConvergencePlot(xlabel='Element Size ($h$)', ylabel='$L_2$ Error')
+        fig.plot(df1, label=labels, marker='o', markersize=8, num_fitted_points=3, slope_precision=1)
+        fig.save('no-diffusion-hllc.png')
+        for key,value in fig.label_to_slope.items():
+            print("%s, %f" % (key, value))
+            self.assertTrue(fuzzyAbsoluteEqual(value, 1., .05))
+
 if __name__ == '__main__':
     unittest.main(__name__, verbosity=2)
     # Test1DUpwind().test()
     # Test1DAverage().test()
     # Test1DHLLC().test()
+    # Test1DUpwindNoDiffusion().test()
+    # Test1DHLLCNoDiffusion().test()
