### 2021-02-17 58a0f1a
Remove extraneous error info
refs #17030
