+++ modules/ray_tracing/src/userobjects/RayTracingStudy.C
@@ -170,7 +170,6 @@ RayTracingStudy::RayTracingStudy(const InputParameters & parameters)
 
     _called_initial_setup(false),
     _need_to_associate_registered_rays(false),
-    _verify_ray_error_prefix("While moving Ray to buffer in " + _error_prefix),
 
     _elem_index_helper(_mesh.getMesh(), name() + "_elem_index")
 {
@@ -197,9 +196,8 @@ RayTracingStudy::RayTracingStudy(const InputParameters & parameters)
                  "mix RayKernels that contribute to the Jacobian/residual with those that do not.");
 
     if (_app.useEigenvalue())
-      mooseError(_error_prefix,
-                 ": Execution on residual and Jacobian evaluation\n",
-                 "(execute_on = PRE_KERNELS) is not supported for an eigenvalue solve.");
+      mooseError("Execution on residual and Jacobian evaluation (execute_on = PRE_KERNELS)\n",
+                 "is not supported for an eigenvalue solve.");
   }
 
   resetUniqueRayIDs();
@@ -245,8 +243,7 @@ RayTracingStudy::initialSetup()
   getRayKernels(ray_kernels, 0);
   for (const auto & rkb : ray_kernels)
     if (dynamic_cast<RayKernel *>(rkb) && !_execute_enum.contains(EXEC_PRE_KERNELS))
-      mooseError(_error_prefix,
-                 ": Has RayKernel objects that contribute to residuals and Jacobians.",
+      mooseError("Has RayKernel objects that contribute to residuals and Jacobians.",
                  "\nIn this case, the study should use the execute_on = PRE_KERNELS");
 
   // Build 1D quadrature rule for along a segment
@@ -324,7 +321,7 @@ RayTracingStudy::coverageChecks()
     if (!missing.empty() && !ray_kernel_blocks.count(Moose::ANY_BLOCK_ID))
     {
       std::ostringstream error;
-      error << _error_prefix << ": Subdomains { ";
+      error << "Subdomains { ";
       std::copy(missing.begin(), missing.end(), std::ostream_iterator<SubdomainID>(error, " "));
       error << "} do not have RayKernels defined!";
 
@@ -360,16 +357,7 @@ RayTracingStudy::verifyDependenciesExist(const std::vector<RayTracingObject *> &
         }
 
       if (!found)
-      {
-        const auto object_type = rto->parameters().get<std::string>("_moose_warehouse_system_name");
-
-        mooseError(rto->errorPrefix(),
-                   " depends on ",
-                   object_type,
-                   " '",
-                   dep_name,
-                   "' but such an object does not exist");
-      }
+        rto->paramError("depends_on", "The object '", dep_name, "' does not exist");
     }
 }
 
@@ -382,14 +370,12 @@ RayTracingStudy::traceableMeshChecks()
     if (_fe_problem.adaptivity().isOn())
     {
       if (!TraceRayTools::isAdaptivityTraceableElem(elem))
-        mooseError(_error_prefix,
-                   ": Element type ",
+        mooseError("Element type ",
                    Utility::enum_to_string(elem->type()),
                    " is not supported in ray tracing with adaptivity");
     }
     else if (!TraceRayTools::isTraceableElem(elem))
-      mooseError(_error_prefix,
-                 ": Element type ",
+      mooseError("Element type ",
                  Utility::enum_to_string(elem->type()),
                  " is not supported in ray tracing");
   }
@@ -434,8 +420,7 @@ RayTracingStudy::internalSidesetSetup()
       continue;
 
     if (neighbor->subdomain_id() == elem->subdomain_id())
-      mooseError(_error_prefix,
-                 ":\n\nRayBCs exist on internal sidesets that are not bounded by a different",
[…]
