### 2021-02-17 78ffaa5
Improve ray_tracing error testing
refs #17030
