+++ modules/ray_tracing/test/src/userobjects/RayTracingStudyTest.C
@@ -18,11 +18,60 @@ RayTracingStudyTest::validParams()
 {
   auto params = RayTracingStudy::validParams();
 
-  params.addParam<bool>("add_duplicate_ray", "Add a duplicate Ray to the buffer");
-  params.addParam<bool>("add_local_non_unique_id_ray",
-                        "Add two Rays on each processor with the same ID");
-  params.addParam<bool>("add_global_non_unique_id_ray",
-                        "Add a Ray on each processor with the same ID");
+  params.addParam<bool>("add_duplicate_ray", false, "Add a duplicate Ray to the buffer");
+  params.addParam<bool>(
+      "add_local_non_unique_id_ray", false, "Add two Rays on each processor with the same ID");
+  params.addParam<bool>(
+      "add_global_non_unique_id_ray", false, "Add a Ray on each processor with the same ID");
+
+  params.addParam<bool>(
+      "ray_at_end_without_set", false, "Test Ray::atEnd() without the end being set");
+  params.addParam<bool>(
+      "ray_end_point_without_set", false, "Test Ray::endPoint() without the end being set");
+  params.addParam<bool>(
+      "ray_set_start_again", false, "Test setting a Ray's start point multiple times");
+  params.addParam<bool>("ray_set_start_fail_bbox",
+                        false,
+                        "Test setting a Ray's start point to a point that fails the bbox check");
+  params.addParam<bool>("ray_set_side_without_elem",
+                        false,
+                        "Test setting a Ray's incoming side without a starting element");
+  params.addParam<bool>(
+      "ray_set_invalid_side", false, "Test setting a Ray's incoming side to an invalid one");
+  params.addParam<bool>(
+      "ray_set_bad_side", false, "Test setting a Ray's incoming side to a wrong side");
+  params.addParam<bool>(
+      "ray_set_bad_start",
+      false,
+      "Test setting a Ray's start point to one that is not within the starting element");
+  params.addParam<bool>("ray_set_direction_before_start",
+                        false,
+                        "Test setting a Ray's direction before its start point");
+  params.addParam<bool>(
+      "ray_set_direction_again", false, "Test setting a Ray's direction multiple times");
+  params.addParam<bool>(
+      "ray_set_zero_direction", false, "Test setting a Ray's direction to the zero vector");
+  params.addParam<bool>(
+      "ray_set_end_before_start", false, "Test setting a Ray's end point before its start point");
+  params.addParam<bool>(
+      "ray_set_end_equal_start", false, "Test setting a Ray's end point to its start point");
+  params.addParam<bool>("ray_set_end_with_direction",
+                        false,
+                        "Test setting a Ray's end point after setting its direction");
+  params.addParam<bool>("ray_set_end_with_distance",
+                        false,
+                        "Test setting a Ray's end point after setting its max distance");
+  params.addParam<bool>("ray_set_end_fail_bbox",
+                        false,
+                        "Test setting a Ray's end point to a point that fails the bbox check");
+  params.addParam<bool>("ray_set_distance_before_start",
+                        false,
+                        "Test setting a Ray's max distance before its start point");
+  params.addParam<bool>(
[…]
