### 2021-02-17 bd03b17
INSFVMomentumGravity -> NSFVMomentumGravity
And make rho a material property such that this object can be used
equally well for both compressible and incompressible simulations
