### 2021-02-17 da8d4b0
XFEM geometric cut switching material system
Add SwitchingMaterial, ComboCutUserObject, GeometricCutSubdomainIDAux and tests

close #17041
