+++ modules/xfem/include/userobjects/MovingLineSegmentCutSetUserObject.h
@@ -34,6 +34,18 @@ public:
 
   virtual Real cutFraction(unsigned int cut_num, Real time) const override;
 
+  virtual GeometricCutSubdomainID getGeometricCutSubdomainID(const Node * node) const override;
+
+protected:
+  /// Calculate the signed distance function at a point
+  virtual Real calculateSignedDistance(Point p) const;
+
   /// Pointer to XFEMMovingInterfaceVelocityBase object
   const XFEMMovingInterfaceVelocityBase * _interface_velocity;
+
+  /// The GeometricCutSubdomainID for the negative side of the cut
+  const GeometricCutSubdomainID _negative_id;
+
+  /// The GeometricCutSubdomainID for the positive side of the cut
+  const GeometricCutSubdomainID _positive_id;
 };
