### 2021-02-17 3d7f5bf
Add SQA docs to RDG for use in applications
