### 2021-02-17 892da21
Handle issue/commit names with non-letters in MooseDocs sqa extension
refs #16868
