+++ framework/src/meshgenerators/MeshGenerator.C
@@ -18,7 +18,13 @@ MeshGenerator::validParams()
 {
   InputParameters params = MooseObject::validParams();
 
+  params.addParam<bool>("show_info",
+                        false,
+                        "Whether or not to show mesh info after generating the mesh "
+                        "(bounding box, element types, sidesets, nodesets, subdomains, etc)");
+
   params.registerBase("MeshGenerator");
+
   return params;
 }
 
@@ -74,4 +80,27 @@ MeshGenerator::buildDistributedMesh(unsigned int dim)
     mooseError("We need a MooseMesh object in order to build DistributedMesh objects through the "
                "buildDistributedMesh API");
   return _mesh->buildTypedMesh<DistributedMesh>(dim);
+}
+
+std::unique_ptr<MeshBase>
+MeshGenerator::generateInternal()
+{
+  auto mesh = generate();
+
+  if (getParam<bool>("show_info"))
+  {
+    const auto mesh_info = mesh->get_info(/* verbosity = */ 2);
+
+    // We will prefix all information with "type() 'name()':" because this could potentially
+    // output a ton of information and looks a bit better with a prefix
+    std::stringstream oss;
+    const auto split = MooseUtils::split(mesh_info, "\n");
+    if (split.size())
+      for (std::size_t i = 0; i < split.size() - 1; ++i) // ignore the last line break
+        oss << COLOR_CYAN << "" << type() << " '" << name() << "': " << COLOR_DEFAULT << split[i]
+            << std::endl;
+    _console << oss.str() << std::flush;
+  }
+
+  return mesh;
 }
