### 2021-02-17 6342d6d
make massflow boundary condition such that the inlet pressure drop is uniform
Refs #29206
