### 2021-02-17 80aa1b7
make massflow boundary condition such that the inlet pressure drop is uniform
Refs #29206
