### 2021-02-17 cd5eba5
create make install target
Also refactor find dep apps script and add a generic moose_test_runner
script for installation.

Fixes #10786.
