### 2021-02-17 a6bf6ca
add --tests flag to moose
