### 2021-02-18 c710ba4
Remove additional legacy disps. Refs #16982.
