+++ modules/tensor_mechanics/src/actions/PressureAction.C
@@ -23,10 +23,6 @@ PressureAction::validParams()
   params.addRequiredParam<std::vector<BoundaryName>>(
       "boundary", "The list of boundary IDs from the mesh where the pressure will be applied");
 
-  params.addParam<VariableName>("disp_x", "The x displacement");
-  params.addParam<VariableName>("disp_y", "The y displacement");
-  params.addParam<VariableName>("disp_z", "The z displacement");
-
   params.addParam<std::vector<VariableName>>(
       "displacements",
       "The displacements appropriate for the simulation geometry and coordinate system");
@@ -68,23 +64,7 @@ PressureAction::act()
 
   std::string kernel_name = ad_prepend + "Pressure";
 
-  std::vector<VariableName> displacements;
-  if (isParamValid("displacements"))
-    displacements = getParam<std::vector<VariableName>>("displacements");
-  else
-  {
-    // Legacy parameter scheme for displacements
-    if (!isParamValid("disp_x"))
-      mooseError("Specify displacement variables using the `displacements` parameter.");
-    displacements.push_back(getParam<VariableName>("disp_x"));
-
-    if (isParamValid("disp_y"))
-    {
-      displacements.push_back(getParam<VariableName>("disp_y"));
-      if (isParamValid("disp_z"))
-        displacements.push_back(getParam<VariableName>("disp_z"));
-    }
-  }
+  std::vector<VariableName> displacements = getParam<std::vector<VariableName>>("displacements");
 
   // Create pressure BCs
   for (unsigned int i = 0; i < displacements.size(); ++i)
