+++ modules/navier_stokes/src/fvkernels/INSFVAdvectionBase.C
@@ -1,0 +1,404 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#include "INSFVAdvectionBase.h"
+
+#include "INSFVPressureVariable.h"
+#include "INSFVVelocityVariable.h"
+#include "SystemBase.h"
+#include "MooseMesh.h"
+#include "FVDirichletBC.h"
+#include "INSFVFlowBC.h"
+#include "INSFVFullyDevelopedFlowBC.h"
+#include "INSFVNoSlipWallBC.h"
+#include "INSFVSlipWallBC.h"
+#include "INSFVSymmetryBC.h"
+#include "INSFVAttributes.h"
+#include "MooseUtils.h"
+
+#include "libmesh/dof_map.h"
+#include "libmesh/elem.h"
+#include "libmesh/numeric_vector.h"
+#include "libmesh/vector_value.h"
+
+#include <algorithm>
+
+std::unordered_map<const MooseApp *,
+                   std::vector<std::unordered_map<const Elem *, VectorValue<ADReal>>>>
+    INSFVAdvectionBase::_rc_a_coeffs;
+
+InputParameters
+INSFVAdvectionBase::validParams()
+{
+  InputParameters params = FVMatAdvection::validParams();
+  params.addRequiredCoupledVar("pressure", "The pressure variable.");
+  params.addRequiredCoupledVar("u", "The velocity in the x direction.");
+  params.addCoupledVar("v", "The velocity in the y direction.");
+  params.addCoupledVar("w", "The velocity in the z direction.");
+
+  MooseEnum velocity_interp_method("average rc", "rc");
+
+  params.addParam<MooseEnum>(
+      "velocity_interp_method",
+      velocity_interp_method,
+      "The interpolation to use for the velocity. Options are "
+      "'average' and 'rc' which stands for Rhie-Chow. The default is Rhie-Chow.");
+
+  params.addRequiredParam<MaterialPropertyName>("mu", "The viscosity");
+
+  // We need 2 ghost layers for the Rhie-Chow interpolation
+  params.set<unsigned short>("ghost_layers") = 2;
+
+  params.addClassDescription("Base class for advecting momentum, e.g. rho*u");
+
+  return params;
+}
+
+INSFVAdvectionBase::INSFVAdvectionBase(const InputParameters & params)
+  : FVMatAdvection(params),
+    _mu_elem(getADMaterialProperty<Real>("mu")),
+    _mu_neighbor(getNeighborADMaterialProperty<Real>("mu")),
+    _p_var(dynamic_cast<const INSFVPressureVariable *>(getFieldVar("pressure", 0))),
+    _u_var(dynamic_cast<const INSFVVelocityVariable *>(getFieldVar("u", 0))),
+    _v_var(params.isParamValid("v")
+               ? dynamic_cast<const INSFVVelocityVariable *>(getFieldVar("v", 0))
+               : nullptr),
+    _w_var(params.isParamValid("w")
+               ? dynamic_cast<const INSFVVelocityVariable *>(getFieldVar("w", 0))
+               : nullptr),
+    _dim(_subproblem.mesh().dimension())
+{
+#ifndef MOOSE_GLOBAL_AD_INDEXING
+  mooseError("INSFV is not supported by local AD indexing. In order to use INSFV, please run the "
+             "configure script in the root MOOSE directory with the configure option "
+             "'--with-ad-indexing-type=global'");
+#endif
+
+  if (!_p_var)
[…]

[…]
