### 2021-02-18 470fc7c
Add multi block deletion to BlockDeletionGenerator #17052
