+++ framework/src/meshgenerators/PatternedMeshGenerator.C
@@ -70,21 +70,13 @@ PatternedMeshGenerator::PatternedMeshGenerator(const InputParameters & parameter
                        " is larger than the the maximum possible index, which is determined by the "
                        "number of MeshGenerators provided in inputs");
 
-  _mesh_ptrs.reserve(_input_names.size());
-  for (auto & input_name : _input_names)
-    _mesh_ptrs.push_back(&getMeshByName(input_name));
-}
-
-std::unique_ptr<MeshBase>
-PatternedMeshGenerator::generate()
-{
   // Reserve spaces for all the meshes
   _meshes.reserve(_input_names.size());
 
   // Read in all of the meshes
   for (MooseIndex(_input_names) i = 0; i < _input_names.size(); ++i)
   {
-    std::unique_ptr<ReplicatedMesh> mesh = dynamic_pointer_cast<ReplicatedMesh>(*_mesh_ptrs[i]);
+    std::unique_ptr<ReplicatedMesh> mesh = dynamic_pointer_cast<ReplicatedMesh>(getMeshByName(_input_names[i]));
     if (!mesh)
       paramError("inputs",
                  "The input mesh '",
@@ -95,7 +87,11 @@ PatternedMeshGenerator::generate()
                  "Try running without distributed mesh.");
     _meshes.push_back(std::move(mesh));
   }
+}
 
+std::unique_ptr<MeshBase>
+PatternedMeshGenerator::generate()
+{
   // Data structure that holds each row
   _row_meshes.resize(_pattern.size());
 
@@ -142,7 +138,6 @@ PatternedMeshGenerator::generate()
       // If this is the first cell of the row initialize the row mesh
       if (j == 0)
       {
-        //_row_meshes[i] = _mesh_ptrs[_pattern[i][j]]->clone();
         auto clone = _meshes[_pattern[i][j]]->clone();
         _row_meshes[i] = dynamic_pointer_cast<ReplicatedMesh>(clone);
 
