+++ framework/include/meshgenerators/SubdomainIDGenerator.h
@@ -30,9 +30,9 @@ public:
   std::unique_ptr<MeshBase> generate() override;
 
 protected:
+  /// the input mesh, which may be output by another mesh generator
   std::unique_ptr<MeshBase> & _input;
 
   /// The subdomain ID to assign to every elemennt
   SubdomainID _subdomain_id;
 };
-

[…]
