### 2021-02-18 6898a49
Add missing class descriptions (#24)
