### 2021-02-18 4b95766
Add missing class descriptions (#21090)
