+++ modules/electromagnetics/src/bcs/CoupledGradBC.C
@@ -8,14 +8,16 @@ InputParameters
 validParams<CoupledGradBC>()
 {
   InputParameters params = validParams<IntegratedBC>();
+  params.addClassDescription(
+      "Imposes the integrated boundary condition arising from integration by parts of a Helmholtz "
+      "equation, when that term is set equal to the gradient of a coupled variable. ");
   params.addParam<Real>(
       "sign",
       1.0,
       "Sign of the coupled gradient term in the weak form (1.0, -1.0, positive default)");
-  params.addParam<Real>(
-      "coefficient", 1.0, "Optional coefficient multiplier for coupled gradient term.");
+  params.addParam<Real>("coefficient", 1.0, "Optional coefficient for coupled gradient term.");
   params.addParam<FunctionName>(
-      "func", 1.0, "Optional function multiplier for coupled gradient term.");
+      "func", 1.0, "Optional function coefficient for coupled gradient term.");
   params.addRequiredCoupledVar("coupled_field", "Coupled field variable.");
   return params;
 }
