+++ modules/electromagnetics/src/bcs/ReflectionBC.C
@@ -10,7 +10,8 @@ InputParameters
 validParams<ReflectionBC>()
 {
   InputParameters params = validParams<IntegratedBC>();
-
+  params.addClassDescription("First order Reflection BC based on 'The Finite Element Method in "
+                             "Electromagnetics' by JM Jin for scalar variables.");
   params.addRequiredCoupledVar("field_real", "Real component of field.");
   params.addRequiredCoupledVar("field_imaginary", "Imaginary component of field.");
   MooseEnum component("real imaginary");
@@ -46,7 +47,7 @@ ReflectionBC::computeQpResidual()
 
   std::complex<double> field(_field_real[_qp], _field_imag[_qp]);
   std::complex<double> func(_func_real.value(_t, _q_point[_qp]),
-                             _func_imag.value(_t, _q_point[_qp]));
+                            _func_imag.value(_t, _q_point[_qp]));
   std::complex<double> coeff(_coeff_real, _coeff_imag);
   std::complex<double> jay(0, 1);
 
