### 2021-02-19 feb47ce
Add missing requirements, design, and issues entries in test specs (#24)
