+++ modules/electromagnetics/doc/content/benchmarks/DipoleAntenna.md
@@ -1,0 +1,3 @@
+# Dipole Antenna Benchmark
+
+This still needs to be documented. Please bug Casey Icenhour (casey.icenhour@inl.gov) to do this!
