### 2021-02-19 154c664
Allow upwinding a vector
