### 2021-02-19 41639c5
Make setDofValue for FV variables populate _u and _ad_u
This makes it so that if one FV aux variable is used in the computation
of another FV aux variable, then the used aux variable will have the
correct value for its solution
