### 2021-02-19 6d08c3c
Introduce mat advection kernel/bc that applies same interpolation scheme to entire flux
