+++ modules/navier_stokes/src/materials/ConservedVarMaterial.C
@@ -17,48 +17,50 @@
 // MOOSE includes
 #include "AuxiliarySystem.h"
 
-registerADMooseObject("NavierStokesApp", ConservedVarMaterial);
+registerMooseObject("NavierStokesApp", ConservedVarMaterial);
 
-namespace nms = NS;
-
-defineADValidParams(
-    ConservedVarMaterial,
-    VarMaterialBase,
-    params.addRequiredCoupledVar(nms::density, "density");
-    params.addRequiredCoupledVar(nms::total_energy_density, "total fluid energy");
-    params.addRequiredCoupledVar(nms::momentum_x, "rhou");
-    params.addCoupledVar(nms::momentum_y, "rhov");
-    params.addCoupledVar(nms::momentum_z, "rhow");
-    params.addCoupledVar(
-        nms::T_fluid,
-        "optionally use for approximating second spatial derivatives for stabilization");
-    params.addClassDescription("Provides access to variables for a conserved variable set "
-      "of density, total fluid energy, and momentum"););
+InputParameters
+ConservedVarMaterial::validParams()
+{
+  auto params = VarMaterialBase::validParams();
+  params.addRequiredCoupledVar(NS::density, "density");
+  params.addRequiredCoupledVar(NS::total_energy_density, "total fluid energy");
+  params.addRequiredCoupledVar(NS::momentum_x, "rhou");
+  params.addCoupledVar(NS::momentum_y, "rhov");
+  params.addCoupledVar(NS::momentum_z, "rhow");
+  params.addCoupledVar(
+      NS::T_fluid,
+      "optionally use for approximating second spatial derivatives for stabilization");
+  params.addClassDescription("Provides access to variables for a conserved variable set "
+    "of density, total fluid energy, and momentum");
+  params.addParam<MaterialPropertyName>(NS::porosity, "the porosity");
+  return params;
+}
 
 using MetaPhysicL::raw_value;
 
 ConservedVarMaterial::ConservedVarMaterial(const InputParameters & params)
   : VarMaterialBase(params),
-    _var_rho(adCoupledValue(nms::density)),
-    _var_total_energy_density(adCoupledValue(nms::total_energy_density)),
-    _var_rho_u(adCoupledValue(nms::momentum_x)),
-    _var_rho_v(isCoupled(nms::momentum_y) ? adCoupledValue(nms::momentum_y) : _ad_zero),
-    _var_rho_w(isCoupled(nms::momentum_z) ? adCoupledValue(nms::momentum_z) : _ad_zero),
-    _var_grad_rho(adCoupledGradient(nms::density)),
-    _var_grad_rho_et(adCoupledGradient(nms::total_energy_density)),
-    _var_grad_rho_u(adCoupledGradient(nms::momentum_x)),
-    _var_grad_rho_v(isCoupled(nms::momentum_y) ? adCoupledGradient(nms::momentum_y) : _ad_grad_zero),
-    _var_grad_rho_w(isCoupled(nms::momentum_z) ? adCoupledGradient(nms::momentum_z) : _ad_grad_zero),
-    _var_grad_grad_rho(adCoupledSecond(nms::density)),
-    _var_grad_grad_rho_u(adCoupledSecond(nms::momentum_x)),
-    _var_grad_grad_rho_v(isCoupled(nms::momentum_y) ? adCoupledSecond(nms::momentum_y) : _ad_second_zero),
-    _var_grad_grad_rho_w(isCoupled(nms::momentum_z) ? adCoupledSecond(nms::momentum_z) : _ad_second_zero),
[…]
