### 2021-02-19 bdfecaf
Modernize var materials
