### 2021-02-19 afd2da9
Remove porosity derivative term
