+++ modules/navier_stokes/include/fvkernels/PINSFVEnergyTimeDerivative.h
@@ -27,7 +27,7 @@ protected:
   /// the porosity
   const VariableValue & _eps;
   /// the time derivative of the porosity
-  const VariableValue & _eps_dot;
+  // const VariableValue & _eps_dot;
   /// whether this kernel is being used for a solid or a fluid temperature
   const bool _is_solid;
 };
