### 2021-02-19 b2a0840
Remove porosity derivative term
