### 2021-02-19 4f49d30
Refactoring the solver class
- Creating a parent class with the basic building blocks of the solver algorithm.
- Creating 2 child classes: 1 for liquid metal, 1 for liquid water

Refs #0
