+++ modules/subchannel/src/problems/LiquidWaterSubChannel1PhaseProblem.C
@@ -1,0 +1,42 @@
+#include "LiquidWaterSubChannel1PhaseProblem.h"
+
+registerMooseObject("SubChannelApp", LiquidWaterSubChannel1PhaseProblem);
+
+InputParameters
+LiquidWaterSubChannel1PhaseProblem::validParams()
+{
+  InputParameters params = SubChannel1PhaseProblemBase::validParams();
+  return params;
+}
+
+LiquidWaterSubChannel1PhaseProblem::LiquidWaterSubChannel1PhaseProblem(
+    const InputParameters & params)
+  : SubChannel1PhaseProblemBase(params)
+{
+}
+
+double
+LiquidWaterSubChannel1PhaseProblem::computeFrictionFactor(double Re)
+{
+  double a, b;
+  if (Re < 1)
+  {
+    return 64.0;
+  }
+  else if (Re >= 1 and Re < 5000)
+  {
+    a = 64.0;
+    b = -1.0;
+  }
+  else if (Re >= 5000 and Re < 30000)
+  {
+    a = 0.316;
+    b = -0.25;
+  }
+  else
+  {
+    a = 0.184;
+    b = -0.20;
+  }
+  return a * std::pow(Re, b);
+}
