+++ modules/subchannel/src/problems/LiquidMetalSubChannel1PhaseProblem.C
@@ -1,0 +1,23 @@
+#include "LiquidMetalSubChannel1PhaseProblem.h"
+
+registerMooseObject("SubChannelApp", LiquidMetalSubChannel1PhaseProblem);
+
+InputParameters
+LiquidMetalSubChannel1PhaseProblem::validParams()
+{
+  InputParameters params = SubChannel1PhaseProblemBase::validParams();
+  return params;
+}
+
+LiquidMetalSubChannel1PhaseProblem::LiquidMetalSubChannel1PhaseProblem(
+    const InputParameters & params)
+  : SubChannel1PhaseProblemBase(params)
+{
+}
+
+double
+LiquidMetalSubChannel1PhaseProblem::computeFrictionFactor(double /*Re*/)
+{
+  // FIXME: implement this
+  return 0;
+}
