+++ modules/subchannel/src/problems/SubChannel1PhaseProblemBase.C
@@ -1,4 +1,4 @@
-#include "SubChannel1PhaseProblem.h"
+#include "SubChannel1PhaseProblemBase.h"
 #include "SystemBase.h"
 #include "libmesh/petsc_vector.h"
 #include <petscdm.h>
@@ -13,17 +13,15 @@
 #include "SolutionHandle.h"
 #include "AuxiliarySystem.h"
 
-registerMooseObject("SubChannelApp", SubChannel1PhaseProblem);
-
 InputParameters
-SubChannel1PhaseProblem::validParams()
+SubChannel1PhaseProblemBase::validParams()
 {
   InputParameters params = ExternalProblem::validParams();
   params.addRequiredParam<UserObjectName>("fp", "Fluid properties user object name");
   return params;
 }
 
-SubChannel1PhaseProblem::SubChannel1PhaseProblem(const InputParameters & params)
+SubChannel1PhaseProblemBase::SubChannel1PhaseProblemBase(const InputParameters & params)
   : ExternalProblem(params),
     _g_grav(9.87),
     _subchannel_mesh(dynamic_cast<SubChannelMeshBase &>(_mesh)),
@@ -45,7 +43,7 @@ SubChannel1PhaseProblem::SubChannel1PhaseProblem(const InputParameters & params)
 }
 
 void
-SubChannel1PhaseProblem::initialSetup()
+SubChannel1PhaseProblemBase::initialSetup()
 {
   ExternalProblem::initialSetup();
   // Read in Wij_global for null transient only at the first run of externalSolve
@@ -90,7 +88,7 @@ SubChannel1PhaseProblem::initialSetup()
   q_prime_soln = new SolutionHandle(getVariable(0, "q_prime"));
 }
 
-SubChannel1PhaseProblem::~SubChannel1PhaseProblem()
+SubChannel1PhaseProblemBase::~SubChannel1PhaseProblemBase()
 {
   delete mdot_soln;
   delete SumWij_soln;
@@ -105,39 +103,13 @@ SubChannel1PhaseProblem::~SubChannel1PhaseProblem()
 }
 
 bool
-SubChannel1PhaseProblem::converged()
+SubChannel1PhaseProblemBase::converged()
 {
   return true;
 }
 
-double
-SubChannel1PhaseProblem::computeFF(double Re)
-{
-  double a, b;
-  if (Re < 1)
-  {
-    return 64.0;
-  }
-  else if (Re >= 1 and Re < 5000)
-  {
-    a = 64.0;
-    b = -1.0;
-  }
-  else if (Re >= 5000 and Re < 30000)
-  {
-    a = 0.316;
-    b = -0.25;
-  }
-  else
-  {
-    a = 0.184;
-    b = -0.20;
-  }
-  return a * std::pow(Re, b);
-}
-
 void
-SubChannel1PhaseProblem::computeWij(int iz)
+SubChannel1PhaseProblemBase::computeWij(int iz)
 {
   if (iz == 0)
   {
@@ -267,7 +239,7 @@ SubChannel1PhaseProblem::computeWij(int iz)
 }
 
 void
-SubChannel1PhaseProblem::computeSumWij(int iz)
+SubChannel1PhaseProblemBase::computeSumWij(int iz)
 {
   if (iz == 0)
   {
@@ -293,7 +265,7 @@ SubChannel1PhaseProblem::computeSumWij(int iz)
 }
 
 void
-SubChannel1PhaseProblem::computeMdot(int iz)
+SubChannel1PhaseProblemBase::computeMdot(int iz)
 {
   if (iz == 0)
   {
[…]
