### 2021-02-19 53366b5
Small fixes due to rebase
(refs #14830)
