### 2021-02-19 7dab401
Use coupledValues for BuildArrayVariableAux
