### 2021-02-19 68ee0fd
Create NSFVPorosityMatAdvection(OutflowBC)
This object is used in momentum equation formulations in which the real
velocity has been replaced by the superficial velocity, e.g. u -> u_d
