### 2021-02-21 1698ad0
Add more ray_tracing coverage
refs #16028
