+++ modules/ray_tracing/test/include/userobjects/RayTracingStudyTest.h
@@ -20,4 +20,6 @@ public:
 
 protected:
   void generateRays() override;
+
+  void postExecuteStudy() override;
 };
