### 2021-02-21 #3031 @arfon
@jedbrown - I realize you're not able to devote a bunch of time to editing for JOSS right now but would you be willing to take this one on sometime in early March. @WilkAndy - many of our editors are very busy right now and so there may be some additional delays before we send this submission out for review.
