### 2021-02-21 4c3f2fc
Fix some porous kernels
