+++ modules/navier_stokes/src/fvkernels/PINSFVEnergyTimeDerivative.C
@@ -23,7 +23,7 @@ PINSFVEnergyTimeDerivative::validParams()
   params.declareControllable("rho");
   params.addParam<MaterialPropertyName>("cp_name", "cp", "The name of the specific heat capacity");
   params.addRequiredCoupledVar("porosity", "Porosity variable");
-  params.addRequiredParam<bool>("is_solid", "Kernel for the solid temperature ?");
+  params.addRequiredParam<bool>("is_solid", "Whether this kernel acts on the solid temperature");
   return params;
 }
 
@@ -32,7 +32,6 @@ PINSFVEnergyTimeDerivative::PINSFVEnergyTimeDerivative(const InputParameters & p
   _rho(getParam<Real>("rho")),
   _cp(getADMaterialProperty<Real>("cp_name")),
   _eps(coupledValue("porosity")),
-  // _eps_dot(coupledDot("porosity")),
   _is_solid(getParam<bool>("is_solid"))
 {
 }
@@ -40,9 +39,6 @@ PINSFVEnergyTimeDerivative::PINSFVEnergyTimeDerivative(const InputParameters & p
 ADReal
 PINSFVEnergyTimeDerivative::computeQpResidual()
 {
-  //FIXME Neglecting time derivative of _cp
-  //FIXME Possibly use a VarMat instead of querying variables as parameters
-  //FIXME Separate derivative of porosity since it's uncommon to need it
-  return (_is_solid ? 1 - _eps[_qp] : _eps[_qp]) * _rho * _cp[_qp] * FVTimeKernel::computeQpResidual();// +
-     // (_is_solid ? - _eps_dot[_qp] : _eps_dot[_qp]) * _rho * _cp[_qp] * _u[_qp];
+  // Note: This is neglecting time derivative of both _cp and _rho
+  return (_is_solid ? 1 - _eps[_qp] : _eps[_qp]) * _rho * _cp[_qp] * FVTimeKernel::computeQpResidual();
 }
