### 2021-02-21 3efecf4
Fix some porous kernels
