### 2021-02-21 27a9599
Fix some porous kernels
