+++ modules/navier_stokes/src/fvkernels/PINSFVEnergyConvection.C
@@ -6,10 +6,11 @@ InputParameters
 PINSFVEnergyConvection::validParams()
 {
   InputParameters params = FVElementalKernel::validParams();
-  params.addClassDescription("Implements the solid-fluid convection term.");
-  params.addRequiredCoupledVar("porosity", "Porosity variable");
-  params.addRequiredParam<MaterialPropertyName>("h_solid_fluid", "Name of the convective heat transfer coefficient");
-  params.addRequiredParam<bool>("is_solid", "Kernel for the fluid or solid temperature ?");
+  params.addClassDescription("Implements the solid-fluid convection term in the porous "
+                             "media Navier Stokes energy equation.");
+  params.addRequiredParam<MaterialPropertyName>("h_solid_fluid", "Name of the convective heat "
+      "transfer coefficient. This coefficient should include the influence of porosity.");
+  params.addRequiredParam<bool>("is_solid", "Whether this kernel acts on the solid temperature");
   params.addRequiredCoupledVar("temp_fluid", "Fluid temperature");
   params.addRequiredCoupledVar("temp_solid", "Solid temperature");
   return params;
@@ -17,7 +18,6 @@ PINSFVEnergyConvection::validParams()
 
 PINSFVEnergyConvection::PINSFVEnergyConvection(const InputParameters & parameters)
   : FVElementalKernel(parameters),
-  _eps(coupledValue("porosity")),
   _h_solid_fluid(getADMaterialProperty<Real>("h_solid_fluid")),
   _temp_fluid(adCoupledValue("temp_fluid")),
   _temp_solid(adCoupledValue("temp_solid")),
@@ -31,5 +31,5 @@ PINSFVEnergyConvection::computeQpResidual()
   if (_is_solid)
     return -_h_solid_fluid[_qp] * (_temp_fluid[_qp] - _temp_solid[_qp]);
   else
-    return _eps[_qp] * _h_solid_fluid[_qp] * (_temp_fluid[_qp] - _temp_solid[_qp]);
+    return _h_solid_fluid[_qp] * (_temp_fluid[_qp] - _temp_solid[_qp]);
 }
