### 2021-02-22 b30ac77
fix sample results transfer for two or more subapps ref #17079
