### 2021-02-22 39257b3
Implement straight upwinding including pressure at interfaces
This achieves the correct result for a step porosity change on a
compressible flow problem. I verified against hand calculations. For a
STP gas, momentum conservation will be determined entirely from the
change in pressure. Consequently the pressure must double if the
porosity halves. Plugging this result into the energy equation you find
that that Darcy velocity must halve in order to have energy
conservation. And then finally from the mass equation this then requires
that the density double.

Currently I'm having an issue with a t
[…]
