### 2021-02-22 4d46fe8
Add SQA docs to ray tracing
(refs #15968)
