+++ modules/ray_tracing/doc/content/sqa/ray_tracing_far.md
@@ -1,0 +1,1 @@
+!template load file=sqa/app_far.md.template app=RayTracingApp category=ray_tracing
