+++ modules/ray_tracing/doc/sqa_reports.yml
@@ -1,9 +1,9 @@
 Applications:
     ray_tracing:
         exe_directory: ${MOOSE_DIR}/modules/ray_tracing
-        app_types:
-            - RayTracingApp
         content_directory: ${MOOSE_DIR}/modules/ray_tracing/doc/content
+        remove:
+            - ${MOOSE_DIR}/framework/doc/remove.yml
 
 Documents:
     software_requirements_specification: sqa/ray_tracing_srs.md
@@ -11,23 +11,13 @@ Documents:
     software_test_plan: sqa/ray_tracing_stp.md
     requirements_traceablity_matrix: sqa/ray_tracing_rtm.md
     verification_validation_plan: sqa/ray_tracing_vvr.md
-    log_safety_software_determination: WARNING
-    log_quality_level_determination: WARNING
-    log_enterprise_architecture_entry: WARNING
-    log_project_management_plan: WARNING
-    log_software_quality_plan: WARNING
-    log_configuration_management_plan: WARNING
-    log_asset_management_plan: WARNING
-    log_verification_validation_plan: WARNING
-    log_user_manual: WARNING
-    log_theory_manual: WARNING
-    log_software_requirements_specification: WARNING
-    log_software_design_description: WARNING
-    log_software_test_plan: WARNING
-    log_requirements_traceablity_matrix: WARNING
+    failure_analysis_report: sqa/ray_tracing_far.md
+    log_default: WARNING
+    show_warning: false
 
 Requirements:
     ray_tracing:
         directories:
-            - ${MOOSE_DIR}/modules/ray_tracing
-        log_testable: WARNING
+            - ${ROOT_DIR}/test
+        log_default: WARNING
+        show_warning: false
