+++ modules/navier_stokes/src/materials/ConservedVarMaterial.C
@@ -29,10 +29,9 @@ ConservedVarMaterial::validParams()
   params.addCoupledVar(NS::momentum_y, "rhov");
   params.addCoupledVar(NS::momentum_z, "rhow");
   params.addCoupledVar(
-      NS::T_fluid,
-      "optionally use for approximating second spatial derivatives for stabilization");
+      NS::T_fluid, "optionally use for approximating second spatial derivatives for stabilization");
   params.addClassDescription("Provides access to variables for a conserved variable set "
-    "of density, total fluid energy, and momentum");
+                             "of density, total fluid energy, and momentum");
   params.addParam<MaterialPropertyName>(NS::porosity, "the porosity");
   params.addParam<bool>("velocity_is_superficial",
                         "Whether u in rho_u corresponds to a superficial velocity or a primitive "
@@ -56,34 +55,39 @@ ConservedVarMaterial::ConservedVarMaterial(const InputParameters & params)
     _var_grad_rho_w(isCoupled(NS::momentum_z) ? adCoupledGradient(NS::momentum_z) : _ad_grad_zero),
     _var_grad_grad_rho(adCoupledSecond(NS::density)),
     _var_grad_grad_rho_u(adCoupledSecond(NS::momentum_x)),
-    _var_grad_grad_rho_v(isCoupled(NS::momentum_y) ? adCoupledSecond(NS::momentum_y) : _ad_second_zero),
-    _var_grad_grad_rho_w(isCoupled(NS::momentum_z) ? adCoupledSecond(NS::momentum_z) : _ad_second_zero),
+    _var_grad_grad_rho_v(isCoupled(NS::momentum_y) ? adCoupledSecond(NS::momentum_y)
+                                                   : _ad_second_zero),
+    _var_grad_grad_rho_w(isCoupled(NS::momentum_z) ? adCoupledSecond(NS::momentum_z)
+                                                   : _ad_second_zero),
     _var_rho_dot(_is_transient ? adCoupledDot(NS::density) : _ad_zero),
     _var_rho_et_dot(_is_transient ? adCoupledDot(NS::total_energy_density) : _ad_zero),
     _var_rho_u_dot(_is_transient ? adCoupledDot(NS::momentum_x) : _ad_zero),
     _var_rho_v_dot(isCoupled(NS::momentum_y) && _is_transient ? adCoupledDot(NS::momentum_y)
-                                                          : _ad_zero),
+                                                              : _ad_zero),
     _var_rho_w_dot(isCoupled(NS::momentum_z) && _is_transient ? adCoupledDot(NS::momentum_z)
-                                                          : _ad_zero),
-    _var_grad_grad_T_fluid(isCoupled(NS::T_fluid) ? adCoupledSecond(NS::T_fluid)
-                                                   : _ad_second_zero),
+                                                              : _ad_zero),
+    _var_grad_grad_T_fluid(isCoupled(NS::T_fluid) ? adCoupledSecond(NS::T_fluid) : _ad_second_zero),
     _have_porosity(isParamValid(NS::porosity)),
     _epsilon(_have_porosity ? &getMaterialProperty<Real>(NS::porosity) : nullptr),
[…]
