### 2021-02-23 4e64381
Remove the restriction on Python 3.8
Not really understanding the syntax. I have found issues with Python 3.8 and had
hoped to skip supporting that version all together... but I don't see how that is
possible given the following is being refused by conda:

  >=3.6,<3.8,>=3.9

Maybe order doesn't matter. Or there exists a logic operator (!) I am not familiar
with...

Refs #17105
