+++ modules/thm/doc/content/sqa/thm_rtm.md
@@ -1,0 +1,1 @@
+!template load file=sqa/app_rtm.md.template app=THMApp category=thm
