+++ modules/navier_stokes/src/fvkernels/NSFVPorosityMatAdvection.C
@@ -31,31 +31,23 @@ NSFVPorosityMatAdvection::NSFVPorosityMatAdvection(const InputParameters & param
 ADReal
 NSFVPorosityMatAdvection::computeQpResidual()
 {
-  // Check for large porosity change. If we do have a large change we are going to upwind everything
-  // to prevent oscillations due to the discontinuity
-  if (std::abs(_eps_elem[_qp] - _eps_neighbor[_qp]) > 0.1)
+  ADReal u_interface;
+  Real elem_coeff, neighbor_coeff;
+
+  interpolate(
+      Moose::FV::InterpMethod::Average, _v, _vel_elem[_qp], _vel_neighbor[_qp], *_face_info, true);
+  if (MetaPhysicL::raw_value(_v) * _face_info->normal() > 0)
   {
-    ADReal u_interface;
-
-    interpolate(Moose::FV::InterpMethod::Average,
-                _v,
-                _vel_elem[_qp],
-                _vel_neighbor[_qp],
-                *_face_info,
-                true);
-    if (MetaPhysicL::raw_value(_v) * _face_info->normal() > 0)
-    {
-      _v = _vel_elem[_qp];
-      u_interface = _adv_quant_elem[_qp];
-    }
-    else
-    {
-      _v = _vel_neighbor[_qp];
-      u_interface = _adv_quant_neighbor[_qp];
-    }
-
-    return _normal * _v * u_interface;
+    elem_coeff = 1;
+    neighbor_coeff = 0;
   }
   else
-    return FVMatAdvection::computeQpResidual();
+  {
+    elem_coeff = 0;
+    neighbor_coeff = 1;
+  }
+  _v = _vel_elem[_qp] * elem_coeff + _vel_neighbor[_qp] * neighbor_coeff;
+  u_interface = _adv_quant_elem[_qp] * elem_coeff + _adv_quant_neighbor[_qp] * neighbor_coeff;
+
+  return _normal * _v * u_interface;
 }
