+++ modules/navier_stokes/src/materials/PorousConservedVarMaterial.C
@@ -53,7 +53,8 @@ PorousConservedVarMaterial::PorousConservedVarMaterial(const InputParameters & p
     _v(declareADProperty<Real>(NS::v)),
     _specific_internal_energy(declareADProperty<Real>(NS::specific_internal_energy)),
     _pressure(declareADProperty<Real>(NS::pressure)),
-    _specific_total_enthalpy(declareADProperty<Real>(NS::specific_total_enthalpy))
+    _specific_total_enthalpy(declareADProperty<Real>(NS::specific_total_enthalpy)),
+    _rho_ht(declareADProperty<Real>(NS::total_enthalpy_density))
 {
 }
 
@@ -77,4 +78,5 @@ PorousConservedVarMaterial::computeQpProperties()
   _pressure[_qp] = _fluid.p_from_v_e(_v[_qp], _specific_internal_energy[_qp]);
 
   _specific_total_enthalpy[_qp] = (_total_energy_density[_qp] + _pressure[_qp]) / _rho[_qp];
+  _rho_ht[_qp] = _specific_total_enthalpy[_qp] * _rho[_qp];
 }
