+++ modules/navier_stokes/include/materials/PorousConservedVarMaterial.h
@@ -40,4 +40,5 @@ protected:
   ADMaterialProperty<Real> & _specific_internal_energy;
   ADMaterialProperty<Real> & _pressure;
   ADMaterialProperty<Real> & _specific_total_enthalpy;
+  ADMaterialProperty<Real> & _rho_ht;
 };
