+++ modules/thm/src/components/FlowChannelBase.C
@@ -1,13 +1,5 @@
 #include "FlowChannelBase.h"
-#include "FlowModelSinglePhase.h"
-#include "FlowModelTwoPhase.h"
-#include "FlowModelTwoPhaseNCG.h"
-#include "TwoPhaseFluidProperties.h"
-#include "VaporMixtureFluidProperties.h"
-#include "TwoPhaseNCGFluidProperties.h"
 #include "HeatTransferBase.h"
-#include "HeatTransfer1PhaseBase.h"
-#include "HeatTransfer2PhaseBase.h"
 #include "ClosuresBase.h"
 #include "THMApp.h"
 #include "THMMesh.h"
