+++ modules/thm/include/bcs/OneDEnergyHRhoUBC.h
@@ -16,17 +16,12 @@ protected:
   virtual Real computeQpJacobian();
   virtual Real computeQpOffDiagJacobian(unsigned int jvar);
 
-  /// Volume fraction
-  const MaterialProperty<Real> & _alpha;
-  const MaterialProperty<Real> * const _dalpha_dbeta;
   /// Specified enthalpy
   const Real & _H;
   /// Specified momentum
   const Real & _rhou;
   /// Cross-sectional area
   const VariableValue & _area;
-
-  unsigned int _beta_var_num;
 
 public:
   static InputParameters validParams();
