+++ modules/thm/src/components/HeatTransferFromHeatFlux1Phase.C
@@ -1,6 +1,5 @@
 #include "HeatTransferFromHeatFlux1Phase.h"
 #include "FlowModelSinglePhase.h"
-#include "FlowModelTwoPhase.h"
 
 registerMooseObject("THMApp", HeatTransferFromHeatFlux1Phase);
 

[…]
