+++ modules/navier_stokes/doc/content/source/fvkernels/INSFVMixingLengthScalarDiffusion.md
@@ -1,4 +1,4 @@
-# INSFVMixingLengthReynoldsStress
+# INSFVMixingLengthScalarDiffusion
 
 This kernel uses a mixing-length model to compute the turbulent diffusion of a
 passive scalar, $\nabla \cdot \overline{ c' \vec u'}$, which appears in
@@ -8,8 +8,8 @@ Example passive scalars include energy and trace chemicals. In the case of
 energy, the turbulent Schmidt number parameter is actually referred to as the
 turbulent Prandtl number.
 
-!syntax parameters /FVKernels/INSFVMixingLengthReynoldsStress
+!syntax parameters /FVKernels/INSFVMixingLengthScalarDiffusion
 
-!syntax inputs /FVKernels/INSFVMixingLengthReynoldsStress
+!syntax inputs /FVKernels/INSFVMixingLengthScalarDiffusion
 
-!syntax children /FVKernels/INSFVMixingLengthReynoldsStress
+!syntax children /FVKernels/INSFVMixingLengthScalarDiffusion
