### 2021-02-23 059629a
Remove test object stub documentation files and correct registration for test objects (#24)
