+++ modules/electromagnetics/doc/content/source/functions/MMSTestFunc.md
@@ -1,23 +1,0 @@
-# MMSTestFunc
-
-!alert construction title=Undocumented Class
-The MMSTestFunc has not been documented. The content listed below should be used as a starting point for
-documenting the class, which includes the typical automatic documentation associated with a
-MooseObject; however, what is contained is ultimately determined by what is necessary to make the
-documentation clear for users.
-
-!syntax description /Functions/MMSTestFunc
-
-## Overview
-
-!! Replace these lines with information regarding the MMSTestFunc object.
-
-## Example Input File Syntax
-
-!! Describe and include an example of how to use the MMSTestFunc object.
-
-!syntax parameters /Functions/MMSTestFunc
-
-!syntax inputs /Functions/MMSTestFunc
-
-!syntax children /Functions/MMSTestFunc
