### 2021-02-23 e86992e
Remove test object stub documentation files and correct registration for test objects (#21090)
