+++ modules/electromagnetics/doc/content/source/functions/ElectricalContactTestFunc.md
@@ -1,23 +1,0 @@
-# ElectricalContactTestFunc
-
-!alert construction title=Undocumented Class
-The ElectricalContactTestFunc has not been documented. The content listed below should be used as a starting point for
-documenting the class, which includes the typical automatic documentation associated with a
-MooseObject; however, what is contained is ultimately determined by what is necessary to make the
-documentation clear for users.
-
-!syntax description /Functions/ElectricalContactTestFunc
-
-## Overview
-
-!! Replace these lines with information regarding the ElectricalContactTestFunc object.
-
-## Example Input File Syntax
-
-!! Describe and include an example of how to use the ElectricalContactTestFunc object.
-
-!syntax parameters /Functions/ElectricalContactTestFunc
-
-!syntax inputs /Functions/ElectricalContactTestFunc
-
-!syntax children /Functions/ElectricalContactTestFunc
