### 2021-02-23 600085b
Adding a Clone transfer
Mainly for testing, but could be useful in the future.

(refs #16166)
