### 2021-02-23 3034d35
Set the context of the ParallelStudy to itself
refs #17112
