### 2021-02-23 041091b
Set the context of the ParallelStudy to itself
refs #17112
