+++ modules/ray_tracing/include/utils/ParallelStudy.h
@@ -25,7 +25,7 @@
 // libMesh Includes
 #include "libmesh/parallel_object.h"
 
-template <typename WorkType, typename ParallelDataType, typename Context>
+template <typename WorkType, typename ParallelDataType>
 class ParallelStudy : public ParallelObject
 {
 public:
@@ -36,7 +36,6 @@ public:
   static InputParameters validParams();
 
   ParallelStudy(const libMesh::Parallel::Communicator & comm,
-                Context * const context,
                 const InputParameters & params,
                 const std::string & name);
 
@@ -85,7 +84,8 @@ public:
   /**
    * Gets the receive buffer.
    */
-  const ReceiveBuffer<ParallelDataType, Context> & receiveBuffer() const
+  const ReceiveBuffer<ParallelDataType, ParallelStudy<WorkType, ParallelDataType>> &
+  receiveBuffer() const
   {
     return *_receive_buffer;
   }
@@ -251,8 +251,6 @@ protected:
    */
   bool buffersAreEmpty() const;
 
-  /// The context
-  Context * const _context;
   /// This rank
   const processor_id_type _pid;
   /// Name for this object for use in error handling
@@ -333,9 +331,12 @@ private:
   /// Buffer for executing work
   const std::unique_ptr<MooseUtils::Buffer<WorkType>> _work_buffer;
   /// The receive buffer
-  const std::unique_ptr<ReceiveBuffer<ParallelDataType, Context>> _receive_buffer;
+  const std::unique_ptr<ReceiveBuffer<ParallelDataType, ParallelStudy<WorkType, ParallelDataType>>>
+      _receive_buffer;
   /// Send buffers for each processor
-  std::unordered_map<processor_id_type, std::unique_ptr<SendBuffer<ParallelDataType, Context>>>
+  std::unordered_map<
+      processor_id_type,
+      std::unique_ptr<SendBuffer<ParallelDataType, ParallelStudy<WorkType, ParallelDataType>>>>
       _send_buffers;
 
   /// Number of chunks of work executed on this processor
@@ -359,14 +360,12 @@ private:
   bool _currently_executing_work;
 };
 
-template <typename WorkType, typename ParallelDataType, typename Context>
-ParallelStudy<WorkType, ParallelDataType, Context>::ParallelStudy(
+template <typename WorkType, typename ParallelDataType>
+ParallelStudy<WorkType, ParallelDataType>::ParallelStudy(
     const libMesh::Parallel::Communicator & comm,
-    Context * const context,
     const InputParameters & params,
     const std::string & name)
   : ParallelObject(comm),
-    _context(context),
     _pid(comm.rank()),
     _name(name),
     _params(params),
@@ -390,8 +389,9 @@ ParallelStudy<WorkType, ParallelDataType, Context>::ParallelStudy(
     _parallel_data_pools(libMesh::n_threads()),
     _temp_threaded_work(libMesh::n_threads()),
     _work_buffer(createWorkBuffer()),
-    _receive_buffer(libmesh_make_unique<ReceiveBuffer<ParallelDataType, Context>>(
-        comm, context, _method, _clicks_per_receive, _parallel_data_buffer_tag)),
+    _receive_buffer(libmesh_make_unique<
+                    ReceiveBuffer<ParallelDataType, ParallelStudy<WorkType, ParallelDataType>>>(
+        comm, this, _method, _clicks_per_receive, _parallel_data_buffer_tag)),
 
     _currently_executing(false),
     _currently_pre_executing(false),
@@ -408,9 +408,9 @@ ParallelStudy<WorkType, ParallelDataType, Context>::ParallelStudy(
                "('allow_new_work_during_execution = true'), the method must be SMART");
 }
 
-template <typename WorkType, typename ParallelDataType, typename Context>
[…]
