+++ framework/src/transfers/MultiAppNearestNodeTransfer.C
@@ -22,6 +22,10 @@
 #include "libmesh/parallel_algebra.h"
 #include "libmesh/dof_object.h"
 
+// TIMPI includes
+#include "timpi/communicator.h"
+#include "timpi/parallel_sync.h"
+
 registerMooseObject("MooseApp", MultiAppNearestNodeTransfer);
 
 defineLegacyParams(MultiAppNearestNodeTransfer);
@@ -55,9 +59,10 @@ MultiAppNearestNodeTransfer::MultiAppNearestNodeTransfer(const InputParameters &
     _node_map(declareRestartableData<std::map<dof_id_type, Node *>>("node_map")),
     _distance_map(declareRestartableData<std::map<dof_id_type, Real>>("distance_map")),
     _neighbors_cached(declareRestartableData<bool>("neighbors_cached", false)),
-    _cached_froms(declareRestartableData<std::vector<std::vector<unsigned int>>>("cached_froms")),
-    _cached_dof_ids(
-        declareRestartableData<std::vector<std::vector<dof_id_type>>>("cached_dof_ids")),
+    _cached_froms(declareRestartableData<std::map<processor_id_type, std::vector<unsigned int>>>(
+        "cached_froms")),
+    _cached_dof_ids(declareRestartableData<std::map<processor_id_type, std::vector<dof_id_type>>>(
+        "cached_dof_ids")),
     _cached_from_inds(
         declareRestartableData<std::map<dof_id_type, unsigned int>>("cached_from_ids")),
     _cached_qp_inds(declareRestartableData<std::map<dof_id_type, unsigned int>>("cached_qp_inds"))
@@ -101,11 +106,13 @@ MultiAppNearestNodeTransfer::execute()
   ////////////////////
 
   // outgoing_qps = nodes/centroids we'll send to other processors.
-  std::vector<std::vector<Point>> outgoing_qps(n_processors());
+  // Map processor to quadrature points. We will send these points to remote processors
+  std::map<processor_id_type, std::vector<Point>> outgoing_qps;
   // When we get results back, node_index_map will tell us which results go with
   // which points
-  std::vector<std::map<std::pair<unsigned int, dof_id_type>, dof_id_type>> node_index_map(
-      n_processors());
+  // <processor, <system_id, node_i>> --> point_id
+  std::map<processor_id_type, std::map<std::pair<unsigned int, dof_id_type>, dof_id_type>>
+      node_index_map;
 
   if (!_neighbors_cached)
   {
@@ -180,6 +187,7 @@ MultiAppNearestNodeTransfer::execute()
               if (distance <= nearest_max_distance || bboxes[i_from].contains_point(*node))
               {
                 std::pair<unsigned int, dof_id_type> key(i_to, node->id());
+                // Record a local ID for each quadrature point
                 node_index_map[i_proc][key] = outgoing_qps[i_proc].size();
                 outgoing_qps[i_proc].push_back(*node + _to_positions[i_to]);
                 qp_found = true;
@@ -299,23 +307,14 @@ MultiAppNearestNodeTransfer::execute()
   // requested that point.
   ////////////////////
 
-  std::vector<std::vector<Real>> incoming_evals(n_processors());
-  std::vector<Parallel::Request> send_qps(n_processors());
-  std::vector<Parallel::Request> send_evals(n_processors());
+  std::map<processor_id_type, std::vector<Real>> incoming_evals;
 
   // Create these here so that they live the entire life of this function
   // and are NOT reused per processor.
-  std::vector<std::vector<Real>> processor_outgoing_evals(n_processors());
+  std::map<processor_id_type, std::vector<Real>> processor_outgoing_evals;
 
   if (!_neighbors_cached)
   {
-    for (processor_id_type i_proc = 0; i_proc < n_processors(); i_proc++)
-    {
[…]
