+++ framework/src/transfers/MultiAppProjectionTransfer.C
@@ -27,6 +27,10 @@
 #include "libmesh/string_to_enum.h"
 #include "libmesh/default_coupling.h"
 
+// TIMPI includes
+#include "timpi/communicator.h"
+#include "timpi/parallel_sync.h"
+
 void
 assemble_l2(EquationSystems & es, const std::string & system_name)
 {
@@ -120,12 +124,6 @@ MultiAppProjectionTransfer::initialSetup()
 
     // Reinitialize EquationSystems since we added a system.
     to_es.reinit();
-  }
-
-  if (_fixed_meshes)
-  {
-    _cached_qps.resize(n_processors());
-    _cached_index_map.resize(n_processors());
   }
 }
 
@@ -239,9 +237,9 @@ MultiAppProjectionTransfer::execute()
   // Figure out how many "from" domains each processor owns.
   std::vector<unsigned int> froms_per_proc = getFromsPerProc();
 
-  std::vector<std::vector<Point>> outgoing_qps(n_processors());
-  std::vector<std::map<std::pair<unsigned int, unsigned int>, unsigned int>> element_index_map(
-      n_processors());
+  std::map<processor_id_type, std::vector<Point>> outgoing_qps;
+  std::map<processor_id_type, std::map<std::pair<unsigned int, unsigned int>, unsigned int>>
+      element_index_map;
   // element_index_map[i_to, element_id] = index
   // outgoing_qps[index] is the first quadrature point in element
 
@@ -309,13 +307,6 @@ MultiAppProjectionTransfer::execute()
   // requests sent to this processor.
   ////////////////////
 
-  // Non-blocking send quadrature points to other processors.
-  std::vector<Parallel::Request> send_qps(n_processors());
-  if (!_qps_cached)
-    for (processor_id_type i_proc = 0; i_proc < n_processors(); i_proc++)
-      if (i_proc != processor_id())
-        _communicator.send(i_proc, outgoing_qps[i_proc], send_qps[i_proc]);
-
   // Get the local bounding boxes.
   std::vector<BoundingBox> local_bboxes(froms_per_proc[processor_id()]);
   {
@@ -349,37 +340,38 @@ MultiAppProjectionTransfer::execute()
 
   // Recieve quadrature points from other processors, evaluate mesh frunctions
   // at those points, and send the values back.
-  std::vector<Parallel::Request> send_evals(n_processors());
-  std::vector<Parallel::Request> send_ids(n_processors());
-  std::vector<std::vector<Real>> outgoing_evals(n_processors());
-  std::vector<std::vector<unsigned int>> outgoing_ids(n_processors());
-  std::vector<std::vector<Real>> incoming_evals(n_processors());
-  std::vector<std::vector<unsigned int>> incoming_app_ids(n_processors());
-  for (processor_id_type i_proc = 0; i_proc < n_processors(); i_proc++)
-  {
-    // Use the cached qps if they're available.
-    std::vector<Point> incoming_qps;
-    if (!_qps_cached)
-    {
-      if (i_proc == processor_id())
-        incoming_qps = outgoing_qps[i_proc];
-      else
-        _communicator.receive(i_proc, incoming_qps);
-      // Cache these qps for later if _fixed_meshes
-      if (_fixed_meshes)
-        _cached_qps[i_proc] = incoming_qps;
-    }
-    else
-    {
-      incoming_qps = _cached_qps[i_proc];
-    }
+  std::map<processor_id_type, std::vector<std::pair<Real, unsigned int>>> outgoing_evals_ids;
 
-    outgoing_evals[i_proc].resize(incoming_qps.size(), OutOfMeshValue);
-    if (_current_direction == FROM_MULTIAPP)
-      outgoing_ids[i_proc].resize(incoming_qps.size(), libMesh::invalid_uint);
[…]
