### 2021-02-24 e9296d9
Check for discontinuites if using a point sampler at faces, refs #17115 -Use a point locator that returns all the possible elements -Avoid unitializatied pointer, add partitioned and consistency of evaluation tests to point sampler -Remove bounding box check to allow for global communications when solving for who owns the points -More selective disabling of test warnings for point sampling -Revert to previous bounding box optimization for the continuous variable case
