### 2021-02-24 cd2caa3
Check for discontinuites if using a point sampler at faces, refs #17115
