### 2021-02-24 c601514
support vector parameter for MultiAppCommandLineControl ref #17116
