+++ framework/include/restart/DataIO.h
@@ -344,6 +344,12 @@ void dataStore(std::ostream & stream, Real & v, void * context);
 template <>
 void dataStore(std::ostream & stream, std::string & v, void * context);
 template <>
+void dataStore(std::ostream & stream, bool & v, void * context);
+// Vectors of bools are special
+// https://en.wikipedia.org/w/index.php?title=Sequence_container_(C%2B%2B)&oldid=767869909#Specialization_for_bool
+template <>
+void dataStore(std::ostream & stream, std::vector<bool> & v, void * context);
+template <>
 void dataStore(std::ostream & stream, const Elem *& e, void * context);
 template <>
 void dataStore(std::ostream & stream, const Node *& n, void * context);
@@ -609,6 +615,12 @@ template <>
 void dataLoad(std::istream & stream, Real & v, void * /*context*/);
 template <>
 void dataLoad(std::istream & stream, std::string & v, void * /*context*/);
+template <>
+void dataLoad(std::istream & stream, bool & v, void * /*context*/);
+// Vectors of bools are special
+// https://en.wikipedia.org/w/index.php?title=Sequence_container_(C%2B%2B)&oldid=767869909#Specialization_for_bool
+template <>
+void dataLoad(std::istream & stream, std::vector<bool> & v, void * /*context*/);
 template <>
 void dataLoad(std::istream & stream, const Elem *& e, void * context);
 template <>
