### 2021-02-24 910aef8
Add tests and enum
