### 2021-02-24 09b9a81
Replaced makefile with moosetools HIT, removed old HIT unit tests
