+++ unit/src/HitTests.C
@@ -1,626 +1,0 @@
-//* This file is part of the MOOSE framework
-//* https://www.mooseframework.org
-//*
-//* All rights reserved, see COPYRIGHT for full restrictions
-//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
-//*
-//* Licensed under LGPL 2.1, please see LICENSE for details
-//* https://www.gnu.org/licenses/lgpl-2.1.html
-
-#include "hit.h"
-#include "Parser.h"
-
-#include "gtest_include.h"
-
-#include <iostream>
-#include <vector>
-
-// TODO:
-//
-// * don't include trailing newlines in tokens (e.g. unquoted strings, strings, numbers, etc.)
-// * build proper nesting structure
-// * throw errors for bad/invalid number formats - both during parsing and retrieval (i.e. param
-// func)
-// * "${}" expressions get tokenized into strings correctly (including nesting, with spaces, etc.)
-// * Lots of tests for tree merging
-//
-//     - existing param gets overridden in dst
-//     - non-existing param gets duplicated into dst (not moved)
-//     - existing tree gets overridden correctly
-//     - non-existing tree gets duplicated into dst (not moved)
-
-struct PassFailCase
-{
-  std::string name;
-  std::string input;
-};
-
-TEST(HitTests, FailCases)
-{
-  PassFailCase cases[] = {
-      {"comment in path", "[hello#world] []"},
-      {"comment in field", "hello#world=foo"},
-      {"missing string", "[hello] foo = []"},
-      {"missing string 2", "[hello] foo = \n bar = 42[]"},
-      {"invalid path char '='", "[hello=world] []"},
-      {"invalid path char '&'", "[hello&world] []"},
-      {"invalid path char '['", "[hello[world] []"},
-      {"invalid field char '&'", "hello&world=foo"},
-      {"invalid field char '['", "hello[world=foo"},
-      {"unfinished field", "hello\nfoo=bar"},
-      {"unterminated section", "[hello]"},
-      {"unterminated section", "[hello][./]"},
-      {"extra section close", "[]"},
-      {"extra section close 2", "[../]"},
-      {"empty  dotslash section name", "[./][]"},
-      {"mismatched consecutive string literal quotes", "foo='bar'\"baz\""},
-  };
-
-  for (size_t i = 0; i < sizeof(cases) / sizeof(PassFailCase); i++)
-  {
-    auto test = cases[i];
-    EXPECT_THROW(hit::parse("TESTCASE", test.input), hit::Error)
-        << "case " << i + 1 << " FAIL (" << test.name << "): parser failed to error on bad input '"
-        << test.input << "'";
-  }
-}
-
-struct LineCase
-{
-  std::string input;
-  std::vector<int> line_nums;
-};
-
-class LineWalker : public hit::Walker
-{
-public:
-  LineWalker(int i, const std::vector<int> & want_lines) : _case(i), _want(want_lines){};
-  virtual void
-  walk(const std::string & fullpath, const std::string & /*nodepath*/, hit::Node * n) override
-  {
-    if (n->type() == hit::NodeType::Blank || fullpath == "")
-      return;
-
-    if (_count >= _want.size())
-      FAIL() << "case " << _case + 1 << " has more nodes than expected";
-    EXPECT_EQ(_want[_count], n->line())
[…]
