### 2021-02-24 599c093
Avoid Eigen's ambiguous operator error.
- Eigen matrices will be non-AD. For now, they can be considered as user input.
- Some clean up. Apply hit format to some inputs.
