### 2021-02-25 69184ce
Use a point locator that returns all the possible elements
