### 2021-02-25 68a777d
Apply suggestions from code review
Co-authored-by: Logan Harbour <loganhharbour@gmail.com>
