+++ modules/navier_stokes/include/materials/PorousConservedVarMaterial.h
@@ -30,15 +30,21 @@ protected:
   const ADVariableValue & _var_total_energy_density;
   const MaterialProperty<Real> & _epsilon;
   ADMaterialProperty<Real> & _rho;
+  ADMaterialProperty<Real> & _superficial_rho;
   ADMaterialProperty<RealVectorValue> & _mass_flux;
   ADMaterialProperty<Real> & _total_energy_density;
   ADMaterialProperty<RealVectorValue> & _velocity;
+  ADMaterialProperty<RealVectorValue> & _superficial_velocity;
   ADMaterialProperty<Real> & _vel_x;
   ADMaterialProperty<Real> & _vel_y;
   ADMaterialProperty<Real> & _vel_z;
+  ADMaterialProperty<Real> & _rhou;
+  ADMaterialProperty<Real> & _rhov;
+  ADMaterialProperty<Real> & _rhow;
   ADMaterialProperty<Real> & _v;
   ADMaterialProperty<Real> & _specific_internal_energy;
   ADMaterialProperty<Real> & _pressure;
   ADMaterialProperty<Real> & _specific_total_enthalpy;
   ADMaterialProperty<Real> & _rho_ht;
+  ADMaterialProperty<Real> & _superficial_rho_ht;
 };
