+++ modules/navier_stokes/src/materials/PorousConservedVarMaterial.C
@@ -44,17 +44,23 @@ PorousConservedVarMaterial::PorousConservedVarMaterial(const InputParameters & p
     _var_total_energy_density(adCoupledValue(NS::total_energy_density)),
     _epsilon(getMaterialProperty<Real>(NS::porosity)),
     _rho(declareADProperty<Real>(NS::density)),
+    _superficial_rho(declareADProperty<Real>(NS::superficial_density)),
     _mass_flux(declareADProperty<RealVectorValue>(NS::mass_flux)),
     _total_energy_density(declareADProperty<Real>(NS::total_energy_density)),
     _velocity(declareADProperty<RealVectorValue>(NS::velocity)),
+    _superficial_velocity(declareADProperty<RealVectorValue>(NS::superficial_velocity)),
     _vel_x(declareADProperty<Real>(NS::velocity_x)),
     _vel_y(declareADProperty<Real>(NS::velocity_y)),
     _vel_z(declareADProperty<Real>(NS::velocity_z)),
+    _rhou(declareADProperty<Real>(NS::momentum_x)),
+    _rhov(declareADProperty<Real>(NS::momentum_y)),
+    _rhow(declareADProperty<Real>(NS::momentum_z)),
     _v(declareADProperty<Real>(NS::v)),
     _specific_internal_energy(declareADProperty<Real>(NS::specific_internal_energy)),
     _pressure(declareADProperty<Real>(NS::pressure)),
     _specific_total_enthalpy(declareADProperty<Real>(NS::specific_total_enthalpy)),
-    _rho_ht(declareADProperty<Real>(NS::total_enthalpy_density))
+    _rho_ht(declareADProperty<Real>(NS::total_enthalpy_density)),
+    _superficial_rho_ht(declareADProperty<Real>(NS::superficial_total_enthalpy_density))
 {
 }
 
@@ -62,13 +68,18 @@ void
 PorousConservedVarMaterial::computeQpProperties()
 {
   _rho[_qp] = _var_rho[_qp];
+  _superficial_rho[_qp] = _rho[_qp] * _epsilon[_qp];
   _mass_flux[_qp] = {_var_rho_ud[_qp], _var_rho_vd[_qp], _var_rho_wd[_qp]};
   _total_energy_density[_qp] = _var_total_energy_density[_qp];
 
-  _velocity[_qp] = _mass_flux[_qp] / _rho[_qp] / _epsilon[_qp];
+  _superficial_velocity[_qp] = _mass_flux[_qp] / _rho[_qp];
+  _velocity[_qp] = _superficial_velocity[_qp] / _epsilon[_qp];
   _vel_x[_qp] = _velocity[_qp](0);
   _vel_y[_qp] = _velocity[_qp](1);
   _vel_z[_qp] = _velocity[_qp](2);
+  _rhou[_qp] = _vel_x[_qp] * _rho[_qp];
+  _rhov[_qp] = _vel_y[_qp] * _rho[_qp];
+  _rhow[_qp] = _vel_z[_qp] * _rho[_qp];
 
   _v[_qp] = 1 / _rho[_qp];
 
@@ -79,4 +90,5 @@ PorousConservedVarMaterial::computeQpProperties()
 
   _specific_total_enthalpy[_qp] = (_total_energy_density[_qp] + _pressure[_qp]) / _rho[_qp];
   _rho_ht[_qp] = _specific_total_enthalpy[_qp] * _rho[_qp];
+  _superficial_rho_ht[_qp] = _rho_ht[_qp] * _epsilon[_qp];
 }
