+++ modules/geochemistry/unit/src/GeochemicalDatabaseReaderTest.C
@@ -114,7 +114,7 @@ TEST(GeochemicalDatabaseReaderTest, getNeutralSpeciesActivity)
   GeochemicalDatabaseReader database("database/moose_testdb.json");
 
   // Get the neutral species activity coefficients from the database
-  // and compare with the expected valules
+  // and compare with the expected values
   auto nsa = database.getNeutralSpeciesActivity();
 
   auto ns = nsa["co2"];
@@ -139,6 +139,21 @@ TEST(GeochemicalDatabaseReaderTest, getNeutralSpeciesActivity)
   EXPECT_EQ(ns.b, b_gold);
   EXPECT_EQ(ns.c, c_gold);
   EXPECT_EQ(ns.d, d_gold);
+
+  // check error is thrown if database has no neutral species
+  GeochemicalDatabaseReader database_non("database/faultydbs/no_neutral_species.json");
+  try
+  {
+    auto nsa = database_non.getNeutralSpeciesActivity();
+    FAIL() << "Missing expected exception.";
+  }
+  catch (const std::exception & e)
+  {
+    std::string msg(e.what());
+    ASSERT_TRUE(msg.find("No neutral species activity coefficients in database") !=
+                std::string::npos)
+        << "Failed with unexpected error message: " << msg;
+  }
 }
 
 TEST(GeochemicalDatabaseReaderTest, getElements)
@@ -202,6 +217,20 @@ TEST(GeochemicalDatabaseReaderTest, getBasisSpecies)
   EXPECT_EQ(species.charge, 1);
   EXPECT_EQ(species.molecular_weight, 1.0079);
   EXPECT_EQ(species.elements, els_gold);
+
+  // check that an error is thrown if the species does not exist
+  try
+  {
+    auto bs_not = database.getBasisSpecies({"does_not_exist"});
+    FAIL() << "Missing expected exception.";
+  }
+  catch (const std::exception & e)
+  {
+    std::string msg(e.what());
+    ASSERT_TRUE(msg.find("does_not_exist does not exist in database database/moose_testdb.json") !=
+                std::string::npos)
+        << "Failed with unexpected error message: " << msg;
+  }
 }
 
 TEST(GeochemicalDatabaseReaderTest, getEquilibriumSpecies)
@@ -274,6 +303,20 @@ TEST(GeochemicalDatabaseReaderTest, getEquilibriumSpecies)
   EXPECT_EQ(sspecies.molecular_weight, 17.0073);
   EXPECT_EQ(sspecies.basis_species, bs_gold);
   EXPECT_EQ(sspecies.equilibrium_const, logk_gold);
+
+  // check that an error is thrown if the species does not exist
+  try
+  {
+    auto bs_not = database.getEquilibriumSpecies({"does_not_exist"});
+    FAIL() << "Missing expected exception.";
+  }
+  catch (const std::exception & e)
+  {
+    std::string msg(e.what());
+    ASSERT_TRUE(msg.find("does_not_exist does not exist in database database/moose_testdb.json") !=
+                std::string::npos)
+        << "Failed with unexpected error message: " << msg;
+  }
 }
 
 TEST(GeochemicalDatabaseReaderTest, getMineralSpecies)
@@ -316,6 +359,20 @@ TEST(GeochemicalDatabaseReaderTest, getMineralSpecies)
   EXPECT_EQ(mspecies.surface_area, 600.0);
   bs_gold = {{">(s)FeOH", 0.005}, {">(w)FeOH", 0.2}};
   EXPECT_EQ(mspecies.sorption_sites, bs_gold);
+
+  // check that an error is thrown if the species does not exist
+  try
+  {
[…]
