+++ modules/geochemistry/src/utils/GeochemicalDatabaseReader.C
@@ -61,23 +61,14 @@ GeochemicalDatabaseReader::validate(const FileName filename, const nlohmann::jso
 void
 GeochemicalDatabaseReader::reexpressFreeElectron()
 {
-  if (!_root.contains("free electron"))
+  if (!_root.contains("free electron") || !_root["free electron"].contains("e-") ||
+      !_root["free electron"]["e-"]["species"].contains("O2(g)"))
     return;
-  if (!_root["free electron"].contains("e-"))
+  if (!_root.contains("basis species") || !_root["basis species"].contains("O2(aq)"))
     return;
-  if (!_root["free electron"]["e-"]["species"].contains("O2(g)"))
-    return;
-  if (!_root.contains("basis species"))
-    return;
-  if (!_root["basis species"].contains("O2(aq)"))
-    return;
-  if (!_root.contains("gas species"))
-    return;
-  if (!_root["gas species"].contains("O2(g)"))
-    return;
-  if (!_root["gas species"]["O2(g)"]["species"].contains("O2(aq)"))
-    return;
-  if (_root["gas species"]["O2(g)"]["species"].size() != 1)
+  if (!_root.contains("gas species") || !_root["gas species"].contains("O2(g)") ||
+      !_root["gas species"]["O2(g)"]["species"].contains("O2(aq)") ||
+      (_root["gas species"]["O2(g)"]["species"].size() != 1))
     return;
 
   // remove O2(g) in the "e-" and replace with O2(aq)
@@ -665,12 +656,9 @@ GeochemicalDatabaseReader::oxideReactions(const std::vector<std::string> & names
 
 std::vector<std::string>
 GeochemicalDatabaseReader::printReactions(
-    std::vector<std::string> names, std::vector<std::map<std::string, Real>> basis_species) const
+    const std::vector<std::string> & names,
+    const std::vector<std::map<std::string, Real>> & basis_species) const
 {
-  // Length of the input vectors must be equal
-  if (names.size() != basis_species.size())
-    mooseError("Input vectors for printReactions() must be equal size");
-
   std::vector<std::string> reactions;
 
   for (unsigned int i = 0; i < names.size(); ++i)
