+++ modules/geochemistry/include/utils/GeochemicalDatabaseReader.h
@@ -461,16 +461,6 @@ protected:
   void removeExtrapolatedSecondarySpecies();
 
   /**
-   * Generates a formatted vector of strings representing all reactions
-   * @param names list of reaction species
-   * @param basis species list of basis species for each reaction species
-   * @return formatted reaction equations
-   */
-  std::vector<std::string>
-  printReactions(std::vector<std::string> names,
-                 std::vector<std::map<std::string, Real>> basis_species) const;
-
-  /**
    * Copy the temperature points (if any) found in the database into _temperature_points.  This
    * method is called in the constructor
    */
@@ -522,6 +512,19 @@ protected:
   GeochemistryDebyeHuckel _debye_huckel;
   /// Neutral species activity coefficients
   std::map<std::string, GeochemistryNeutralSpeciesActivity> _neutral_species_activity;
-  // Helper for converting json node to Real from string, if needed
+  // Helper for converting json node to Real from string, if needed (see
+  // https://www.urbandictionary.com/define.php?term=get%20real)
   static Real getReal(const nlohmann::json & node);
+
+private:
+  /**
+   * Generates a formatted vector of strings representing all reactions
+   * @param names list of reaction species
+   * @param basis species list of basis species for each reaction species (this vector must be of
+   * same size as names)
+   * @return formatted reaction equations
+   */
+  std::vector<std::string>
+  printReactions(const std::vector<std::string> & names,
+                 const std::vector<std::map<std::string, Real>> & basis_species) const;
 };

[…]
