### 2021-02-25 f9939fa
Avoid unitializatied pointer, add partitioned and consistency of evaluation tests to point sampler
