+++ framework/contrib/hit/hit.h
@@ -1,10 +1,0 @@
-
-#ifndef HIT_HIT
-#define HIT_HIT
-
-#include "parse.h"
-#include "braceexpr.h"
-
-#include <fstream> // TODO: remove this - hit parser doesn't need it at all
-
-#endif
