+++ unit/src/HitTests.C
@@ -1,0 +1,170 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#include "hit.h"
+#include "Parser.h"
+
+#include "gtest_include.h"
+
+#include <iostream>
+#include <vector>
+
+// TODO:
+//
+// * don't include trailing newlines in tokens (e.g. unquoted strings, strings, numbers, etc.)
+// * build proper nesting structure
+// * throw errors for bad/invalid number formats - both during parsing and retrieval (i.e. param
+// func)
+// * "${}" expressions get tokenized into strings correctly (including nesting, with spaces, etc.)
+// * Lots of tests for tree merging
+//
+//     - existing param gets overridden in dst
+//     - non-existing param gets duplicated into dst (not moved)
+//     - existing tree gets overridden correctly
+//     - non-existing tree gets duplicated into dst (not moved)
+
+struct PassFailCase
+{
+  std::string name;
+  std::string input;
+};
+
+TEST(HitTests, BraceExpressions)
+{
+  ValCase cases[] = {
+      {"substitute string", "foo=bar boo=${foo}", "boo", "bar", hit::Field::Kind::String},
+      {"substitute string explicit",
+       "foo=bar boo=${replace foo}",
+       "boo",
+       "bar",
+       hit::Field::Kind::String},
+      {"trailing space", "foo=bar boo=${foo} ", "boo", "bar", hit::Field::Kind::String},
+      {"substute number", "foo=42 boo=${foo}", "boo", "42", hit::Field::Kind::Int},
+      {"multiple replacements",
+       "foo=42 boo='${foo} ${foo}'",
+       "boo",
+       "42 42",
+       hit::Field::Kind::String},
+      {"nested", "foo=bar [hello] boo='${foo}' []", "hello/boo", "bar", hit::Field::Kind::String},
+      {"repl-header-before",
+       "src=foo [src] bar='${src}' []",
+       "src/bar",
+       "foo",
+       hit::Field::Kind::String},
+      {"repl-header-missing", "[src] bar='${src}' []", "src/bar", "${src}", hit::Field::Kind::None},
+      {"repl-header-shadow",
+       "[src] bar='${src}' [] src=foo",
+       "src/bar",
+       "${src}",
+       hit::Field::Kind::None},
+      {"nested shadow",
+       "foo=bar [hello] foo=baz boo='${foo}' []",
+       "hello/boo",
+       "baz",
+       hit::Field::Kind::String},
+      {"multi-line brace expression",
+       "foo=${raw 4\n"
+       "          2\n"
+       "     }",
+       "foo",
+       "42",
+       hit::Field::Kind::String},
+      {"fparse", "foo=${fparse 40 + 2}\n", "foo", "42", hit::Field::Kind::Float},
+      {"fparse-dep-chain",
+       "foo=${fparse 42} bar=${fparse foo}",
+       "bar",
+       "42",
+       hit::Field::Kind::Float},
+      {"fparse-dep-chain-quoted",
+       "foo='${fparse 42}' bar=${fparse foo}",
+       "bar",
+       "42",
+       hit::Field::Kind::Float},
+      {"fparse-with-pi", "foo=${fparse cos(pi)}\n", "foo", "-1", hit::Field::Kind::Float},
[…]
