+++ __init__.py
@@ -1,8 +1,0 @@
-#* This file is part of MOOSETOOLS repository
-#* https://www.github.com/idaholab/moosetools
-#*
-#* All rights reserved, see COPYRIGHT for full restrictions
-#* https://github.com/idaholab/moose/blob/master/COPYRIGHT
-#*
-#* Licensed under LGPL 2.1, please see LICENSE for details
-#* https://www.gnu.org/licenses/lgpl-2.1.html

[…]
