### 2021-02-28 05c78bb
More selective disabling of test warnings for point sampling Style fix
