### 2021-02-28 1704bfe
Remove bounding box check to allow for global communications when solving for who owns the points
