### 2021-02-28 74c31f7
Revert to previous bounding box optimization for the continuous variable case Fix last (?) porous test fail
