+++ framework/src/vectorpostprocessors/PointSamplerBase.C
@@ -98,27 +98,32 @@ PointSamplerBase::execute()
   {
     Point & p = _points[i];
 
-    auto & values = _point_values[i];
-
-    if (values.empty())
-      values.resize(_coupled_moose_vars.size());
-
-    // First find the element the hit lands in
-    const Elem * elem = getLocalElemContainingPoint(p);
-
-    if (elem)
+    // Do a bounding box check so we're not doing unnecessary PointLocator lookups
+    // If discontinuous variables, do all lookups to have global consensus on point ownership
+    if (bbox.contains_point(p) || _discontinuous_variables)
     {
-      // We have to pass a vector of points into reinitElemPhys
-      point_vec[0] = p;
+      auto & values = _point_values[i];
 
-      _subproblem.setCurrentSubdomainID(elem, 0);
-      _subproblem.reinitElemPhys(elem, point_vec, 0); // Zero is for tid
+      if (values.empty())
+        values.resize(_coupled_moose_vars.size());
 
-      for (MooseIndex(_coupled_moose_vars) j = 0; j < _coupled_moose_vars.size(); ++j)
-        values[j] = (dynamic_cast<MooseVariableField<Real> *>(_coupled_moose_vars[j]))->sln()[0] *
-                    _pp_value; // The zero is for the "qp"
+      // First find the element the hit lands in
+      const Elem * elem = getLocalElemContainingPoint(p);
 
-      _found_points[i] = true;
+      if (elem)
+      {
+        // We have to pass a vector of points into reinitElemPhys
+        point_vec[0] = p;
+
+        _subproblem.setCurrentSubdomainID(elem, 0);
+        _subproblem.reinitElemPhys(elem, point_vec, 0); // Zero is for tid
+
+        for (MooseIndex(_coupled_moose_vars) j = 0; j < _coupled_moose_vars.size(); ++j)
+          values[j] = (dynamic_cast<MooseVariableField<Real> *>(_coupled_moose_vars[j]))->sln()[0] *
+                      _pp_value; // The zero is for the "qp"
+
+        _found_points[i] = true;
+      }
     }
   }
 }
