### 2021-02-28 6fb2a48
Add effective thermal diffusion kernel
