### 2021-02-28 912f7da
Add effective thermal diffusion kernel
