### 2021-02-28 115a5d4
Improve params documentation Compute porosity on the face for symmetry BCs
