### 2021-02-28 f45748d
Improve params documentation Compute porosity on the face for symmetry BCs
