+++ modules/navier_stokes/src/fvkernels/PINSFVMomentumAdvection.C
@@ -16,7 +16,8 @@ InputParameters
 PINSFVMomentumAdvection::validParams()
 {
   auto params = INSFVMomentumAdvection::validParams();
-  params.addClassDescription("Object for advecting mass in porous media mass equation");
+  params.addClassDescription("Object for advecting superficial momentum, e.g. rho*u_d, "
+                             "in the porous media momentum equation");
   params.addRequiredCoupledVar("porosity", "Porosity auxiliary variable");
 
   return params;
@@ -141,60 +142,64 @@ PINSFVMomentumAdvection::coeffCalculator(const Elem & elem, const ADReal & mu) c
                   "I think some of my logic might depend on my implicit assumption that we have "
                   "one boundary ID at most per face");
       const auto bc_id = *fi->boundaryIDs().begin();
-
-      if (_no_slip_wall_boundaries.find(bc_id) != _no_slip_wall_boundaries.end())
+      // Find the boundary id that has an associated INSFV boundary condition
+      // if a face has more than one bc_id
+      for (const auto bc_id : fi->boundaryIDs())
       {
-        // Need to account for viscous shear stress from wall
-        for (const auto i : make_range(_dim))
-          coeff(i) += mu / eps_face * surface_vector.norm() /
-                      std::abs((fi->faceCentroid() - rc_centroid) * normal) *
-                      (1 - normal(i) * normal(i));
+        if (_no_slip_wall_boundaries.find(bc_id) != _no_slip_wall_boundaries.end())
+        {
+          // Need to account for viscous shear stress from wall
+          for (const auto i : make_range(_dim))
+            coeff(i) += mu / eps_face * surface_vector.norm() /
+                        std::abs((fi->faceCentroid() - rc_centroid) * normal) *
+                        (1 - normal(i) * normal(i));
 
-        // No flow normal to wall, so no contribution to coefficient from the advection term
-        return;
-      }
+          // No flow normal to wall, so no contribution to coefficient from the advection term
+          return;
+        }
 
-      if (_slip_wall_boundaries.find(bc_id) != _slip_wall_boundaries.end())
-        // In the case of a slip wall we neither have viscous shear stress from the wall nor
-        // normal outflow, so our contribution to the coefficient is zero
-        return;
+        if (_slip_wall_boundaries.find(bc_id) != _slip_wall_boundaries.end())
+          // In the case of a slip wall we neither have viscous shear stress from the wall nor
+          // normal outflow, so our contribution to the coefficient is zero
+          return;
 
-      if (_flow_boundaries.find(bc_id) != _flow_boundaries.end())
-      {
-        ADRealVectorValue face_velocity(_u_var->getBoundaryFaceValue(*fi));
-        if (_v_var)
-          face_velocity(1) = _v_var->getBoundaryFaceValue(*fi);
-        if (_w_var)
-          face_velocity(2) = _w_var->getBoundaryFaceValue(*fi);
+        if (_flow_boundaries.find(bc_id) != _flow_boundaries.end())
+        {
+          ADRealVectorValue face_velocity(_u_var->getBoundaryFaceValue(*fi));
+          if (_v_var)
+            face_velocity(1) = _v_var->getBoundaryFaceValue(*fi);
+          if (_w_var)
+            face_velocity(2) = _w_var->getBoundaryFaceValue(*fi);
 
-        const auto advection_coeffs =
-            Moose::FV::interpCoeffs(_advected_interp_method, *fi, elem_has_info, face_velocity);
[…]
