### 2021-02-28 5c22624
Improve params documentation Compute porosity on the face for symmetry BCs
