+++ modules/navier_stokes/doc/content/source/fvbcs/PINSFVSymmetryVelocityBC.md
@@ -1,0 +1,23 @@
+# PINSFVSymmetryVelocityBC
+
+!alert construction title=Undocumented Class
+The PINSFVSymmetryVelocityBC has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /FVBCs/PINSFVSymmetryVelocityBC
+
+## Overview
+
+!! Replace these lines with information regarding the PINSFVSymmetryVelocityBC object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the PINSFVSymmetryVelocityBC object.
+
+!syntax parameters /FVBCs/PINSFVSymmetryVelocityBC
+
+!syntax inputs /FVBCs/PINSFVSymmetryVelocityBC
+
+!syntax children /FVBCs/PINSFVSymmetryVelocityBC
