+++ modules/navier_stokes/doc/content/source/fvkernels/PINSFVEnergyDiffusion.md
@@ -1,0 +1,23 @@
+# PINSFVEnergyDiffusion
+
+!alert construction title=Undocumented Class
+The PINSFVEnergyDiffusion has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /FVKernels/PINSFVEnergyDiffusion
+
+## Overview
+
+!! Replace these lines with information regarding the PINSFVEnergyDiffusion object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the PINSFVEnergyDiffusion object.
+
+!syntax parameters /FVKernels/PINSFVEnergyDiffusion
+
+!syntax inputs /FVKernels/PINSFVEnergyDiffusion
+
+!syntax children /FVKernels/PINSFVEnergyDiffusion
