+++ modules/navier_stokes/doc/content/source/fvkernels/PINSFVMomentumDiffusion.md
@@ -1,0 +1,23 @@
+# PINSFVMomentumDiffusion
+
+!alert construction title=Undocumented Class
+The PINSFVMomentumDiffusion has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /FVKernels/PINSFVMomentumDiffusion
+
+## Overview
+
+!! Replace these lines with information regarding the PINSFVMomentumDiffusion object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the PINSFVMomentumDiffusion object.
+
+!syntax parameters /FVKernels/PINSFVMomentumDiffusion
+
+!syntax inputs /FVKernels/PINSFVMomentumDiffusion
+
+!syntax children /FVKernels/PINSFVMomentumDiffusion
