+++ modules/navier_stokes/doc/content/source/fvkernels/INSFVMomentumBoussinesq.md
@@ -1,0 +1,23 @@
+# INSFVMomentumBoussinesq
+
+!alert construction title=Undocumented Class
+The INSFVMomentumBoussinesq has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /FVKernels/INSFVMomentumBoussinesq
+
+## Overview
+
+!! Replace these lines with information regarding the INSFVMomentumBoussinesq object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the INSFVMomentumBoussinesq object.
+
+!syntax parameters /FVKernels/INSFVMomentumBoussinesq
+
+!syntax inputs /FVKernels/INSFVMomentumBoussinesq
+
+!syntax children /FVKernels/INSFVMomentumBoussinesq
