+++ modules/navier_stokes/doc/content/source/fvkernels/PINSFVMomentumBoussinesq.md
@@ -1,0 +1,23 @@
+# PINSFVMomentumBoussinesq
+
+!alert construction title=Undocumented Class
+The PINSFVMomentumBoussinesq has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /FVKernels/PINSFVMomentumBoussinesq
+
+## Overview
+
+!! Replace these lines with information regarding the PINSFVMomentumBoussinesq object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the PINSFVMomentumBoussinesq object.
+
+!syntax parameters /FVKernels/PINSFVMomentumBoussinesq
+
+!syntax inputs /FVKernels/PINSFVMomentumBoussinesq
+
+!syntax children /FVKernels/PINSFVMomentumBoussinesq
