+++ modules/navier_stokes/doc/content/source/fvkernels/PINSFVEnergyEffectiveDiffusion.md
@@ -1,0 +1,23 @@
+# PINSFVEnergyEffectiveDiffusion
+
+!alert construction title=Undocumented Class
+The PINSFVEnergyEffectiveDiffusion has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /FVKernels/PINSFVEnergyEffectiveDiffusion
+
+## Overview
+
+!! Replace these lines with information regarding the PINSFVEnergyEffectiveDiffusion object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the PINSFVEnergyEffectiveDiffusion object.
+
+!syntax parameters /FVKernels/PINSFVEnergyEffectiveDiffusion
+
+!syntax inputs /FVKernels/PINSFVEnergyEffectiveDiffusion
+
+!syntax children /FVKernels/PINSFVEnergyEffectiveDiffusion
