+++ modules/navier_stokes/doc/content/source/fvkernels/PINSFVMomentumPressure.md
@@ -1,0 +1,23 @@
+# PINSFVMomentumPressure
+
+!alert construction title=Undocumented Class
+The PINSFVMomentumPressure has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /FVKernels/PINSFVMomentumPressure
+
+## Overview
+
+!! Replace these lines with information regarding the PINSFVMomentumPressure object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the PINSFVMomentumPressure object.
+
+!syntax parameters /FVKernels/PINSFVMomentumPressure
+
+!syntax inputs /FVKernels/PINSFVMomentumPressure
+
+!syntax children /FVKernels/PINSFVMomentumPressure
