+++ modules/navier_stokes/doc/content/source/fvkernels/PINSFVEnergyAdvection.md
@@ -1,0 +1,23 @@
+# PINSFVEnergyAdvection
+
+!alert construction title=Undocumented Class
+The PINSFVEnergyAdvection has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /FVKernels/PINSFVEnergyAdvection
+
+## Overview
+
+!! Replace these lines with information regarding the PINSFVEnergyAdvection object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the PINSFVEnergyAdvection object.
+
+!syntax parameters /FVKernels/PINSFVEnergyAdvection
+
+!syntax inputs /FVKernels/PINSFVEnergyAdvection
+
+!syntax children /FVKernels/PINSFVEnergyAdvection
