+++ modules/navier_stokes/doc/content/source/fvkernels/PINSFVMassAdvection.md
@@ -1,0 +1,23 @@
+# PINSFVMassAdvection
+
+!alert construction title=Undocumented Class
+The PINSFVMassAdvection has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /FVKernels/PINSFVMassAdvection
+
+## Overview
+
+!! Replace these lines with information regarding the PINSFVMassAdvection object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the PINSFVMassAdvection object.
+
+!syntax parameters /FVKernels/PINSFVMassAdvection
+
+!syntax inputs /FVKernels/PINSFVMassAdvection
+
+!syntax children /FVKernels/PINSFVMassAdvection
