+++ modules/navier_stokes/doc/content/source/fvkernels/INSFVMomentumGravity.md
@@ -1,0 +1,23 @@
+# INSFVMomentumGravity
+
+!alert construction title=Undocumented Class
+The INSFVMomentumGravity has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /FVKernels/INSFVMomentumGravity
+
+## Overview
+
+!! Replace these lines with information regarding the INSFVMomentumGravity object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the INSFVMomentumGravity object.
+
+!syntax parameters /FVKernels/INSFVMomentumGravity
+
+!syntax inputs /FVKernels/INSFVMomentumGravity
+
+!syntax children /FVKernels/INSFVMomentumGravity
