+++ modules/navier_stokes/doc/content/source/fvkernels/PINSFVMassTimeDerivative.md
@@ -1,0 +1,23 @@
+# PINSFVMassTimeDerivative
+
+!alert construction title=Undocumented Class
+The PINSFVMassTimeDerivative has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /FVKernels/PINSFVMassTimeDerivative
+
+## Overview
+
+!! Replace these lines with information regarding the PINSFVMassTimeDerivative object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the PINSFVMassTimeDerivative object.
+
+!syntax parameters /FVKernels/PINSFVMassTimeDerivative
+
+!syntax inputs /FVKernels/PINSFVMassTimeDerivative
+
+!syntax children /FVKernels/PINSFVMassTimeDerivative
