+++ modules/navier_stokes/doc/content/source/fvkernels/PINSFVEnergyTimeDerivative.md
@@ -1,0 +1,23 @@
+# PINSFVEnergyTimeDerivative
+
+!alert construction title=Undocumented Class
+The PINSFVEnergyTimeDerivative has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /FVKernels/PINSFVEnergyTimeDerivative
+
+## Overview
+
+!! Replace these lines with information regarding the PINSFVEnergyTimeDerivative object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the PINSFVEnergyTimeDerivative object.
+
+!syntax parameters /FVKernels/PINSFVEnergyTimeDerivative
+
+!syntax inputs /FVKernels/PINSFVEnergyTimeDerivative
+
+!syntax children /FVKernels/PINSFVEnergyTimeDerivative
