+++ modules/navier_stokes/doc/content/source/fvkernels/PINSFVMomentumFriction.md
@@ -1,0 +1,23 @@
+# PINSFVMomentumFriction
+
+!alert construction title=Undocumented Class
+The PINSFVMomentumFriction has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /FVKernels/PINSFVMomentumFriction
+
+## Overview
+
+!! Replace these lines with information regarding the PINSFVMomentumFriction object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the PINSFVMomentumFriction object.
+
+!syntax parameters /FVKernels/PINSFVMomentumFriction
+
+!syntax inputs /FVKernels/PINSFVMomentumFriction
+
+!syntax children /FVKernels/PINSFVMomentumFriction
