+++ modules/navier_stokes/doc/content/source/fvkernels/PINSFVMomentumAdvection.md
@@ -1,0 +1,23 @@
+# PINSFVMomentumAdvection
+
+!alert construction title=Undocumented Class
+The PINSFVMomentumAdvection has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /FVKernels/PINSFVMomentumAdvection
+
+## Overview
+
+!! Replace these lines with information regarding the PINSFVMomentumAdvection object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the PINSFVMomentumAdvection object.
+
+!syntax parameters /FVKernels/PINSFVMomentumAdvection
+
+!syntax inputs /FVKernels/PINSFVMomentumAdvection
+
+!syntax children /FVKernels/PINSFVMomentumAdvection
