+++ modules/navier_stokes/doc/content/source/fvbcs/PINSFVMomentumAdvectionOutflowBC.md
@@ -1,0 +1,23 @@
+# PINSFVMomentumAdvectionOutflowBC
+
+!alert construction title=Undocumented Class
+The PINSFVMomentumAdvectionOutflowBC has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+!syntax description /FVBCs/PINSFVMomentumAdvectionOutflowBC
+
+## Overview
+
+!! Replace these lines with information regarding the PINSFVMomentumAdvectionOutflowBC object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the PINSFVMomentumAdvectionOutflowBC object.
+
+!syntax parameters /FVBCs/PINSFVMomentumAdvectionOutflowBC
+
+!syntax inputs /FVBCs/PINSFVMomentumAdvectionOutflowBC
+
+!syntax children /FVBCs/PINSFVMomentumAdvectionOutflowBC
