+++ modules/tensor_mechanics/include/materials/czm/SalehaniIrani3DCTractionIncremental.h
@@ -9,32 +9,28 @@
 
 #pragma once
 
-#include "CZMConstitutiveModelTotalBase.h"
+#include "CZMConstitutiveModelIncrementalBase.h"
 
 /**
  * Implementation of the non-stateful exponential traction separation law
  * proposed by Salehani, Mohsen Khajeh and Irani, Nilgoon 2018
  **/
-class SalehaniIrani3DCTraction : public CZMConstitutiveModelTotalBase
+class SalehaniIrani3DCTractionIncremental : public CZMConstitutiveModelIncrementalBase
 {
 public:
   static InputParameters validParams();
-  SalehaniIrani3DCTraction(const InputParameters & parameters);
+  SalehaniIrani3DCTractionIncremental(const InputParameters & parameters);
 
 protected:
-  /// method computing the total traction and its derivatives
-  void computeInterfaceTractionAndDerivatives() override;
-
-  /// method computing the total traction
-  RealVectorValue computeTraction();
-
-  /// method computing the total traction derivatives w.r.t. the interface
-  /// displacement jump
-  RankTwoTensor computeTractionDerivatives();
+  void computeInterfaceTractionIncrementAndDerivatives() override;
 
   /// the displacement jump associated to the maximum traction
   const std::vector<Real> _delta_u0;
 
   /// the vector representing the maximum allowed traction in each direction
   const std::vector<Real> _max_allowable_traction;
+
+private:
+  RealVectorValue computeTraction();
+  RankTwoTensor computeTractionDerivatives();
 };
