### 2021-03-01 6640bc8
Sort interface materials
closes #17171
