### 2021-03-01 72bfa0e
Implement AD material property getters in InterfaceMaterial
We must ensure that these methods exist otherwise the base
`MaterialPropertyInterface` methods will be used and then we won't
register requested material properties for dependency sorting
