+++ test/include/interfacekernels/ADMaterialPropertySource.h
@@ -1,0 +1,25 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#pragma once
+
+#include "ADInterfaceKernel.h"
+
+class ADMaterialPropertySource : public ADInterfaceKernel
+{
+public:
+  static InputParameters validParams();
+
+  ADMaterialPropertySource(const InputParameters & parameters);
+
+protected:
+  virtual ADReal computeQpResidual(Moose::DGResidualType type) override;
+
+  const ADMaterialProperty<Real> & _source;
+};
