### 2021-03-01 fa83bf6
minor edits to tutorial docs
(refs #15232)
