### 2021-03-01 ee0ee00
minor edits to tutorial docs
(refs #15232)
