+++ modules/doc/content/getting_started/examples_and_tutorials/tutorial01_app_development/step09_mat_props.md
@@ -36,19 +36,19 @@ The `"outputs"` parameter defaults to `none` and, therefore, must be explicitly 
 
 ## Demonstration id=demo
 
-Now that the basic purpose of the Materials System has been established, another object in the divergence-free weak form of Darcy's equation can be identified, i.e.,
+Now that the basic purpose of the Materials System has been established, another object in the isotropic, divergence-free weak form of Darcy's equation can be identified, i.e.,
 
 !equation id=darcy-weak
-\underbrace{(\nabla \psi, \underbrace{\dfrac{K}{\mu_{f}}}_{\clap{Material}} \nabla p)}_{\clap{Kernel}} = 0
+\underbrace{(\nabla \psi, \underbrace{\dfrac{K}{\mu}}_{\clap{Material}} \nabla p)}_{\clap{Kernel}} = 0
 
-For this demonstration, a new `Material` object will be developed to produce the $K / \mu_f$ term in [darcy-weak] and the `DarcyPressure` class will be modified to consume it. The new `MooseObject` shall accept an input for the radius of steel spheres, $r_{s}$, tightly packed to form a porous medium. The [tutorial01_app_development/problem_statement.md#mats] section of the Problem Statement provided the following linear relationship between the permeability ($\text{m}^{2}$) of such media and the sphere radius ($\text{mm}$):
+For this demonstration, a new `Material` object will be developed to produce the $K / \mu$ term in [darcy-weak] and the `DarcyPressure` class will be modified to consume it. The new `MooseObject` shall accept an input for the radius, $r$, of steel spheres tightly packed to form a porous medium. The [tutorial01_app_development/problem_statement.md#mats] section of the Problem Statement provided the following linear relationship between the permeability ($\text{m}^{2}$) of such media and the sphere radius ($\text{mm}$):
 
 !equation id=permeability
-K(r_{s}) = \frac{1}{2} \begin{bmatrix} -r_{s} + 3 & r_{s} - 1 \end{bmatrix} \begin{Bmatrix} 0.8451 \\ 8.968 \end{Bmatrix} \times 10^{-9}, \, \, \, \forall \, r_{s} \in [1, 3]
+K(r) = \frac{1}{2} \begin{bmatrix} -r + 3 & r - 1 \end{bmatrix} \begin{Bmatrix} 0.8451 \\ 8.968 \end{Bmatrix} \times 10^{-9}, \, \, \, \forall \, r \in [1, 3]
 
-The new object shall restrict the input radius to the domain of [permeability] and enforce the requirement that $\mu_{f} \; {=}\mathllap{\small{/}\,} \; 0$.
+The new object shall restrict the input radius to the domain of [permeability] and enforce the requirement that $\mu \; {=}\mathllap{\small{/}\,} \; 0$.
 
-Since the pressure vessel model uses $r_{s} = 1 \, \text{mm}$, `pressure_diffusion.i` shall be modified to use this value as input. It will be shown that this input reproduces the isotropic permeability value, $K = 0.8451 \times 10^{-9} \, \textrm{m}^{2}$, introduced in [Step 5](tutorial01_app_development/step05_kernel_object.md#demo). Thus, the results of the `darcy_pressure_test.i` file created in the [previous step](tutorial01_app_development/step08_test_harness.md#test-demo) should still match the gold file after modifying its inputs.
[…]
