+++ modules/doc/content/getting_started/examples_and_tutorials/tutorial01_app_development/problem_statement.md
@@ -4,7 +4,7 @@
 
 # Problem Statement
 
-Consider a system containing two pressure vessels at differing temperatures as in the experiments presented by [!cite](pamuk2012friction). The vessels are connected via a pipe that contains a filter consisting of close-packed steel spheres as shown in [problem-schematic]. Predict the velocity and temperature of the fluid inside the filter. The pipe is 0.304 m in length and 0.0514 m in diameter. The fluid inside the system is liquid water.
+Consider a system containing two pressure vessels at differing temperatures as in the experiments presented by [!cite](pamuk2012friction). The vessels are connected via a pipe that contains a filter consisting of close-packed steel spheres as shown in [problem-schematic]. Predict the velocity and temperature of the fluid inside the filter. The pipe is 0.304 m in length and 0.0514 m in diameter. The fluid inside the system is liquid water at approximately 30 degrees Celcius.
 
 !media tutorial01_app_development/problem_schematic.png
        style=width:80%;margin-left:auto;margin-right:auto;
@@ -20,24 +20,24 @@ To solve this problem, the following physics must be considered:
 - Conservation of Mass:
 
 !equation id=mass
-\nabla \cdot \bar{u} = 0
+\nabla \cdot \vec{u} = 0
 
 - Conservation of Energy:
 
 !equation id=energy
-C \left(\dfrac{\partial T}{\partial t} + \epsilon \bar{u} \cdot \nabla T \right) - \nabla \cdot k \nabla T = 0
+C \left(\dfrac{\partial T}{\partial t} + \epsilon \vec{u} \cdot \nabla T \right) - \nabla \cdot k \nabla T = 0
 
 - Darcy's Law:
 
 !equation id=darcy
-\bar{u} = -\dfrac{\mathbf{K}}{\mu} \left(\nabla p - \rho \bar{g} \right)
+\vec{u} = -\dfrac{\mathbf{K}}{\mu} \left(\nabla p - \rho \vec{g} \right)
 
 The variables shown in [mass], [energy], and [darcy] denote the properties listed in [variables].
 
 !table id=variables caption=Disambiguation of problem variables.
 | Symbol | Property |
 | - | - |
-| $\bar{u}$ | Velocity |
+| $\vec{u}$ | Velocity (Volumetric Flux) Vector |
 | $C$ | Heat Capacity |
 | $T$ | Temperature |
 | $t$ | Time |
@@ -47,15 +47,15 @@ The variables shown in [mass], [energy], and [darcy] denote the properties liste
 | $\mu$ | Dynamic Viscosity |
 | $p$ | Pressure |
 | $\rho$ | Density |
-| $\bar{g}$ | Gravity Vector |
+| $\vec{g}$ | Gravity Vector |
 
-If a zero-gravity condition is assumed, i.e., if $\bar{g} = 0$, and if the divergence-free condition of [mass] is imposed onto [darcy],
+If a zero-gravity condition is assumed, i.e., if $\vec{g} = 0$, and if the divergence-free condition of [mass] is imposed onto [darcy],
 then it follows that the two unknowns, $p$ and $T$, must satisfy the following system of two [!ac](PDEs):
 
 !equation id=system
 -\nabla \cdot \dfrac{\mathbf{K}}{\mu} \nabla p = 0
 \newline \, \newline
-C \left(\dfrac{\partial T}{\partial t} + \epsilon \bar{u} \cdot \nabla T \right) - \nabla \cdot k \nabla T = 0
[…]

[…]
