+++ modules/doc/content/getting_started/examples_and_tutorials/tutorial01_app_development/step10_auxkernels.md
@@ -1,6 +1,250 @@
+!content pagination previous=tutorial01_app_development/step09_mat_props.md
+                    next=tutorial01_app_development/step11_transient_kernel.md
+                    margin-bottom=0px
+
 # Step 10: Develop an AuxKernel Object
 
-!alert construction
-The remainder of this tutorial is currently being developed. More content should be available soon. For now, refer back to the [examples_and_tutorials/index.md] page for other helpful training materials or check out the MOOSE [application_development/index.md] pages for more information.
+This step introduces the auxiliary system in MOOSE, which is based on the `AuxKernel` class and operates on `AuxVariables`. These two components are analogous to `Kernel` objects operating on `Variables`, e.g., a `DarcyPressure` object operating on a pressure variable. However, `AuxVariables` do not directly affect the [source/systems/NonlinearSystem.md]---the various methods that contribute to solving
+[!ac](PDEs) by the Galerkin [!ac](FEM).
+
+To demonstrate the basic use of this system, a new `VectorAuxKernel` object that computes the volumetric flux (velocity) associated with a given pressure gradient will be developed in accordance with Darcy's Law. A simple regression test for this new class will be created. It will also be implemented in the pressure vessel model of `pressure_diffusion.i`.
+
+## Auxiliary Variables
+
+In [Step 5](tutorial01_app_development/step05_kernel_object.md#kernels), it was mentioned that kernels compute residual contributions from the volumetric terms in a [!ac](PDE). The weak form of these terms exposes the primary dependent variable of a given problem, i.e., that which is multiplied by the test function gradient $\nabla \psi$, and which can only be determined by enforcing Dirichlet boundary conditions and solving the resulting system of equations. For example, the primary variable of Darcy's equation is pressure. In the case of solid mechanics (referred to as [Tensor Mechanics](modules/tensor_mechanics/index.md) in MOOSE), it is displacement. In MOOSE lingo, these types of variables are referred to as *nonlinear variables*.
+
+In contrast to nonlinear variables, *auxiliary variables* are never the primary dependent. These variables can always be computed explicitly based on given information during any part of a solve procedure. But, similar to nonlinear variables, they are continuous and possibly differentiable over individual finite elements and vary based on an assumed shape function. Often, they depend on nonlinear variables and/or vice versa. For example, one auxiliary variable in Darcy's equation is the volumetric flux, which depends on the gradient of a scalar pressure field. In the case of solid mechanics, stress and strain, which both depend on the gradient of a displacement field, are considered as auxiliary variables. In structural dynamics, acceleration and velocity are considered to be auxiliary, but the displacement at any given moment depends on their values. In other cases still, auxiliary variables may have no relation to the nonlinear ones.
+
+Auxiliary variables are declared in the [`[AuxVariables]`](syntax/AuxVariables/index.md) block and have access to the same parameters that their nonlinear counterparts do, e.g., `"order"` and `"family"`. They are designated as one of the following two types depending on their input for the `"family"` parameter:
+
+- Elemental (`family = MONOMIAL` or `family = MONOMIAL_VEC`)
+- Nodal (`family = LAGRANGE` or `family = LAGRANGE_VEC`)
+
+It is necessary to distinguish `AuxVariables` based on [!ac](FE) shape function family because their selection dictates how they are treated by `AuxKernel` objects, which are discussed in the next section.
+
[…]
