### 2021-03-02 eb54f3b
Restore CIVET badges to SQA pages
(closes #17177)
