### 2021-03-02 197386b
Adding junction for parallel channels
- This component can be used to connect multiple channels that have the
same direction
- Inlet and outlet pipes can have different flow area

Closes #19841
