+++ modules/thm/src/components/JunctionParallelChannels1Phase.C
@@ -1,0 +1,183 @@
+#include "VolumeJunction1Phase.h"
+#include "JunctionParallelChannels1Phase.h"
+#include "FlowModelSinglePhase.h"
+#include "SinglePhaseFluidProperties.h"
+#include "Function.h"
+#include "GeometricalFlowComponent.h"
+
+registerMooseObject("THMApp", JunctionParallelChannels1Phase);
+
+InputParameters
+JunctionParallelChannels1Phase::validParams()
+{
+  InputParameters params = VolumeJunction1Phase::validParams();
+
+  params.addClassDescription("Junction between 1-phase flow channels that are parallel");
+
+  return params;
+}
+
+JunctionParallelChannels1Phase::JunctionParallelChannels1Phase(const InputParameters & params)
+  : VolumeJunction1Phase(params)
+
+{
+}
+
+void
+JunctionParallelChannels1Phase::init()
+{
+  FlowConnection::init();
+  for (const auto & connection : getConnections())
+  {
+    const std::string & comp_name = connection._geometrical_component_name;
+    if (hasComponentByName<GeometricalFlowComponent>(comp_name))
+    {
+      const GeometricalFlowComponent & gc = getComponentByName<GeometricalFlowComponent>(comp_name);
+      _directions.push_back(gc.getDirection());
+    }
+  }
+}
+
+void
+JunctionParallelChannels1Phase::addVariables()
+{
+  auto connected_subdomains = getConnectedSubdomainNames();
+
+  _sim.addSimVariable(
+      true, _rhoV_var_name, FEType(FIRST, SCALAR), connected_subdomains, _scaling_factor_rhoV);
+  _sim.addSimVariable(
+      true, _rhouV_var_name, FEType(FIRST, SCALAR), connected_subdomains, _scaling_factor_rhouV);
+  _sim.addSimVariable(
+      true, _rhovV_var_name, FEType(FIRST, SCALAR), connected_subdomains, _scaling_factor_rhovV);
+  _sim.addSimVariable(
+      true, _rhowV_var_name, FEType(FIRST, SCALAR), connected_subdomains, _scaling_factor_rhowV);
+  _sim.addSimVariable(
+      true, _rhoEV_var_name, FEType(FIRST, SCALAR), connected_subdomains, _scaling_factor_rhoEV);
+
+  if (isParamValid("initial_p") && isParamValid("initial_T") && isParamValid("initial_vel_x") &&
+      isParamValid("initial_vel_y") && isParamValid("initial_vel_z"))
+  {
+    Function & initial_p_fn = _sim.getFunction(getParam<FunctionName>("initial_p"));
+    Function & initial_T_fn = _sim.getFunction(getParam<FunctionName>("initial_T"));
+    Function & initial_vel_x_fn = _sim.getFunction(getParam<FunctionName>("initial_vel_x"));
+    Function & initial_vel_y_fn = _sim.getFunction(getParam<FunctionName>("initial_vel_y"));
+    Function & initial_vel_z_fn = _sim.getFunction(getParam<FunctionName>("initial_vel_z"));
+
+    initial_p_fn.initialSetup();
+    initial_T_fn.initialSetup();
+    initial_vel_x_fn.initialSetup();
+    initial_vel_y_fn.initialSetup();
+    initial_vel_z_fn.initialSetup();
+
+    const Real initial_p = initial_p_fn.value(0, _position);
+    const Real initial_T = initial_T_fn.value(0, _position);
+    const Real initial_vel_x = initial_vel_x_fn.value(0, _position);
+    const Real initial_vel_y = initial_vel_y_fn.value(0, _position);
[…]
