+++ modules/thm/src/userobjects/JunctionParallelChannels1PhaseUserObject.C
@@ -1,0 +1,287 @@
+#include "JunctionParallelChannels1PhaseUserObject.h"
+#include "SinglePhaseFluidProperties.h"
+#include "THMIndices3Eqn.h"
+#include "VolumeJunction1Phase.h"
+#include "NumericalFlux3EqnBase.h"
+#include "Numerics.h"
+#include "LoggingInterface.h"
+
+registerMooseObject("THMApp", JunctionParallelChannels1PhaseUserObject);
+
+InputParameters
+JunctionParallelChannels1PhaseUserObject::validParams()
+{
+  InputParameters params = VolumeJunction1PhaseUserObject::validParams();
+  params.addRequiredParam<std::string>("component_name", "Name of the associated component");
+  params.addRequiredParam<RealVectorValue>("dir_c0", "Direction of the first connection");
+
+  params.addClassDescription("Computes and caches flux and residual vectors for a 1-phase junction "
+                             "that connects flow channels that are parallel");
+
+  return params;
+}
+
+JunctionParallelChannels1PhaseUserObject::JunctionParallelChannels1PhaseUserObject(
+    const InputParameters & params)
+  : DerivativeMaterialInterfaceTHM<VolumeJunction1PhaseUserObject>(params),
+
+    _p(getMaterialProperty<Real>("p")),
+    _dp_drhoA(getMaterialPropertyDerivativeTHM<Real>("p", "rhoA")),
+    _dp_drhouA(getMaterialPropertyDerivativeTHM<Real>("p", "rhouA")),
+    _dp_drhoEA(getMaterialPropertyDerivativeTHM<Real>("p", "rhoEA")),
+
+    _dir_c0(getParam<RealVectorValue>("dir_c0")),
+
+    _stored_pA(_n_connections),
+    _stored_dp_drhoA(_n_connections),
+    _stored_dp_drhouA(_n_connections),
+    _stored_dp_drhoEA(_n_connections),
+    _areas(_n_connections),
+    _directions(_n_connections),
+    _is_inlet(_n_connections),
+    _connection_indices(_n_connections),
+
+    _component_name(getParam<std::string>("component_name"))
+{
+}
+
+void
+JunctionParallelChannels1PhaseUserObject::initialize()
+{
+  VolumeJunctionBaseUserObject::initialize();
+
+  _connection_indices.clear();
+}
+
+void
+JunctionParallelChannels1PhaseUserObject::storeConnectionData()
+{
+  VolumeJunctionBaseUserObject::storeConnectionData();
+
+  const unsigned int c = getBoundaryIDIndex();
+  _connection_indices.push_back(c);
+}
+
+void
+JunctionParallelChannels1PhaseUserObject::execute()
+{
+  storeConnectionData();
+
+  const unsigned int c = getBoundaryIDIndex();
+
+  computeFluxesAndResiduals(c);
+}
+
+void
+JunctionParallelChannels1PhaseUserObject::computeFluxesAndResiduals(const unsigned int & c)
+{
+  VolumeJunction1PhaseUserObject::computeFluxesAndResiduals(c);
+
+  const Real din = _normal[c];
+  const Point di = _dir[0];
[…]
