### 2021-03-02 b74ef69
Make PerfGraphRegistry be local to the app instance refs #15444
