### 2021-03-02 a65f347
Making sure StochasticReporter can be used for statistics #16166
