### 2021-03-02 eb35440
Revert "Make PerfGraphRegistry be local to the app instance refs #15444"
This reverts commit 8f5c78611a8ac11cdfcddd286a55120086dd48bb.
