+++ framework/include/interfaces/PerfGraphInterface.h
@@ -44,7 +44,7 @@ template <>
 InputParameters validParams<PerfGraphInterface>();
 
 /**
- * Interface for objects that needs PerfGraph capabilities
+ * Interface for objects that needs transient capabilities
  */
 class PerfGraphInterface
 {
