### 2021-03-02 64e32aa
The paper specifies sphere diameter, not radius. Corrected all such references.
(closes idaholab/babbler#6) (refs #15232)
