+++ tutorials/tutorial01_app_development/step10_auxkernels/src/materials/PackedColumn.C
@@ -7,16 +7,16 @@ PackedColumn::validParams()
 {
   InputParameters params = Material::validParams();
   params.addClassDescription("Computes the permeability of a porous medium made up of packed "
-                             "steel spheres of a specified radius in accordance with Pamuk and "
+                             "steel spheres of a specified diameter in accordance with Pamuk and "
                              "Ozdemir (2012). This also provides a specified dynamic viscosity of "
                              "the fluid in the medium.");
 
-  // Optional params for sphere radius and viscosity - inputs must satisfy range checked conditions
+  // Optional params for ball diameter and viscosity - inputs must satisfy range checked conditions
   params.addRangeCheckedParam<Real>(
-      "radius",
+      "diameter",
       1.0,
-      "(1 <= radius) & (radius <= 3)",
-      "The radius (mm) of the steel spheres packed within the column that is used to compute "
+      "(1 <= diameter) & (diameter <= 3)",
+      "The diameter (mm) of the steel spheres packed within the column that is used to compute "
       "its permeability. The input must be in the range [1, 3]. The default value is 1 mm.");
   params.addRangeCheckedParam<Real>(
       "viscosity",
@@ -30,7 +30,7 @@ PackedColumn::validParams()
 
 PackedColumn::PackedColumn(const InputParameters & parameters)
   : Material(parameters),
-    _radius(getParam<Real>("radius")),
+    _diameter(getParam<Real>("diameter")),
     _input_viscosity(getParam<Real>("viscosity")),
 
     // Declare two material properties by getting a reference from the MOOSE Material system
@@ -48,6 +48,6 @@ PackedColumn::PackedColumn(const InputParameters & parameters)
 void
 PackedColumn::computeQpProperties()
 {
-  _permeability[_qp] = _permeability_interpolation.sample(_radius);
+  _permeability[_qp] = _permeability_interpolation.sample(_diameter);
   _viscosity[_qp] = _input_viscosity;
 }
