### 2021-03-02 a1a5e4e
The paper specifies sphere diameter, not radius. Corrected all such references.
(closes idaholab/babbler#6) (refs #15232)
