+++ modules/doc/content/getting_started/examples_and_tutorials/tutorial01_app_development/problem_statement.md
@@ -4,7 +4,7 @@
 
 # Problem Statement
 
-Consider a system containing two pressure vessels at differing temperatures as in the experiments presented by [!cite](pamuk2012friction). The vessels are connected via a pipe that contains a filter consisting of close-packed steel spheres as shown in [problem-schematic]. Predict the velocity and temperature of the fluid inside the filter. The pipe is 0.304 m in length and 0.0514 m in diameter. The fluid inside the system is liquid water at approximately 30 degrees Celcius.
+Consider a system containing two pressure vessels at differing temperatures as in the experiments presented by [!cite](pamuk2012friction). The vessels are connected via a pipe that contains a filter consisting of close-packed steel spheres as shown in [problem-schematic]. Predict the velocity and temperature of the fluid inside the filter. The pipe is 0.304 m in length and 0.0514 m in diameter. The fluid inside the system is liquid water at approximately 30 degrees Celcius. Each of the steel spheres are 1 mm in diameter.
 
 !media tutorial01_app_development/problem_schematic.png
        style=width:80%;margin-left:auto;margin-right:auto;
@@ -75,9 +75,9 @@ Here, $c_{p}$ denotes specific heat and the subscripts, $f$ and $s$ refer to the
 The material properties of the fluid, $f$, and the solid, $s$, are given in [mats]. The permeability of the packed steel sphere medium that is present throughout the pipe is assumed to be isotropic. [!cite](pamuk2012friction) provides the following relationship:
 
 !equation id=permeability
-K(r) = \frac{1}{2} \begin{bmatrix} -r + 3 & r - 1 \end{bmatrix} \begin{Bmatrix} 0.8451 \\ 8.968 \end{Bmatrix} \times 10^{-9}, \, \, \, \forall \, r \in [1, 3]
+K(d) = \frac{1}{2} \begin{bmatrix} -d + 3 & d - 1 \end{bmatrix} \begin{Bmatrix} 0.8451 \\ 8.968 \end{Bmatrix} \times 10^{-9}, \, \, \, \forall \, d \in [1, 3]
 
-where $r$ is the radius ($\textrm{mm}$) of the spheres and $K$ denotes the scalar permeability ($\textrm{m}^{2}$), which is a linear function of $r$.
+where $d$ is the diameter ($\textrm{mm}$) of the spheres and $K$ denotes the scalar permeability ($\textrm{m}^{2}$), which is a linear function of $d$.
 
 !!!
 TODO: The properties listed here are somewhat inconsistent with what we input in the code. First off,
@@ -91,7 +91,7 @@ assumptions are...
 | Property | Value | Units |
 | :- | :- | :- |
 | Viscosity of water, $\mu_{f}$ | $7.98\times10^{-4}$ |  $\textrm{Pa}\cdot\textrm{s}$ |
-| Steel sphere radius, $r_{s}$ | $1$ | $\textrm{mm}$ |
+| Steel sphere diameter, $d_{s}$ | $1$ | $\textrm{mm}$ |
 | Density of water, $\rho_f$ | 996 | $\textrm{kg}/\textrm{m}^3$ |
 | Density of steel, $\rho_s$ | 8000 | $\textrm{kg}/\textrm{m}^3$ |
 | Thermal conductivity of water, $k_f$ | 0.600 | $\textrm{W}/\textrm{m}\,\textrm{K}$ |
