+++ framework/src/systems/NonlinearSystemBase.C
@@ -1275,8 +1275,6 @@ NonlinearSystemBase::constraintResiduals(NumericVector<Number> & residual, bool 
     }
   }
 
-  mortarConstraints(displaced);
-
   // go over element-element constraint interface
   std::map<unsigned int, std::shared_ptr<ElementPairLocator>> * element_pair_locators = nullptr;
 
@@ -1576,6 +1574,11 @@ NonlinearSystemBase::computeResidualInternal(const std::set<TagID> & tags)
   }
   PARALLEL_CATCH;
 
+  // undisplaced mortar constraints
+  mortarConstraints(false);
+  // displaced mortar constraints
+  mortarConstraints(true);
+
   if (_need_residual_copy)
   {
     _Re_non_time->close();
@@ -1599,7 +1602,7 @@ NonlinearSystemBase::computeResidualInternal(const std::set<TagID> & tags)
     _Re_non_time->close();
   }
 
-  // Add in Residual contributions from Constraints
+  // Add in Residual contributions from other Constraints
   if (_fe_problem._has_constraints)
   {
     PARALLEL_TRY
@@ -2099,8 +2102,6 @@ NonlinearSystemBase::constraintJacobians(bool displaced)
     }
   }
 
-  mortarConstraints(displaced);
-
   THREAD_ID tid = 0;
   // go over element-element constraint interface
   std::map<unsigned int, std::shared_ptr<ElementPairLocator>> * element_pair_locators = nullptr;
@@ -2537,20 +2538,18 @@ NonlinearSystemBase::computeJacobianInternal(const std::set<TagID> & tags)
   }
   PARALLEL_CATCH;
 
-  if (_fe_problem.hasMortarCoupling())
-    // don't shrink the memory allocation because our constraints may have additional sparsity
-    // pattern that hasn't been used up to this point. Note that some other types of constraints
-    // seem to rely on the matrix actually being closed, so we still close for any constraint types
-    // other mortar
-    flushTaggedMatrices(tags);
-  else
-    closeTaggedMatrices(tags);
+  // undisplaced mortar constraints
+  mortarConstraints(false);
+  // displaced mortar constraints
+  mortarConstraints(true);
+
+  closeTaggedMatrices(tags);
 
   // Have no idea how to have constraints work
   // with the tag system
   PARALLEL_TRY
   {
-    // Add in Jacobian contributions from Constraints
+    // Add in Jacobian contributions from other Constraints
     if (_fe_problem._has_constraints)
     {
       // Nodal Constraints
