### 2021-03-02 538f9fa
Remove unused temperature parameter for TensorMechanics action. Refs #17186
