### 2021-03-02 3d1cec6
new algorithm and test for computing rod positions
