### 2021-03-02 c535789
new algorithm and test for computing rod positions
