+++ modules/contact/src/constraints/ApplyPenetrationConstraintLMMechanicalContact.C
@@ -1,0 +1,203 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#include "ApplyPenetrationConstraintLMMechanicalContact.h"
+#include "PenetrationLocator.h"
+#include "PenetrationInfo.h"
+#include "SystemBase.h"
+#include "Assembly.h"
+
+#include "DualRealOps.h"
+
+using MetaPhysicL::DualNumber;
+
+registerMooseObject("ContactApp", ApplyPenetrationConstraintLMMechanicalContact);
+
+InputParameters
+ApplyPenetrationConstraintLMMechanicalContact::validParams()
+{
+  InputParameters params = NodeFaceConstraint::validParams();
+  params.set<bool>("use_displaced_mesh") = true;
+
+  params.addCoupledVar("disp_y", "The y displacement");
+  params.addCoupledVar("disp_z", "The z displacement");
+  params.addParam<Real>("c", 1, "Parameter for balancing the size of the gap and contact pressure");
+
+  params.addClassDescription(
+      "Implements the KKT conditions for normal contact using an NCP "
+      "function, in this case just the min function. This function enforces that either "
+      "the gap distance or the normal contact pressure (represented by the value of `variable`) is "
+      "zero.");
+  return params;
+}
+
+ApplyPenetrationConstraintLMMechanicalContact::ApplyPenetrationConstraintLMMechanicalContact(
+    const InputParameters & parameters)
+  : NodeFaceConstraint(parameters),
+    _c(getParam<Real>("c")),
+    _fe_problem(*getCheckedPointerParam<FEProblemBase *>("_fe_problem_base")),
+    _displaced_problem(_fe_problem.getDisplacedProblem().get()),
+    _residual_copy(_sys.residualGhosted())
+{
+}
+
+bool
+ApplyPenetrationConstraintLMMechanicalContact::shouldApply()
+{
+  _dof_number = _current_node->dof_number(_sys.number(), _var.number(), /*comp=*/0);
+  return _dof_number != DofObject::invalid_id;
+}
+
+bool
+ApplyPenetrationConstraintLMMechanicalContact::overwriteSecondaryResidual()
+{
+  // If we didn't project the secondary node onto the primary surface, then gap has no meaning and
+  // we just want to drive our Lagrange Multiplier to zero. If we did, then our LM drives the
+  // residual if it is less than the gap (we're using min for our NCP function)
+  return (!_projection_successful) ||
+         (_u_secondary[_qp] < _node_to_integrated_gap.at(_current_node));
+}
+
+void
+ApplyPenetrationConstraintLMMechanicalContact::residualSetup()
+{
+  NodeFaceConstraint::residualSetup();
+  _node_to_integrated_gap.clear();
+}
+
+Real
+ApplyPenetrationConstraintLMMechanicalContact::computeQpSecondaryValue()
+{
+  return _u_secondary[_qp];
+}
+
+void
+ApplyPenetrationConstraintLMMechanicalContact::computeResidual()
+{
+  DenseVector<Number> & re = _assembly.residualBlock(_var.number());
+
+  _qp = 0;
+
+  _secondary_residual = re(0) = computeQpResidual(Moose::Secondary);
+  _secondary_residual_computed = true;
+}
+
+void
+ApplyPenetrationConstraintLMMechanicalContact::computeJacobian()
+{
+  _Kee.resize(1, 1);
[…]
