### 2021-03-02 daca20e
Improve error message and handling of app syntax failurs in MooseDocs
(refs #15971)
