+++ framework/doc/content/sqa/inl_records.md
@@ -1,0 +1,23 @@
+# Accessing INL Software Quality Records
+
+Several software quality-related documents and records listed below are viewable by those with access to the INL network. Instructions on how to access these documents are provided below. These are stored in the same way for each MOOSE-based application, so where appropriate, use the name of the actual application of interest when searching.
+
+## PLN-4005, "Software Quality Assurance Plan for MOOSE and MOOSE-Based Applications"
+1. From the +Nucleus+ Home page select +A-Z+
+2. From the alphabetical list select +EDMS+
+3. Enter +PLN-4005+ in the "Document Id" field and select +Submit+.
+4. Select the +Doc Viewer down arrow+ and then select the +Record File pdf+.
+
+## Enterprise Architecture (EA) record for a MOOSE-Based Application id=ea_record
+1. From the Nucleus Home page select +A-Z+
+2. From the alphabetical list select +CTM –Capabilities & Technology Management+
+3. Begin typing the name of the application in the search box in the upper right corner of the screen
+4. When the search is complete, click on the ellipses (three dots) to the left of the application name and then select +View Details+
+
+## Safety Software Determination (SSD) record for a MOOSE-Based Application
+1. From the Enterprise Architecture +Details+ screen for the application (refer to [#ea_record]), select the +General+ tab
+2. Scroll to the bottom of the screen and in the +Related Artifacts+ section select +Safety Software Determination+
+
+## Quality Level Determination (QLD) record for a MOOSE-Based Application
+1. From the Enterprise Architecture +Details+ screen for the application (refer to [#ea_record]), select the +General+ tab
+2. Scroll to the bottom of the screen and in the +Related Artifacts+ section select +Quality Level Determination+
