### 2021-03-03 e8f2d76
Almost all working refs #15444
Fixup a couple of tests refs #15444

Remove debugging statement refs #15444

Clang format #15444

Fixup build error refs #15444
