+++ framework/src/base/MooseApp.C
@@ -305,7 +305,6 @@ MooseApp::MooseApp(InputParameters parameters)
     _pars(parameters),
     _type(getParam<std::string>("_type")),
     _comm(getParam<std::shared_ptr<Parallel::Communicator>>("_comm")),
-    _rank_map(*_comm, _perf_graph),
     _file_base_set_by_user(false),
     _output_position_set(false),
     _start_time_set(false),
@@ -313,6 +312,7 @@ MooseApp::MooseApp(InputParameters parameters)
     _global_time_offset(0.0),
     _output_warehouse(*this),
     _perf_graph(type() + " (" + name() + ')', *this, getParam<bool>("perf_graph_live_all")),
+    _rank_map(*_comm, _perf_graph),
     _input_parameter_warehouse(new InputParameterWarehouse()),
     _action_factory(*this),
     _action_warehouse(*this, _syntax, _action_factory),
