### 2021-03-03 e5563f9
Prevent repeat failed projections of primary nodes
