### 2021-03-03 567a119
Don't call dof_number if not n_dofs
