### 2021-03-03 e3f8f5e
Adjust ncpus appropriately
After applying the math to adjust node count for min_parallel, set ncpus
to what the end user supplied.

Closes #17204
