### 2021-03-03 e436f34
Update framework/doc/content/syntax/Materials/index.md
Co-authored-by: Guillaume Giudicelli <guillaume.giudicelli@gmail.com>
