### 2021-03-03 19c74a8
Use pool unique_ptr as storage type for rays
refs #17208
