### 2021-03-03 c665199
Use pool unique_ptr as storage type for rays
refs #17208
