+++ modules/ray_tracing/src/raybcs/KillRayBC.C
@@ -25,5 +25,5 @@ void
 KillRayBC::onBoundary(const unsigned int /* num_applying */)
 {
   // After RayBCs are completed, ray->shouldContinue() is checked and this will kill the Ray
-  currentRay()->setShouldContinue(false);
+  _current_ray->setShouldContinue(false);
 }

[…]
