+++ modules/ray_tracing/include/raytracing/ClaimRays.h
@@ -23,6 +23,7 @@
 #include "libmesh/bounding_box.h"
 #include "libmesh/parallel_object.h"
 #include "libmesh/point_locator_base.h"
+#include "timpi/request.h"
 
 // Forward declarations
 class RayTracingStudy;
@@ -171,3 +172,189 @@ private:
   /// Whether or not an init is needed (bounding boxes, neighbors)
   bool _needs_init;
 };
+
+// template <typename MapToContainers, typename ActionFunctor, typename Context>
+// struct PushParallelHelper
+// {
+// public:
+//   PushParallelHelper(const TIMPI::Communicator & comm) : _comm(comm), _tag(comm.get_unique_tag())
+//   {}
+//
+//   template <
+//       typename T,
+//       class = typename std::enable_if<!std::is_lvalue_reference<MapToContainers>::value>::type>
+//   void send(MapToContainers data) {}
+//
+//   typedef typename MapToContainers::value_type map_pair_type;
+//   typedef typename map_pair_type::second_type container_type;
+//   typedef typename container_type::value_type nonref_type;
+//   typedef typename std::remove_const<nonref_type>::type nonconst_nonref_type;
+//
+// private:
+//   const TIMPI::Communicator & _comm;
+//
+//   MessageTag _tag;
+//   std::list<TIMPI::Request> reqs;
+// };
+//
+// template <typename MapToContainers, typename ActionFunctor, typename Context>
+// void
+// new_push_parallel_packed_range(const TIMPI::Communicator & comm,
+//                                const MapToContainers & data,
+//                                Context * context,
+//                                const ActionFunctor & act_on_data)
+// {
+//   PushParallelHelper<MapToContainers, ActionFunctor, Context> helper(comm);
+//
+//   // This function must be run on all processors at once
+//   timpi_parallel_only(comm);
+//
+//   // This function implements the "NBX" algorithm from
+//   // https://htor.inf.ethz.ch/publications/img/hoefler-dsde-protocols.pdf
+//
+//   typedef typename MapToContainers::value_type map_pair_type;
+//   typedef typename map_pair_type::second_type container_type;
+//   typedef typename container_type::value_type nonref_type;
+//   typedef typename std::remove_const<nonref_type>::type nonconst_nonref_type;
+//
+//   // We'll grab a tag so we can overlap request sends and receives
+//   // without confusing one for the other
+//   auto tag = comm.get_unique_tag();
+//
+//   // Post all of the sends, non-blocking and synchronous
+//
+//   // Save off the old send_mode so we can restore it after this
+//   auto old_send_mode = comm.send_mode();
+//
+//   // Set the sending to synchronous - this is so that we can know when
+//   // the sends are complete
+//   const_cast<TIMPI::Communicator &>(comm).send_mode(TIMPI::Communicator::SYNCHRONOUS);
+//
+//   // The send requests
+//   std::list<TIMPI::Request> reqs;
+//
+//   processor_id_type num_procs = comm.size();
+//
+//   for (auto & datapair : data)
+//   {
+//     // In the case of data partitioned into more processors than we
+//     // have ranks, we "wrap around"
+//     processor_id_type destid = datapair.first % num_procs;
+//     auto & datum = datapair.second;
+//
+//     // Just act on data if the user requested a send-to-self
+//     if (destid == comm.rank())
+//       act_on_data(destid, datum);
+//     else
+//     {
+//       TIMPI::Request sendreq;
+//       comm.nonblocking_send_packed_range(destid, context, datum.begin(), datum.end(), sendreq,
+//       tag); reqs.push_back(sendreq);
+//     }
+//   }
+//
+//   bool sends_complete = reqs.empty();
+//   bool started_barrier = false;
+//   TIMPI::Request barrier_request;
+//
+//   // Receive
+//
[…]
