+++ modules/ray_tracing/src/raytracing/ClaimRays.C
@@ -20,10 +20,9 @@
 #include "libmesh/mesh_tools.h"
 
 ClaimRays::ClaimRays(RayTracingStudy & study,
-                     ParallelStudy<std::shared_ptr<Ray>, Ray> & parallel_study,
+                     ParallelStudy<MooseUtils::SharedPool<Ray>::PtrType, Ray> & parallel_study,
                      MooseMesh & mesh,
-                     const std::vector<std::shared_ptr<Ray>> & rays,
-                     std::vector<std::shared_ptr<Ray>> & local_rays,
+                     std::vector<MooseUtils::SharedPool<Ray>::PtrType> & rays,
                      const bool do_exchange)
   : ParallelObject(study.comm()),
     MeshChangedInterface(study.parameters()),
@@ -33,7 +32,7 @@ ClaimRays::ClaimRays(RayTracingStudy & study,
     _study(study),
     _parallel_study(parallel_study),
     _rays(rays),
-    _local_rays(local_rays),
+    _point_locator(nullptr),
     _needs_init(true)
 {
 }
@@ -49,57 +48,62 @@ ClaimRays::claim()
 
   preClaim();
 
-  // Clear these as we're about to fill
-  _local_rays.clear();
+  // Move the user's rays into _temp_rays, so that we can look through them
+  // and move (and receive) rays that we claim into _rays
+  _temp_rays.clear();
+  _temp_rays.swap(_rays);
 
   // Grab the point locator
   _point_locator = PointLocatorBase::build(TREE_LOCAL_ELEMENTS, _mesh.getMesh());
   _point_locator->enable_out_of_mesh_mode();
 
   // Exchange: filter Rays into processors that _may_ claim them
-  std::unordered_map<processor_id_type, std::vector<std::shared_ptr<Ray>>> rays_to_send;
+  std::unordered_map<processor_id_type, std::vector<MooseUtils::SharedPool<Ray>::PtrType>>
+      rays_to_send;
   if (_do_exchange)
     for (processor_id_type pid = 0; pid < comm().size(); ++pid)
       if (_pid != pid)
       {
         const BoundingBox & pid_bbox = inflatedBoundingBox(pid);
-        for (auto & ray : _rays)
+        for (const auto & ray : _temp_rays)
           if (pid_bbox.contains_point(ray->currentPoint()))
-            rays_to_send[pid].push_back(ray);
+            rays_to_send[pid].emplace_back(_study.acquireCopiedRay(*ray));
       }
 
   // Functor for possibly claiming a vector of Rays
   auto claim_functor = [&](processor_id_type /* pid */,
-                           const std::vector<std::shared_ptr<Ray>> & rays) {
+                           const std::vector<MooseUtils::SharedPool<Ray>::PtrType> & rays) {
     for (auto & ray : rays)
-      possiblyClaim(ray);
+      possiblyClaim(_study.acquireCopiedRay(*ray));
   };
 
   // Send the relevant Rays to everyone and then attempt to claim the ones that we receive
   if (_do_exchange)
-    Parallel::push_parallel_packed_range(comm(), rays_to_send, &_parallel_study, claim_functor);
+    push_parallel_packed_range(comm(), rays_to_send, &_parallel_study, claim_functor);
 
   // Attempt to claim the locally generated rays in _rays
-  claim_functor(_pid, _rays);
+  claim_functor(_pid, _temp_rays);
 
   // Verify the claiming if the study so desires
   if (_study.verifyRays())
     verifyClaiming();
 
+  _temp_rays.clear();
+
   postClaim();
 }
 
 void
[…]
