+++ modules/ray_tracing/include/raytracing/ClaimRays.h
@@ -23,6 +23,7 @@
 #include "libmesh/bounding_box.h"
 #include "libmesh/parallel_object.h"
 #include "libmesh/point_locator_base.h"
+#include "timpi/request.h"
 
 // Forward declarations
 class RayTracingStudy;
@@ -171,3 +172,234 @@ private:
   /// Whether or not an init is needed (bounding boxes, neighbors)
   bool _needs_init;
 };
+
+template <typename MapToContainers,
+          typename SendFunctor,
+          typename PossiblyReceiveFunctor,
+          typename ActionFunctor>
+void
+push_parallel_helper(const TIMPI::Communicator & comm,
+                     MapToContainers & data,
+                     SendFunctor & send_functor,
+                     PossiblyReceiveFunctor & possibly_receive_functor,
+                     ActionFunctor & act_on_data)
+{
+  typedef typename MapToContainers::value_type::second_type container_type;
+
+  // This function must be run on all processors at once
+  timpi_parallel_only(comm);
+
+  // We'll grab a tag so we can overlap request sends and receives
+  // without confusing one for the other
+  auto tag = comm.get_unique_tag();
+
+  // Save off the old send_mode so we can restore it after this
+  auto old_send_mode = comm.send_mode();
+
+  // Set the sending to synchronous - this is so that we can know when
+  // the sends are complete
+  const_cast<TIMPI::Communicator &>(comm).send_mode(TIMPI::Communicator::SYNCHRONOUS);
+
+  // The send requests
+  std::list<TIMPI::Request> requests;
+
+  for (auto & datapair : data)
+  {
+    // In the case of data partitioned into more processors than we
+    // have ranks, we "wrap around"
+    processor_id_type dest_pid = datapair.first % comm.size();
+    auto & datum = datapair.second;
+
+    // Just act on data if the user requested a send-to-self
+    if (dest_pid == comm.rank())
+      act_on_data(dest_pid, datum);
+    else
+    {
+      requests.emplace_back();
+      send_functor(comm, dest_pid, datum, requests.back(), tag);
+    }
+  }
+
+  // In serial we've now acted on all our data.
+  if (comm.size() == 1)
+    return;
+
+  // Whether or not all of the sends are complete
+  bool sends_complete = requests.empty();
+
+  // Whether or not the nonblocking barrier has started
+  bool started_barrier = false;
+  // Request for the nonblocking barrier
+  TIMPI::Request barrier_request;
+
+  // The pair of src_pid and requests
+  std::list<std::pair<unsigned int, std::shared_ptr<TIMPI::Request>>> receive_requests;
+  auto current_request = std::make_shared<TIMPI::Request>();
+
+  // Storage for the incoming data
+  std::multimap<processor_id_type, std::shared_ptr<container_type>> incoming_data;
+  auto current_incoming_data = std::make_shared<container_type>();
+
+  unsigned int current_src_proc = 0;
+
+  // Keep looking for receives
+  while (true)
+  {
+    // Look for data from anywhere
+    current_src_proc = TIMPI::any_source;
+
+    // Check if there is a message and start receiving it
+    if (possibly_receive_functor(
+            comm, current_src_proc, *current_incoming_data, *current_request, tag))
+    {
+      receive_requests.emplace_back(current_src_proc, current_request);
+      current_request = std::make_shared<TIMPI::Request>();
+
+      // current_src_proc will now hold the src pid for this receive
+      incoming_data.emplace(current_src_proc, current_incoming_data);
+      current_incoming_data = std::make_shared<container_type>();
+    }
+
+    // Clean up outstanding receive requests
+    receive_requests.remove_if(
+        [&act_on_data,
[…]
