+++ modules/ray_tracing/include/utils/ParallelStudy.h
@@ -29,9 +29,9 @@ template <typename WorkType, typename ParallelDataType>
 class ParallelStudy : public ParallelObject
 {
 public:
+  typedef typename MooseUtils::SharedPool<ParallelDataType>::PtrType ParallelDataPoolPtr;
   typedef typename MooseUtils::Buffer<WorkType>::iterator work_iterator;
-  typedef typename MooseUtils::Buffer<std::shared_ptr<ParallelDataType>>::iterator
-      parallel_data_iterator;
+  typedef typename MooseUtils::Buffer<ParallelDataPoolPtr>::iterator parallel_data_iterator;
 
   static InputParameters validParams();
 
@@ -60,8 +60,8 @@ public:
    * During execute(), this method is thread safe and can be used to add work during execution.
    */
   ///@{
-  void moveWorkToBuffer(WorkType & work, const THREAD_ID tid);
-  void moveWorkToBuffer(const work_iterator begin, const work_iterator end, const THREAD_ID tid);
+  void moveWorkToBuffer(WorkType && work, const THREAD_ID tid);
+  void moveWorkToBuffer(work_iterator begin, work_iterator end, const THREAD_ID tid);
   void moveWorkToBuffer(std::vector<WorkType> & work, const THREAD_ID tid);
   ///@}
 
@@ -69,8 +69,7 @@ public:
    * Acquire a parallel data object from the pool.
    */
   template <typename... Args>
-  typename MooseUtils::SharedPool<ParallelDataType>::PtrType
-  acquireParallelData(const THREAD_ID tid, Args &&... args)
+  ParallelDataPoolPtr acquireParallelData(const THREAD_ID tid, Args &&... args)
   {
     return _parallel_data_pools[tid].acquire(std::forward<Args>(args)...);
   }
@@ -78,13 +77,12 @@ public:
   /**
    * Moves parallel data objects to the send buffer to be communicated to processor \p dest_pid.
    */
-  void moveParallelDataToBuffer(std::shared_ptr<ParallelDataType> & data,
-                                const processor_id_type dest_pid);
+  void moveParallelDataToBuffer(ParallelDataPoolPtr && data, const processor_id_type dest_pid);
 
   /**
    * Gets the receive buffer.
    */
-  const ReceiveBuffer<ParallelDataType, ParallelStudy<WorkType, ParallelDataType>> &
+  const ReceiveBuffer<ParallelDataPoolPtr, ParallelStudy<WorkType, ParallelDataType>> &
   receiveBuffer() const
   {
     return *_receive_buffer;
@@ -137,6 +135,10 @@ public:
    * Whether or not this object is between preExecute() and execute().
    */
   bool currentlyPreExecuting() const { return _currently_pre_executing; }
+  /**
+   * Whether or not work is currently being executed within executeAndBuffer().
+   */
+  bool currentlyExecutingWork() const { return _currently_executing_work; }
 
   /**
    * Gets the max buffer size
@@ -177,18 +179,6 @@ public:
 
 protected:
   /**
-   * Enum for providing useful errors during work addition in moveWorkError().
-   */
-  enum MoveWorkError
-  {
-    DURING_EXECUTION_DISABLED,
-    PRE_EXECUTION_AND_EXECUTION_ONLY,
-    PRE_EXECUTION_ONLY,
-    PRE_EXECUTION_THREAD_0_ONLY,
-    CONTINUING_DURING_EXECUTING_WORK
-  };
-
-  /**
    * Creates the work buffer
    *
    * This is virtual so that derived classes can use their own specialized buffers
@@ -198,12 +188,7 @@ protected:
   /**
    * Pure virtual to be overridden that executes a single object of work on a given thread
    */
[…]

[…]
