### 2021-03-03 878800c
Leave smaller jobs alone
Only create 'Chunks' when required. That is, if the job requires less than the
number of cores being specified with --pbs-node-cpus, do not alter the select
statement (select=1:ncpus=min_parallel).

Refs #17204
