### 2021-03-03 ea51616
Make sure perfgraph doesn't print during dump refs #15444
