+++ conda/petsc/conda_build_config.yaml
@@ -17,16 +17,16 @@ moose_mpich:
   - 3.3.2 build_5
 
 libblas:
-  - 3.9.0 8_openblas           # [osx]
+  - 3.9.0 8_mkl                # [osx]
 
 libcblas:
-  - 3.9.0 8_openblas           # [osx]
+  - 3.9.0 8_mkl                # [osx]
 
 liblapack:
-  - 3.9.0 8_openblas           # [osx]
+  - 3.9.0 8_mkl                # [osx]
 
 libopenblas:
-  - 0.3.12 openmp_h63d9170_1   # [osx]
+  - 0.3.13 openmp_hb927a7c_0   # [osx]
 
 pin_run_as_build:
   moose_mpich: x.x.x
