### 2021-03-03 bc617a3
Fix vkt/petsc version dependency on mpich build
Missed a couple things... MPICH gets rebuilt, everything must get rebuilt.
