+++ conda/libmesh-vtk/meta.yaml
@@ -1,4 +1,4 @@
-{% set build = 5 %}
+{% set build = 6 %}
 {% set strbuild = "build_" + build|string %}
 
 {% set vtk_version = vtk_version %}
