+++ conda/libmesh-vtk/conda_build_config.yaml
@@ -29,7 +29,7 @@ mpi:
   - moose-mpich
 
 moose_mpich:
-  - 3.3.2 build_5
+  - 3.3.2 build_6
 
 pin_run_as_build:
   moose_mpich: x.x.x
