+++ framework/src/transfers/MultiAppTransfer.C
@@ -212,12 +212,12 @@ MultiAppTransfer::getFromBoundingBoxes(BoundaryID boundary_id)
     {
       if (bnode->_bnd_id == boundary_id && bnode->_node->processor_id() == processor_id())
       {
+        at_least_one = true;
         const auto & node = *bnode->_node;
         for (unsigned int i = 0; i < LIBMESH_DIM; ++i)
         {
-          min = std::min(min(i), node(i));
-          max = std::max(max(i), node(i));
-          at_least_one = true;
+          min(i) = std::min(min(i), node(i));
+          max(i) = std::max(max(i), node(i));
         }
       }
     }
