+++ python/moosesqa/get_documents.py
@@ -12,20 +12,30 @@ import collections
 import mooseutils
 
 @mooseutils.addProperty('name', ptype=str, required=True)
+@mooseutils.addProperty('title', ptype=str, required=True)
 @mooseutils.addProperty('filename', ptype=str)
 class Document(mooseutils.AutoPropertyMixin):
     pass
 
-INL_DOCUMENTS = ['safety_software_determination', 'quality_level_determination', 'enterprise_architecture_entry', # Initiation
-                 'project_management_plan', 'software_quality_plan', 'configuration_management_plan', # Planning
-                 'software_requirements_specification', # Requirements
-                 'software_design_description', # Design
-                 'software_test_plan', 'requirements_traceablity_matrix', 'asset_management_plan', 'verification_validation_plan', # Testing
-                 'user_manual', 'theory_manual', # Other
-                 'failure_analysis_report'] # PLN-4005 v8
+INL_DOCUMENTS = ['safety_software_determination',
+                 'quality_level_determination',
+                 'enterprise_architecture_entry',
+                 'project_management_plan',
+                 'software_quality_plan',
+                 'configuration_management_plan',
+                 'software_test_plan',
+                 'asset_management_plan',
+                 'verification_validation_plan',
+                 'software_requirements_specification',
+                 'software_design_description',
+                 'requirements_traceablity_matrix',
+                 'verification_validation_report',
+                 'failure_analysis_report',
+                 'user_manual',
+                 'theory_manual']
 
 def get_documents(required_docs=INL_DOCUMENTS, **kwargs):
     """
     Build SQA document dictionary from the provided directories.
     """
-    return [Document(name=name, filename=kwargs.get(name, None)) for name in required_docs]
+    return [Document(name=name, title=name.replace('_', ' ').title(), filename=kwargs.get(name, None)) for name in required_docs]
