+++ test/doc/sqa_reports.yml
@@ -1,40 +1,1 @@
-Applications:
-    framework:
-        exe_directory: test
-        exe_name: moose_test
-        app_types:
-            - MooseApp
-        content_directory: framework/doc/content
-        unregister:
-            - framework/doc/unregister.yml
-        remove:
-            - framework/doc/remove.yml
-            - test/doc/remove.yml
-        log_hidden_syntax: WARNING
-        log_stub_files: WARNING
-        show_warning: false
-
-Documents:
-    software_requirements_specification: sqa/moose_srs.md
-    software_design_description: sqa/moose_sdd.md
-    software_test_plan: sqa/moose_stp.md
-    requirements_traceablity_matrix: sqa/moose_rtm.md
-    verification_validation_plan: sqa/moose_vvr.md
-    log_safety_software_determination: WARNING
-    log_quality_level_determination: WARNING
-    log_enterprise_architecture_entry: WARNING
-    log_project_management_plan: WARNING
-    log_software_quality_plan: WARNING
-    log_configuration_management_plan: WARNING
-    log_asset_management_plan: WARNING
-    log_verification_validation_plan: WARNING
-    log_user_manual: WARNING
-    log_theory_manual: WARNING
-    show_warning: false
-
-Requirements:
-    moose_test:
-        directories:
-            - test/tests
-        log_testable: WARNING
-        show_warning: false
+../../framework/doc/sqa_reports.yml
