+++ modules/stochastic_tools/doc/sqa_reports.yml
@@ -11,7 +11,7 @@ Documents:
     software_design_description: sqa/stochastic_tools_sdd.md
     software_test_plan: sqa/stochastic_tools_stp.md
     requirements_traceablity_matrix: sqa/stochastic_tools_rtm.md
-    verification_validation_plan: sqa/stochastic_tools_vvr.md
+    verification_validation_report: sqa/stochastic_tools_vvr.md
     failure_analysis_report: sqa/stochastic_tools_far.md
     log_safety_software_determination: WARNING
     log_quality_level_determination: WARNING

[…]
