+++ python/MooseDocs/extensions/katex.py
@@ -13,7 +13,7 @@ import moosetree
 from .. import common
 from ..base import components, renderers
 from ..tree import tokens, html, latex
-from . import command, core, floats
+from . import command, core, floats, heading
 
 def make_extension(**kwargs):
     """Create an instance of the Extension object."""
@@ -25,6 +25,7 @@ LatexBlockEquation = tokens.newToken('LatexBlockEquation',
                                      number=None,
                                      bookmark=None)
 LatexInlineEquation = tokens.newToken('LatexInlineEquation', tex=r'', bookmark=None)
+EquationReference = tokens.newToken('EquationReference', label=None, filename=None)
 
 class KatexExtension(command.CommandExtension):
     """
@@ -46,13 +47,16 @@ class KatexExtension(command.CommandExtension):
         """
         Add the necessary components for reading and rendering LaTeX.
         """
-        self.requires(core, floats)
+        self.requires(core, floats, heading)
         self.addCommand(reader, KatexBlockEquationCommand())
+        self.addCommand(reader, EquationReferenceCommand())
+
         reader.addInline(KatexBlockEquationComponent(), location='_begin')
         reader.addInline(KatexInlineEquationComponent(), location='_begin')
         renderer.add('LatexBlockEquation', RenderLatexEquation())
         renderer.add('LatexInlineEquation', RenderLatexEquation())
         renderer.add('ShortcutLink', RenderEquationLink())
+        renderer.add('EquationReference', RenderEquationReference())
 
         if isinstance(renderer, renderers.LatexRenderer):
             renderer.addPackage('amsfonts')
@@ -60,21 +64,21 @@ class KatexExtension(command.CommandExtension):
                 for k, v in self.get('macros').items():
                     renderer.addNewCommand(k, v)
 
+    def initPage(self, page):
+        page['labels'] = dict()
+
     def postTokenize(self, page, ast):
-        labels = set()
+        labels = dict()
         count = 0
         func = lambda n: (n.name == 'LatexBlockEquation') and (n['label'] is not None)
         for node in moosetree.iterate(ast, func):
             count += 1
             node['number'] = count
-            if node['label']:
-                if isinstance(self.translator.renderer, renderers.LatexRenderer):
-                    labels.add(node['label'])
-                else:
-                    core.Shortcut(ast,
-                                  key=node['label'],
-                                  string='{} ({})'.format(self.get('prefix'), count),
-                                  link='#{}'.format(node['bookmark']))
+            labels[node['label']] = (count, node['bookmark'])
+            core.Shortcut(ast,
+                          key=node['label'],
+                          string='{} ({})'.format(self.get('prefix'), count),
+                          link='#{}'.format(node['bookmark']))
 
         page['labels'] = labels
 
@@ -160,6 +164,25 @@ class KatexInlineEquationComponent(components.ReaderComponent):
         LatexInlineEquation(parent, tex=tex, bookmark=eq_id)
         return parent
 
+class EquationReferenceCommand(command.CommandComponent):
+    COMMAND = 'eqref'
+    SUBCOMMAND = None
+    LABEL_RE = re.compile(r'((?P<filename>.*?\.md)#)?(?P<label>.+)', flags=re.UNICODE)
+
+    @staticmethod
+    def defaultSettings():
+        settings = command.CommandComponent.defaultSettings()
+        return settings
+
+    def createToken(self, parent, info, page):
+        content = info['inline'] if ('inline' in info) else info['block']
+        match = self.LABEL_RE.search(content)
+        if match is None:
+            raise common.exceptions.MooseDocsException("Invalid equation label format.")
+
+        EquationReference(parent, label=match.group('label'), filename=match.group('filename'))
[…]
