### 2021-03-04 0e8faf0
Deprecate latex style commands
(refs #17218)
