+++ python/MooseDocs/extensions/katex.py
@@ -9,22 +9,20 @@
 import sys
 import re
 import uuid
+import logging
 import moosetree
 from .. import common
 from ..base import components, renderers
 from ..tree import tokens, html, latex
 from . import command, core, floats, heading
 
+LOG = logging.getLogger(__name__)
+
 def make_extension(**kwargs):
     """Create an instance of the Extension object."""
     return KatexExtension(**kwargs)
 
-LatexBlockEquation = tokens.newToken('LatexBlockEquation',
-                                     tex=r'',
-                                     label=None,
-                                     number=None,
-                                     bookmark=None)
-LatexInlineEquation = tokens.newToken('LatexInlineEquation', tex=r'', bookmark=None)
+Equation = tokens.newToken('Equation', tex=r'', inline=False, label=None, number=None, bookmark=None)
 EquationReference = tokens.newToken('EquationReference', label=None, filename=None)
 
 class KatexExtension(command.CommandExtension):
@@ -48,15 +46,16 @@ class KatexExtension(command.CommandExtension):
         Add the necessary components for reading and rendering LaTeX.
         """
         self.requires(core, floats, heading)
-        self.addCommand(reader, KatexBlockEquationCommand())
+        self.addCommand(reader, EquationCommand())
         self.addCommand(reader, EquationReferenceCommand())
 
+        renderer.add('Equation', RenderEquation())
+        renderer.add('EquationReference', RenderEquationReference())
+
+        # Deprecated
         reader.addInline(KatexBlockEquationComponent(), location='_begin')
         reader.addInline(KatexInlineEquationComponent(), location='_begin')
-        renderer.add('LatexBlockEquation', RenderLatexEquation())
-        renderer.add('LatexInlineEquation', RenderLatexEquation())
         renderer.add('ShortcutLink', RenderEquationLink())
-        renderer.add('EquationReference', RenderEquationReference())
 
         if isinstance(renderer, renderers.LatexRenderer):
             renderer.addPackage('amsfonts')
@@ -65,12 +64,15 @@ class KatexExtension(command.CommandExtension):
                     renderer.addNewCommand(k, v)
 
     def initPage(self, page):
+
+        # Contains a mapping of the equation label ("id=") to a tuple that contains
+        # the equation number and unique bookmark. This used for equation referencing.
         page['labels'] = dict()
 
     def postTokenize(self, page, ast):
         labels = dict()
         count = 0
-        func = lambda n: (n.name == 'LatexBlockEquation') and (n['label'] is not None)
+        func = lambda n: (n.name == 'Equation') and (n['label'] is not None)
         for node in moosetree.iterate(ast, func):
             count += 1
             node['number'] = count
@@ -82,19 +84,18 @@ class KatexExtension(command.CommandExtension):
 
         page['labels'] = labels
 
-        if common.has_tokens(ast, 'LatexBlockEquation', 'LatexInlineEquation'):
-            renderer = self.translator.renderer
-            if isinstance(renderer, renderers.HTMLRenderer):
-                renderer.addCSS('katex', "contrib/katex/katex.min.css", puid=page.uid)
-                renderer.addCSS('katex_moose', "css/katex_moose.css", puid=page.uid)
-                renderer.addJavaScript('katex', "contrib/katex/katex.min.js", head=True, puid=page.uid)
+        renderer = self.translator.renderer
+        if common.has_tokens(ast, 'Equation', 'LatexInlineEquation') and isinstance(renderer, renderers.HTMLRenderer):
+            renderer.addCSS('katex', "contrib/katex/katex.min.css", puid=page.uid)
[…]
