### 2021-03-04 098f795
Remove need to call NonlinearSystemBase::mortarConstraints twice
It's more transparent just to have a single call to `mortarConstraints`
that executes both undisplaced and displaced mortar constraints. The old
way was motivated by having computation of mortar constraints occur in
`constraintResiduals/constraintJacobians` but we now compute mortar
constraints ahead of those methods in order to ensure that mortar
contributions are assembled into the global residual/Jacobian before we
evaluate other constraints
