### 2021-03-04 4f928a1
Use getFieldVar for mixing length kernels
