+++ modules/navier_stokes/src/fvkernels/INSFVMixingLengthScalarDiffusion.C
@@ -32,14 +32,13 @@ INSFVMixingLengthScalarDiffusion::validParams()
 INSFVMixingLengthScalarDiffusion::INSFVMixingLengthScalarDiffusion(const InputParameters & params)
   : FVFluxKernel(params),
     _dim(_subproblem.mesh().dimension()),
-    _u_var(dynamic_cast<const INSFVVelocityVariable *>(
-        getFieldVar(params.get<std::vector<VariableName>>("u").front(), 0))),
-    _v_var(isParamValid("v") ? dynamic_cast<const INSFVVelocityVariable *>(getFieldVar(
-                                   params.get<std::vector<VariableName>>("v").front(), 0))
-                             : nullptr),
-    _w_var(isParamValid("w") ? dynamic_cast<const INSFVVelocityVariable *>(getFieldVar(
-                                   params.get<std::vector<VariableName>>("w").front(), 0))
-                             : nullptr),
+    _u_var(dynamic_cast<const INSFVVelocityVariable *>(getFieldVar("u", 0))),
+    _v_var(params.isParamValid("v")
+               ? dynamic_cast<const INSFVVelocityVariable *>(getFieldVar("v", 0))
+               : nullptr),
+    _w_var(params.isParamValid("w")
+               ? dynamic_cast<const INSFVVelocityVariable *>(getFieldVar("w", 0))
+               : nullptr),
     _mixing_len(coupledValue("mixing_length")),
     _schmidt_number(getParam<Real>("schmidt_number"))
 {
