### 2021-03-04 8f36ec7
Add unit tests for katex extension
(refs #17218)
