+++ python/MooseDocs/test/gold/latex/extensions/katex.tex
@@ -1,43 +1,0 @@
-
-\chapter{\label{katex-extension}KaTeX Extension}
-\section{\label{inline}Inline}
-\par Inline $y=ax+b$.
-\section{\label{block}Block}
-\par 
-\begin{equation*}
-a^2 + b^2 = c^2
-\end{equation*}
-
-\par 
-\begin{equation*}
-G_{\mu\nu} = 8\pi G(T_{\mu\nu} + \rho_{\Lambda} g_{\mu\nu})
-\end{equation*}
-
-\par 
-\begin{equation*}
-t' = t \frac{1}{\sqrt{1 - \frac{\nu^2}{c^2}}}
-\end{equation*}
-
-\par 
-\begin{equation*}
-F = ma
-\end{equation*}
- 
-\par 
-\begin{equation}
-\label{calculus}
- \int_a^b f'(x)dx = f(b) - f(a)
-\end{equation}
-
-\par The fundamental theorem of calculus: Eq.~\eqref{calculus}.
-\section{\label{macros}Macros}
-\par 
-\begin{equation*}
-c = \pm\sqrt{a^2 + b^2}\in\RR
-\end{equation*}
-
-\par 
-\begin{equation*}
-\pf{T}{t} = c
-\end{equation*}
- 
