### 2021-03-04 131426e
Update documentation and remove deprecation warning
(refs #17218)
