### 2021-03-04 50a5a53
make multiapp create apps right after its construction #17223
