### 2021-03-04 1214360
add Aydin's algo and fix the test point sorting
