### 2021-03-04 c665b2b
Make member vars between Density and ADDensity consistent
refs #17229
