### 2021-03-04 e510fab
actually use the new rod position algorithm
fixes #29209
