+++ modules/subchannel/src/mesh/TriSubChannelMesh.C
@@ -20,30 +20,6 @@ TriSubChannelMesh::validParams()
 void
 rodPositions(std::vector<Point> & positions, Real nrings, Real pitch, Point center)
 {
-  positions.clear();
-
-  const Real start_r = (nrings - 1) * pitch;
-  const Real theta0 = 2 * libMesh::pi / 3;
-  const Real startx = start_r * std::cos(theta0);
-  const Real starty = start_r * std::sin(theta0);
-  for (int i = 0; i < nrings; i++)
-  {
-    int n_rods_in_row = nrings + i;
-    const Real y = starty - i * pitch * std::sin(theta0);
-    const Real x_row = startx + i * pitch * std::cos(theta0);
-    for (int j = 0; j < n_rods_in_row; j++)
-    {
-      const Real x = x_row + j * pitch;
-      positions.emplace_back(center(0) + x, center(1) + y);
-      if (i < nrings - 1)
-        positions.emplace_back(center(0) + x, center(1) - y);
-    }
-  }
-}
-
-void
-rodPositions2(std::vector<Point> & positions, Real nrings, Real pitch, Point center)
-{
   Real teta = 0.0;
   Real dteta = 0.0;
   Real distance = 0.0;
@@ -71,32 +47,21 @@ rodPositions2(std::vector<Point> & positions, Real nrings, Real pitch, Point cen
       else
       {
         if (teta > pi / 3.0 && teta <= 2.0 / 3.0 * pi)
-        {
           teta_corrected = teta_corrected + pi / 3.0;
-        }
         else if (teta > 2.0 / 3.0 * pi && teta <= pi)
-        {
           teta_corrected = teta_corrected + 2.0 / 3.0 * pi;
-        }
         else if (teta > pi && teta <= 4.0 / 3.0 * pi)
-        {
           teta_corrected = teta_corrected + pi;
-        }
         else if (teta > 4.0 / 3.0 * pi && teta <= 5.0 / 3.0 * pi)
-        {
           teta_corrected = teta_corrected + 4.0 / 3.0 * pi;
-        }
         else if (teta > 5.0 / 3.0 * pi && teta <= 2.0 * pi)
-        {
           teta_corrected = teta_corrected + 5.0 / 3.0 * pi;
-        }
       }
-      positions.emplace_back(distance * std::cos(teta_corrected),
-                             distance * std::sin(teta_corrected));
+      positions.emplace_back(center(0) + distance * std::cos(teta_corrected),
+                             center(1) + distance * std::sin(teta_corrected));
       teta = teta + dteta;
     } // j
   }   // i
-  //
 }
 
 TriSubChannelMesh::TriSubChannelMesh(const InputParameters & params)
@@ -109,14 +74,6 @@ TriSubChannelMesh::TriSubChannelMesh(const InputParameters & params)
   //  compute the hex mesh variables
   // -------------------------------------------
 
-  // angle in radians
-  Real teta = 0.0;
-  // incremental angle
-  Real dteta = 0.0;
-  // distance between selected fuel rods
-  Real distance = 0.0;
-  // angle in radiance
-  Real teta1 = 0.0;
   // x coordinate for the first position
   Real x0 = 0.0;
   // y coordinate for the first position
@@ -138,8 +95,6 @@ TriSubChannelMesh::TriSubChannelMesh(const InputParameters & params)
   // distance between two points
   Real dist0 = 0.0;
   // integer counter
-  unsigned int k = 0;
-  // integer counter
   unsigned int kgap = 0;
[…]
