### 2021-03-04 9d0be15
Skip check command tests on MacOS, they are too slow (see moosetools#36)
(refs #17191)
