### 2021-03-04 e11d0f9
working on mapping subchannel nodes to duct nodes
