+++ modules/navier_stokes/src/fvkernels/PNSFVMomentumPressureRZ.C
@@ -7,33 +7,36 @@
 //* Licensed under LGPL 2.1, please see LICENSE for details
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 
-#include "NSFVMomentumPressureRZ.h"
+#include "PNSFVMomentumPressureRZ.h"
 
 #include "NS.h"
 
-registerMooseObject("NavierStokesApp", NSFVMomentumPressureRZ);
+registerMooseObject("NavierStokesApp", PNSFVMomentumPressureRZ);
 
 InputParameters
-NSFVMomentumPressureRZ::validParams()
+PNSFVMomentumPressureRZ::validParams()
 {
   InputParameters params = FVElementalKernel::validParams();
-  params.addClassDescription("Adds the $-p/r$ term into the radial component of the Navier-Stokes "
-                             "momentum equation for the problems in the RZ coordinate system.");
+  params.addClassDescription(
+      "Adds the porous $-p/r$ term into the radial component of the Navier-Stokes "
+      "momentum equation for the problems in the RZ coordinate system.");
   return params;
 }
 
-NSFVMomentumPressureRZ::NSFVMomentumPressureRZ(const InputParameters & params)
-  : FVElementalKernel(params), _p(getADMaterialProperty<Real>(NS::pressure))
+PNSFVMomentumPressureRZ::PNSFVMomentumPressureRZ(const InputParameters & params)
+  : FVElementalKernel(params),
+    _p(getADMaterialProperty<Real>(NS::pressure)),
+    _eps(getMaterialProperty<Real>(NS::porosity))
 {
 }
 
 ADReal
-NSFVMomentumPressureRZ::computeQpResidual()
+PNSFVMomentumPressureRZ::computeQpResidual()
 {
   mooseAssert(_subproblem.getCoordSystem(_current_elem->subdomain_id()) == Moose::COORD_RZ,
               "This object should only be active in an RZ coordinate system.");
 
   auto rz_radial_coord = _subproblem.getAxisymmetricRadialCoord();
 
-  return -_p[_qp] / _q_point[_qp](rz_radial_coord);
+  return -_eps[_qp] * _p[_qp] / _q_point[_qp](rz_radial_coord);
 }
