+++ modules/navier_stokes/src/fvbcs/PNSFVMomentumOneOverGammaMatAdvectionOutflowBC.C
@@ -7,13 +7,13 @@
 //* Licensed under LGPL 2.1, please see LICENSE for details
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 
-#include "NSFVPorosityMomentumMatAdvectionOutflowBC.h"
+#include "PNSFVMomentumOneOverGammaMatAdvectionOutflowBC.h"
 #include "NS.h"
 
-registerMooseObject("NavierStokesApp", NSFVPorosityMomentumMatAdvectionOutflowBC);
+registerMooseObject("NavierStokesApp", PNSFVMomentumOneOverGammaMatAdvectionOutflowBC);
 
 InputParameters
-NSFVPorosityMomentumMatAdvectionOutflowBC::validParams()
+PNSFVMomentumOneOverGammaMatAdvectionOutflowBC::validParams()
 {
   InputParameters params = FVMatAdvectionOutflowBC::validParams();
   params.addClassDescription(
@@ -24,22 +24,14 @@ NSFVPorosityMomentumMatAdvectionOutflowBC::validParams()
                              one_over_porosity_interp_method,
                              "The interpolation to use for the one/porosity prefactor. Options are "
                              "'upwind' and 'average', with the default being 'average'.");
-  MooseEnum momentum_component("x=0 y=1 z=2");
-  params.addRequiredParam<MooseEnum>(
-      "momentum_component",
-      momentum_component,
-      "The component of the momentum equation that this kernel applies to.");
   return params;
 }
 
-NSFVPorosityMomentumMatAdvectionOutflowBC::NSFVPorosityMomentumMatAdvectionOutflowBC(
+PNSFVMomentumOneOverGammaMatAdvectionOutflowBC::PNSFVMomentumOneOverGammaMatAdvectionOutflowBC(
     const InputParameters & params)
   : FVMatAdvectionOutflowBC(params),
     _eps_elem(getMaterialProperty<Real>(NS::porosity)),
-    _eps_neighbor(getNeighborMaterialProperty<Real>(NS::porosity)),
-    _p_elem(getADMaterialProperty<Real>(NS::pressure)),
-    _p_neighbor(getNeighborADMaterialProperty<Real>(NS::pressure)),
-    _index(getParam<MooseEnum>("momentum_component"))
+    _eps_neighbor(getNeighborMaterialProperty<Real>(NS::porosity))
 {
   using namespace Moose::FV;
 
@@ -55,7 +47,7 @@ NSFVPorosityMomentumMatAdvectionOutflowBC::NSFVPorosityMomentumMatAdvectionOutfl
 }
 
 ADReal
-NSFVPorosityMomentumMatAdvectionOutflowBC::computeQpResidual()
+PNSFVMomentumOneOverGammaMatAdvectionOutflowBC::computeQpResidual()
 {
   using namespace Moose::FV;
 
@@ -74,13 +66,5 @@ NSFVPorosityMomentumMatAdvectionOutflowBC::computeQpResidual()
               *_face_info,
               true);
 
-  ADReal eps_p_face;
-  interpolate(InterpMethod::Average,
-              eps_p_face,
-              _eps_elem[_qp] * _p_elem[_qp],
-              _eps_neighbor[_qp] * _p_neighbor[_qp],
-              *_face_info,
-              true);
-
-  return residual * one_over_porosity_face + _normal(_index) * eps_p_face;
+  return residual * one_over_porosity_face;
 }
