+++ modules/navier_stokes/src/fvkernels/PNSFVMomentumOneOverGammaUpwindAdvectorAndAdvected.C
@@ -7,15 +7,15 @@
 //* Licensed under LGPL 2.1, please see LICENSE for details
 //* https://www.gnu.org/licenses/lgpl-2.1.html
 
-#include "NSFVPorosityMomentumMatAdvection.h"
+#include "PNSFVMomentumOneOverGammaUpwindAdvectorAndAdvected.h"
 #include "NS.h"
 
-registerMooseObject("NavierStokesApp", NSFVPorosityMomentumMatAdvection);
+registerMooseObject("NavierStokesApp", PNSFVMomentumOneOverGammaUpwindAdvectorAndAdvected);
 
 InputParameters
-NSFVPorosityMomentumMatAdvection::validParams()
+PNSFVMomentumOneOverGammaUpwindAdvectorAndAdvected::validParams()
 {
-  InputParameters params = NSFVPorosityMatAdvection::validParams();
+  InputParameters params = PNSFVUpwindAdvectorAndAdvected::validParams();
   params.addClassDescription("Computes the residual of the momentum advective term with porosity "
                              "using finite volume method.");
   MooseEnum one_over_porosity_interp_method("average upwind", "average");
@@ -24,19 +24,12 @@ NSFVPorosityMomentumMatAdvection::validParams()
                              one_over_porosity_interp_method,
                              "The interpolation to use for the one/porosity prefactor. Options are "
                              "'upwind' and 'average', with the default being 'average'.");
-  MooseEnum momentum_component("x=0 y=1 z=2");
-  params.addRequiredParam<MooseEnum>(
-      "momentum_component",
-      momentum_component,
-      "The component of the momentum equation that this kernel applies to.");
   return params;
 }
 
-NSFVPorosityMomentumMatAdvection::NSFVPorosityMomentumMatAdvection(const InputParameters & params)
-  : NSFVPorosityMatAdvection(params),
-    _p_elem(getADMaterialProperty<Real>(NS::pressure)),
-    _p_neighbor(getNeighborADMaterialProperty<Real>(NS::pressure)),
-    _index(getParam<MooseEnum>("momentum_component"))
+PNSFVMomentumOneOverGammaUpwindAdvectorAndAdvected::
+    PNSFVMomentumOneOverGammaUpwindAdvectorAndAdvected(const InputParameters & params)
+  : PNSFVUpwindAdvectorAndAdvected(params)
 {
   using namespace Moose::FV;
 
@@ -52,7 +45,7 @@ NSFVPorosityMomentumMatAdvection::NSFVPorosityMomentumMatAdvection(const InputPa
 }
 
 ADReal
-NSFVPorosityMomentumMatAdvection::computeQpResidual()
+PNSFVMomentumOneOverGammaUpwindAdvectorAndAdvected::computeQpResidual()
 {
   using namespace Moose::FV;
 
@@ -76,13 +69,5 @@ NSFVPorosityMomentumMatAdvection::computeQpResidual()
   u_interface = _adv_quant_elem[_qp] * elem_coeff + _adv_quant_neighbor[_qp] * neighbor_coeff;
   one_over_porosity_interface = 1 / eps_interface;
 
-  interpolate(InterpMethod::Average,
-              eps_p_interface,
-              _eps_elem[_qp] * _p_elem[_qp],
-              _eps_neighbor[_qp] * _p_neighbor[_qp],
-              *_face_info,
-              true);
-
-  return _normal * _v * u_interface * one_over_porosity_interface +
-         _normal(_index) * eps_p_interface;
+  return _normal * _v * u_interface * one_over_porosity_interface;
 }
