+++ modules/tensor_mechanics/doc/content/source/postprocessors/PolarMomentOfInertia.md
@@ -1,0 +1,17 @@
+# PolarMomentOfInertia
+
+!syntax description /Postprocessors/PolarMomentOfInertia
+
+## Overview
+
+This postprocessor computes the *polar moment of inertia* $J$ (or $I_z$) of a surface
+with respect to a reference point (`origin`) and an axis of rotation (`direction`). Its result
+is needed by the [TwistingMomentBC](TwistingMomentBC.md) boundary condition.
+
+## Example Input File Syntax
+
+!syntax parameters /Postprocessors/PolarMomentOfInertia
+
+!syntax inputs /Postprocessors/PolarMomentOfInertia
+
+!syntax children /Postprocessors/PolarMomentOfInertia
