+++ modules/navier_stokes/include/fvkernels/PINSFVMomentumPressureFlux.h
@@ -1,0 +1,39 @@
+//* This file is part of the MOOSE framework
+//* https://www.mooseframework.org
+//*
+//* All rights reserved, see COPYRIGHT for full restrictions
+//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
+//*
+//* Licensed under LGPL 2.1, please see LICENSE for details
+//* https://www.gnu.org/licenses/lgpl-2.1.html
+
+#pragma once
+
+#include "FVFluxKernel.h"
+
+/**
+ * A flux kernel diffusion of momentum in porous media across cell faces
+ */
+class PINSFVMomentumPressureFlux : public FVFluxKernel
+{
+public:
+  static InputParameters validParams();
+  PINSFVMomentumPressureFlux(const InputParameters & params);
+
+protected:
+  ADReal computeQpResidual() override;
+  bool skipForBoundary(const FaceInfo & fi) const override;
+
+  /// the current element porosity
+  const VariableValue & _eps;
+  /// the neighbor element porosity
+  const VariableValue & _eps_neighbor;
+  /// the porosity
+  const ADVariableValue & _p_elem;
+  /// the neighbor element porosity
+  const ADVariableValue & _p_neighbor;
+  /// the pressure variable
+  const MooseVariableFVReal * const _p_var;
+  /// Which momentum component this kernel applies to
+  const int _index;
+};
