### 2021-03-05 957c56d
Add a porosity discontinuity term which causes a proportional pressure jump
