### 2021-03-05 0590aeb
Add a Generic vector material for testing anisotropic friction kernels
