### 2021-03-05 890e1a1
Add a Generic vector material for testing anisotropic friction kernels
