### 2021-03-05 df5a01c
Add a Generic vector material for testing anisotropic friction kernels
