+++ framework/doc/content/source/materials/GenericConstantVectorMaterial.md
@@ -1,0 +1,36 @@
+# GenericConstantVectorMaterial
+
+!alert! construction title=Undocumented Class
+The GenericConstantVectorMaterial has not been documented. The content listed below should be used as a starting point for
+documenting the class, which includes the typical automatic documentation associated with a
+MooseObject; however, what is contained is ultimately determined by what is necessary to make the
+documentation clear for users.
+
+```markdown
+# GenericConstantVectorMaterial
+
+!syntax description /Materials/GenericConstantVectorMaterial
+
+## Overview
+
+!! Replace these lines with information regarding the GenericConstantVectorMaterial object.
+
+## Example Input File Syntax
+
+!! Describe and include an example of how to use the GenericConstantVectorMaterial object.
+
+!syntax parameters /Materials/GenericConstantVectorMaterial
+
+!syntax inputs /Materials/GenericConstantVectorMaterial
+
+!syntax children /Materials/GenericConstantVectorMaterial
+```
+!alert-end!
+
+!syntax description /Materials/GenericConstantVectorMaterial
+
+!syntax parameters /Materials/GenericConstantVectorMaterial
+
+!syntax inputs /Materials/GenericConstantVectorMaterial
+
+!syntax children /Materials/GenericConstantVectorMaterial
