+++ modules/navier_stokes/include/base/NS.h
@@ -15,137 +15,145 @@
 
 namespace NS
 {
-  // geometric quantities
-  static const std::string pebble_diameter = "pebble_diameter";
-  static const std::string infinite_porosity = "infinite_porosity";
-  static const std::string axis = "axis";
-  static const std::string center = "center";
-  static const std::string wall_porosity = "wall_porosity";
-  static const std::string wall_distance = "wall_distance";
+// geometric quantities
+static const std::string pebble_diameter = "pebble_diameter";
+static const std::string infinite_porosity = "infinite_porosity";
+static const std::string axis = "axis";
+static const std::string center = "center";
+static const std::string wall_porosity = "wall_porosity";
+static const std::string wall_distance = "wall_distance";
 
-  // Names defined in Navier-Stokes
-  static const std::string density = "rho";
-  static const std::string superficial_density = "superficial_rho";
-  static const std::string momentum_x = "rhou";
-  static const std::string momentum_y = "rhov";
-  static const std::string momentum_z = "rhow";
-  static const std::string superficial_momentum_x = "superficial_rhou";
-  static const std::string superficial_momentum_y = "superficial_rhov";
-  static const std::string superficial_momentum_z = "superficial_rhow";
+// Names defined in Navier-Stokes
+static const std::string density = "rho";
+static const std::string superficial_density = "superficial_rho";
+static const std::string momentum_x = "rhou";
+static const std::string momentum_y = "rhov";
+static const std::string momentum_z = "rhow";
+static const std::string superficial_momentum_x = "superficial_rhou";
+static const std::string superficial_momentum_y = "superficial_rhov";
+static const std::string superficial_momentum_z = "superficial_rhow";
 
-  static const std::string velocity = "velocity";
-  static const std::string velocity_x = "vel_x";
-  static const std::string velocity_y = "vel_y";
-  static const std::string velocity_z = "vel_z";
-  static const std::string superficial_velocity_x = "superficial_vel_x";
-  static const std::string superficial_velocity_y = "superficial_vel_y";
-  static const std::string superficial_velocity_z = "superficial_vel_z";
-  static const std::string superficial_velocity = "superficial_velocity";
-  static const std::string pressure = "pressure";
-  static const std::string temperature = "temperature";
+static const std::string velocity = "velocity";
+static const std::string velocity_x = "vel_x";
+static const std::string velocity_y = "vel_y";
+static const std::string velocity_z = "vel_z";
+static const std::string superficial_velocity_x = "superficial_vel_x";
+static const std::string superficial_velocity_y = "superficial_vel_y";
+static const std::string superficial_velocity_z = "superficial_vel_z";
+static const std::string superficial_velocity = "superficial_velocity";
+static const std::string pressure = "pressure";
+static const std::string temperature = "temperature";
 
-  static const std::string specific_internal_energy = "e";
-  static const std::string specific_total_energy = "et";
-  static const std::string internal_energy_density = "rho_e";
-  static const std::string total_energy_density = "rho_et";
+static const std::string specific_internal_energy = "e";
+static const std::string specific_total_energy = "et";
+static const std::string internal_energy_density = "rho_e";
+static const std::string total_energy_density = "rho_et";
 
-  static const std::string specific_enthalpy = "h";
-  static const std::string specific_total_enthalpy = "ht";
-  static const std::string enthalpy_density = "rho_h";
-  static const std::string total_enthalpy_density = "rho_ht";
[…]
