### 2021-03-05 ed39301
Remove code that should stay in Pronghorn
