### 2021-03-05 9611c25
Remove code that should stay in Pronghorn
