+++ modules/navier_stokes/src/materials/PrimitiveVarMaterial.C
@@ -1,126 +1,0 @@
-#include "PrimitiveVarMaterial.h"
-#include "SinglePhaseFluidProperties.h"
-#include "NS.h"
-#include "AuxiliarySystem.h"
-
-namespace nms = NS;
-
-// Full specialization of the validParams function for this object
-registerADMooseObject("NavierStokesApp", PrimitiveVarMaterial);
-
-defineADValidParams(
-    PrimitiveVarMaterial,
-    VarMaterialBase,
-    params.addRequiredCoupledVar(nms::pressure, "pressure");
-    params.addRequiredCoupledVar(nms::T_fluid, "fluid temperature");
-    params.addRequiredCoupledVar(nms::velocity_x, "x-direction velocity");
-    params.addCoupledVar(nms::velocity_y, "y-direction velocity");
-    params.addCoupledVar(nms::velocity_z, "z-direction velocity");
-    params.addClassDescription("Provides access to variables for a primitive variable set "
-      "of pressure, temperature, and interstitial velocity"););
-
-PrimitiveVarMaterial::PrimitiveVarMaterial(const InputParameters & params)
-  : VarMaterialBase(params),
-    _var_T_fluid(adCoupledValue(nms::T_fluid)),
-    _var_pressure(adCoupledValue(nms::pressure)),
-    _var_vel_x(adCoupledValue(nms::velocity_x)),
-    _var_vel_y(isCoupled(nms::velocity_y) ? adCoupledValue(nms::velocity_y) : _ad_zero),
-    _var_vel_z(isCoupled(nms::velocity_z) ? adCoupledValue(nms::velocity_z) : _ad_zero),
-    _var_grad_T_fluid(adCoupledGradient(nms::T_fluid)),
-    _var_grad_pressure(adCoupledGradient(nms::pressure)),
-    _var_grad_vel_x(adCoupledGradient(nms::velocity_x)),
-    _var_grad_vel_y(isCoupled(nms::velocity_y) ? adCoupledGradient(nms::velocity_y) : _ad_grad_zero),
-    _var_grad_vel_z(isCoupled(nms::velocity_z) ? adCoupledGradient(nms::velocity_z) : _ad_grad_zero),
-    _var_T_fluid_dot(_is_transient ? adCoupledDot(nms::T_fluid) : _ad_zero),
-    _var_pressure_dot(_is_transient ? adCoupledDot(nms::pressure) : _ad_zero),
-    _var_vel_x_dot(_is_transient ? adCoupledDot(nms::velocity_x) : _ad_zero),
-    _var_vel_y_dot(isCoupled(nms::velocity_y) && _is_transient ? adCoupledDot(nms::velocity_y)
-                                                          : _ad_zero),
-    _var_vel_z_dot(isCoupled(nms::velocity_z) && _is_transient ? adCoupledDot(nms::velocity_z)
-                                                          : _ad_zero),
-    _var_grad_grad_T_fluid(adCoupledSecond(nms::T_fluid)),
-    _var_grad_grad_vel_x(adCoupledSecond(nms::velocity_x)),
[…]

[…]
