### 2021-03-05 581e9d5
Initialize vars refs #15444
Address comments refs #15444

Another initialization refs #15444
