+++ framework/src/systems/NonlinearSystem.C
@@ -191,8 +191,6 @@ NonlinearSystem::solve()
 
   solver.set_snesmf_reuse_base(_fe_problem.useSNESMFReuseBase());
 
-  //  _console << "\n\n" << std::endl;
-
   if (_time_integrator)
   {
     _time_integrator->solve();
