### 2021-03-05 0fa1b64
Remove TimedPrint refs #15444
Clang format refs #15444
