+++ framework/include/utils/PerfNode.h
@@ -49,13 +49,16 @@ public:
    * Get the currnet start time
    * Only makes sense if this node is running.
    */
-  const std::chrono::time_point<std::chrono::steady_clock> & startTime() { return _start_time; }
+  const std::chrono::time_point<std::chrono::steady_clock> & startTime() const
+  {
+    return _start_time;
+  }
 
   /**
    * Get the current start memory
    * Only makes sense if this node is running.
    */
-  const long unsigned int & startMemory() { return _start_memory; }
+  const long unsigned int & startMemory() const { return _start_memory; }
 
   /**
    * Add some time into this Node by taking the difference with the time passed in
