+++ framework/src/base/MooseInit.C
@@ -13,7 +13,6 @@
 #include "ActionFactory.h"
 #include "Executioner.h"
 #include "MooseRandom.h"
-#include "TimedPrint.h"
 
 // PETSc
 #include "petscsys.h"
