### 2021-03-05 a628955
Add INL records to SQA reports
(refs #17191)
