+++ modules/tensor_mechanics/doc/sqa_reports.yml
@@ -9,19 +9,20 @@ Applications:
         log_stub_files: WARNING
 
 Documents:
+    software_quality_plan: sqa/inl_records.md#software_quality_plan
+    enterprise_architecture_entry: sqa/inl_records.md#enterprise_architecture_entry
+    safety_software_determination: sqa/inl_records.md#safety_software_determination
+    quality_level_determination: sqa/inl_records.md#quality_level_determination
+    log_project_management_plan: WARNING
+    log_configuration_management_plan: WARNING
+    log_asset_management_plan: WARNING
+    log_verification_validation_plan: WARNING
     software_requirements_specification: sqa/tensor_mechanics_srs.md
     software_design_description: sqa/tensor_mechanics_sdd.md
     software_test_plan: sqa/tensor_mechanics_stp.md
     requirements_traceablity_matrix: sqa/tensor_mechanics_rtm.md
-    verification_validation_plan: sqa/tensor_mechanics_vvr.md
-    log_safety_software_determination: WARNING
-    log_quality_level_determination: WARNING
-    log_enterprise_architecture_entry: WARNING
-    log_project_management_plan: WARNING
-    log_software_quality_plan: WARNING
-    log_configuration_management_plan: WARNING
-    log_asset_management_plan: WARNING
-    log_verification_validation_plan: WARNING
+    verification_validation_report: sqa/tensor_mechanics_vvr.md
+    failure_analysis_report: sqa/tensor_mechanics_far.md
     log_user_manual: WARNING
     log_theory_manual: WARNING
 
