+++ modules/functional_expansion_tools/doc/sqa_reports.yml
@@ -6,19 +6,20 @@ Applications:
         content_directory: ${MOOSE_DIR}/modules/functional_expansion_tools/doc/content
 
 Documents:
+    software_quality_plan: sqa/inl_records.md#software_quality_plan
+    enterprise_architecture_entry: sqa/inl_records.md#enterprise_architecture_entry
+    safety_software_determination: sqa/inl_records.md#safety_software_determination
+    quality_level_determination: sqa/inl_records.md#quality_level_determination
+    log_project_management_plan: WARNING
+    log_configuration_management_plan: WARNING
+    log_asset_management_plan: WARNING
+    log_verification_validation_plan: WARNING
     log_software_requirements_specification: WARNING
     log_software_design_description: WARNING
     log_software_test_plan: WARNING
     log_requirements_traceablity_matrix: WARNING
-    log_verification_validation_plan: WARNING
-    log_safety_software_determination: WARNING
-    log_quality_level_determination: WARNING
-    log_enterprise_architecture_entry: WARNING
-    log_project_management_plan: WARNING
-    log_software_quality_plan: WARNING
-    log_configuration_management_plan: WARNING
-    log_asset_management_plan: WARNING
-    log_verification_validation_plan: WARNING
+    log_verification_validation_report: WARNING
+    log_failure_analysis_report: WARNING
     log_user_manual: WARNING
     log_theory_manual: WARNING
 
