+++ modules/subchannel/unit/src/MeshTests.C
@@ -9,7 +9,7 @@ const double tol = 1e-8;
 const double eps = 1e-9;
 
 void
-rodPositionsRef(std::vector<Point> & positions, Real nrings, Real pitch, Point center)
+triRodPositionsRef(std::vector<Point> & positions, Real nrings, Real pitch, Point center)
 {
   positions.clear();
 
@@ -38,10 +38,10 @@ relativeEq(double x, double y)
   if (std::abs(x) < eps && std::abs(y) < eps)
     return true;
 
-  float diff = fabs(x - y);
+  double diff = std::abs(x - y);
   x = std::abs(x);
   y = std::abs(y);
-  float largest = (y > x) ? y : x;
+  double largest = (y > x) ? y : x;
   return diff <= largest * tol;
 }
 
@@ -57,7 +57,7 @@ pointLess(const Point & a, const Point & b)
   return false;
 }
 
-TEST(MeshTests, rodCoordinates)
+TEST(MeshTests, triRodCoordinates)
 {
 
   int nrings = 4;
@@ -65,11 +65,11 @@ TEST(MeshTests, rodCoordinates)
   Point center(0, 0, 0);
 
   std::vector<Point> positions;
-  rodPositionsRef(positions, nrings, pitch, center);
+  triRodPositionsRef(positions, nrings, pitch, center);
   std::sort(positions.begin(), positions.end(), pointLess);
 
   std::vector<Point> positions2;
-  rodPositions(positions2, nrings, pitch, center);
+  triRodPositions(positions2, nrings, pitch, center);
   std::sort(positions2.begin(), positions2.end(), pointLess);
 
   ASSERT_EQ(positions.size(), positions2.size());
@@ -78,7 +78,6 @@ TEST(MeshTests, rodCoordinates)
   for (int i = 0; i < positions.size(); i++)
     if (!positions[i].absolute_fuzzy_equals(positions2[i]))
       msg << "point " << i + 1 << " differs: " << positions[i] << " != " << positions2[i] << "\n";
-  ;
 
   if (msg.str().size() > 0)
     FAIL() << msg.str();
