### 2021-03-05 149d8c7
change one check to make it more strait-forward for #17223
