+++ framework/doc/content/sqa/inl_records.md
@@ -5,31 +5,67 @@ to the [!ac](INL) network. Instructions on how to access these documents are pro
 stored in the same way for each MOOSE-based application, so where appropriate, use the name of the
 actual application of interest when searching.
 
-## PLN-4005, "Software Quality Assurance Plan for MOOSE and MOOSE-Based Applications" id=software_quality_plan
+!alert warning title=[!ac](INL) Network Required
+To access the following documents you must be on the [!ac](INL) network.
+
+## Acronyms
+
+!acronym list
+
+## Software Quality Plan id=software_quality_plan
+
+The [!ac](SQP) use is titled the "Software Quality Assurance Plan for MOOSE and MOOSE-Based
+Applications" (PLN-4005). It includes the [!ac](PMP), [!ac](CMP), and [!ac](VVP). It can be accessed
+from the [!ac](INL) network as follows.
 
 1. From the +Nucleus+ Home page select +A-Z+
 1. From the alphabetical list select +EDMS+
 1. Enter +PLN-4005+ in the "Document Id" field and select +Submit+.
 1. Select the +Doc Viewer down arrow+ and then select the +Record File pdf+.
 
-## [!ac](EA) Record id=enterprise_architecture_entry
+## Enterprise Architecture Record id=enterprise_architecture_entry
+
+The [!ac](EA) record can be accessed from the [!ac](INL) network as follows.
 
 1. From the Nucleus Home page select +A-Z+
-1. From the alphabetical list select +CTM –Capabilities & Technology Management+
+1. From the alphabetical list select +CTM---Capabilities & Technology Management+
 1. Begin typing the name of the application in the search box in the upper right corner of the screen
 1. When the search is complete, click on the ellipses (three dots) to the left of the application
    name and then select +View Details+
 
-## [!ac](SSD) id=safety_software+determination
+## Safety Software Determination id=safety_software+determination
+
+The [!ac](SSD) can be accessed from the [!ac](INL) network as follows.
 
 1. From the [!ac](EA) +Details+ screen for the application (refer to [#enterprise_architecture_entry]),
    select the +General+ tab
 1. Scroll to the bottom of the screen and in the +Related Artifacts+ section select +Safety
    Software Determination+
 
-## [!ac](QLD) id=quality_level_determination
+## Quality Level Determination id=quality_level_determination
 
-1. From the Enterprise Architecture +Details+ screen for the application (refer to [#enterprise_architecture_entry]),
-   select the +General+ tab)
+The [!ac](QLD) can be accessed from the [!ac](INL) network as follows.
+
+1. From the [!ac](EA) +Details+ screen for the application (refer to [#enterprise_architecture_entry]),
+   select the +General+ tab
 1. Scroll to the bottom of the screen and in the +Related Artifacts+ section select +Quality
    Level Determination+
+
+## Project Management Plan id=project_management_plan
+
+The [!ac](PMP) is included in the [!ac](SQP), see [#software_quality_plan].
+
+## Configuration Management Plan id=configuration_management_plan
+
+The [!ac](CMP) is included in the [!ac](SQP), see [#software_quality_plan].
+
+## Verification and Validation Plan id=verification_and_validation_plan
+
+The [!ac](VVP) is included in the [!ac](SQP), see [#software_quality_plan].
+
+## Asset Maintenance Plan id=asset_maintenance_plan
+
+In accordance with [!ac](INL) [!ac](LWP) "Managing Information Technology Assets" (LWP-13620), an
+[!ac](IT) asset maintenance plan, or equivalent, must be created by the end user to specify how the
+software application will be managed after it is deployed for use, see the [!ac](INL) template,
+"IT Asset Maintenance Plan Template" (TEM-215) for further information.
