+++ framework/doc/sqa_reports.yml
@@ -19,10 +19,10 @@ Documents:
     enterprise_architecture_entry: sqa/inl_records.md#enterprise_architecture_entry
     safety_software_determination: sqa/inl_records.md#safety_software_determination
     quality_level_determination: sqa/inl_records.md#quality_level_determination
-    log_project_management_plan: WARNING
-    log_configuration_management_plan: WARNING
-    log_asset_management_plan: WARNING
-    log_verification_validation_plan: WARNING
+    project_management_plan: sqa/inl_records.md#project_management_plan
+    configuration_management_plan: sqa/inl_records.md#configuration_management_plan
+    asset_management_plan: sqa/inl_records.md#asset_management_plan
+    verification_validation_plan: sqa/inl_records.md#verification_and_validation_plan
     software_requirements_specification: sqa/framework_srs.md
     software_design_description: sqa/framework_sdd.md
     software_test_plan: sqa/framework_stp.md
