### 2021-03-05 cbc27af
Add member function to determine target elems.
