### 2021-03-05 eec79ee
FullSolveMultiApp need to be replaced with SamplerFullSolveMultiApp for working with commandline control #17223
