+++ modules/navier_stokes/src/materials/ConservedVarValuesMaterial.C
@@ -54,7 +54,8 @@ ConservedVarValuesMaterial::ConservedVarValuesMaterial(const InputParameters & p
     _specific_internal_energy(declareADProperty<Real>(NS::specific_internal_energy)),
     _pressure(declareADProperty<Real>(NS::pressure)),
     _specific_total_enthalpy(declareADProperty<Real>(NS::specific_total_enthalpy)),
-    _rho_ht(declareADProperty<Real>(NS::total_enthalpy_density))
+    _rho_ht(declareADProperty<Real>(NS::total_enthalpy_density)),
+    _T_fluid(declareADProperty<Real>(NS::T_fluid))
 {
 }
 
@@ -97,4 +98,5 @@ ConservedVarValuesMaterial::computeQpProperties()
 
   _specific_total_enthalpy[_qp] = (_total_energy_density[_qp] + _pressure[_qp]) / _rho[_qp];
   _rho_ht[_qp] = _specific_total_enthalpy[_qp] * _rho[_qp];
+  _T_fluid[_qp] = _fluid.T_from_v_e(_v[_qp], _specific_internal_energy[_qp]);
 }
