### 2021-03-06 738daf7
Compute and use momentum in friction kernel
