### 2021-03-06 8c99fc6
Compute and use momentum in friction kernel
