+++ modules/navier_stokes/src/fvbcs/NSFVFluidEnergySpecifiedTemperatureBC.C
@@ -11,11 +11,11 @@ NSFVFluidEnergySpecifiedTemperatureBC::validParams()
   auto params = FVFluxBC::validParams();
   params.addRequiredParam<FunctionName>(NS::temperature,
                                         "Inlet temperature specified as a function");
-  params.addRequiredParam<FunctionName>(NS::superficial_momentum_x,
+  params.addRequiredParam<FunctionName>(NS::momentum_x,
                                         "The x component of the inlet superficial momentum");
-  params.addParam<FunctionName>(NS::superficial_momentum_y,
+  params.addParam<FunctionName>(NS::momentum_y,
                                 "The y component of the inlet superficial momentum");
-  params.addParam<FunctionName>(NS::superficial_momentum_z,
+  params.addParam<FunctionName>(NS::momentum_z,
                                 "The z component of the inlet superficial momentum");
   params.addRequiredParam<UserObjectName>(NS::fluid, "fluid userobject");
   return params;
@@ -25,23 +25,17 @@ NSFVFluidEnergySpecifiedTemperatureBC::NSFVFluidEnergySpecifiedTemperatureBC(
     const InputParameters & parameters)
   : FVFluxBC(parameters),
     _temperature(getFunction(NS::temperature)),
-    _superficial_rhou(getFunction(NS::superficial_momentum_x)),
-    _superficial_rhov(isParamValid(NS::superficial_momentum_y)
-                          ? &getFunction(NS::superficial_momentum_y)
-                          : nullptr),
-    _superficial_rhow(isParamValid(NS::superficial_momentum_z)
-                          ? &getFunction(NS::superficial_momentum_z)
-                          : nullptr),
+    _rhou(getFunction(NS::momentum_x)),
+    _rhov(isParamValid(NS::momentum_y) ? &getFunction(NS::momentum_y) : nullptr),
+    _rhow(isParamValid(NS::momentum_z) ? &getFunction(NS::momentum_z) : nullptr),
     _rho_elem(getADMaterialProperty<Real>(NS::density)),
     _rho_neighbor(getNeighborADMaterialProperty<Real>(NS::density)),
-    _porosity_elem(getMaterialProperty<Real>(NS::porosity)),
-    _porosity_neighbor(getNeighborMaterialProperty<Real>(NS::porosity)),
     _fluid(UserObjectInterface::getUserObject<SinglePhaseFluidProperties>(NS::fluid))
 {
-  if (_mesh.dimension() > 1 && !_superficial_rhov)
+  if (_mesh.dimension() > 1 && !_rhov)
     mooseError("If the mesh dimension is greater than 1, a function for the y superficial momentum "
                "must be provided");
-  if (_mesh.dimension() > 2 && !_superficial_rhow)
+  if (_mesh.dimension() > 2 && !_rhow)
     mooseError("If the mesh dimension is greater than 2, a function for the z superficial momentum "
                "must be provided");
 }
@@ -49,11 +43,6 @@ NSFVFluidEnergySpecifiedTemperatureBC::NSFVFluidEnergySpecifiedTemperatureBC(
 ADReal
 NSFVFluidEnergySpecifiedTemperatureBC::computeQpResidual()
 {
-  mooseAssert(_porosity_elem[_qp] == _porosity_neighbor[_qp],
-              "It's not good if a ghost cell porosity has a different porosity than the boundary "
-              "cell porosity");
-  const auto & porosity = _porosity_elem[_qp];
-
   ADReal rho;
   using namespace Moose::FV;
[…]
