### 2021-03-06 46d752f
Add 1D porosity jump test, clean up 2D test Add porosity jump treatment to advection kernel Add continuous porosity gradient term from advection kernel Add effective viscosity option to momentum diffusion Add momentum matprop option to friction kernel, switch to momentum for consistency with Pronghorn Fix formulation of porosity gradient term
