+++ modules/navier_stokes/src/fvkernels/PINSFVMomentumAdvection.C
@@ -19,14 +19,17 @@ PINSFVMomentumAdvection::validParams()
   params.addClassDescription("Object for advecting superficial momentum, e.g. rho*u_d, "
                              "in the porous media momentum equation");
   params.addRequiredCoupledVar("porosity", "Porosity auxiliary variable");
+  params.addParam<bool>("smooth_porosity", false, "Whether the porosity has no discontinuities");
 
   return params;
 }
 
 PINSFVMomentumAdvection::PINSFVMomentumAdvection(const InputParameters & params)
   : INSFVMomentumAdvection(params),
+  _eps_var(dynamic_cast<const MooseVariableFVReal *>(getFieldVar("porosity", 0))),
   _eps(coupledValue("porosity")),
-  _eps_neighbor(coupledNeighborValue("porosity"))
+  _eps_neighbor(coupledNeighborValue("porosity")),
+  _smooth_porosity(getParam<bool>("smooth_porosity"))
 {
   if (!dynamic_cast<const PINSFVVelocityVariable *>(_u_var))
     mooseError("PINSFVMomentumAdvection may only be used with a superficial advective velocity, "
@@ -122,11 +125,6 @@ PINSFVMomentumAdvection::coeffCalculator(const Elem & elem, const ADReal & mu) c
           "Let's make sure our normal is what we think it is");
 #endif
 
-    // if (elem_has_info && std::abs(_eps[_qp] - MetaPhysicL::raw_value(dynamic_cast<const MooseVariableFVReal *>( getFieldVar("porosity", 0))->getElemValue(&fi->elem()))) > 1e-8)
-    //   std::cout << _eps[_qp] << " " << dynamic_cast<const MooseVariableFVReal *>( getFieldVar("porosity", 0))->getElemValue(&fi->elem()) << std::endl;
-    // else if (!elem_has_info && std::abs(_eps[_qp] - MetaPhysicL::raw_value(dynamic_cast<const MooseVariableFVReal *>( getFieldVar("porosity", 0))->getElemValue(&fi->neighbor()))) > 1e-8)
-    //   std::cout << _eps[_qp] << " " << dynamic_cast<const MooseVariableFVReal *>( getFieldVar("porosity", 0))->getElemValue(&fi->neighbor()) << std::endl;
-
     // Unless specified otherwise, "elem" here refers to the element we're computing the
     // Rhie-Chow coefficient for. "neighbor" is the element across the current FaceInfo (fi)
     // face from the Rhie-Chow element
@@ -134,8 +132,7 @@ PINSFVMomentumAdvection::coeffCalculator(const Elem & elem, const ADReal & mu) c
     if (onBoundary(*fi))
     {
       // Compute the face porosity
-      Real eps_face = MetaPhysicL::raw_value(dynamic_cast<const MooseVariableFVReal *>(
-          getFieldVar("porosity", 0))->getBoundaryFaceValue(*fi));
+      Real eps_face = MetaPhysicL::raw_value(_eps_var->getBoundaryFaceValue(*fi));
 
       // In my mind there should only be about one bc_id per FaceInfo
       mooseAssert(fi->boundaryIDs().size() == 1,
@@ -212,11 +209,9 @@ PINSFVMomentumAdvection::coeffCalculator(const Elem & elem, const ADReal & mu) c
     // Else we are on an internal face
 
     // Compute the face porosity
-    Real eps_elem = MetaPhysicL::raw_value(elem_has_info ? dynamic_cast<const MooseVariableFVReal *>(
-        getFieldVar("porosity", 0))->getElemValue(&fi->elem()) :  dynamic_cast<const MooseVariableFVReal *>(
-            getFieldVar("porosity", 0))->getElemValue(neighbor));
[…]
