+++ modules/navier_stokes/src/fvkernels/PINSFVMomentumDiffusion.C
@@ -25,31 +25,39 @@ PINSFVMomentumDiffusion::validParams()
       "momentum_component",
       momentum_component,
       "The component of the momentum equation that this kernel applies to.");
-  params.addParam<bool>("smooth_porosity_gradient", false, "Whether to compute the porosity gradient diffusive term");
+  params.addParam<bool>("smooth_porosity", false, "Whether to compute the porosity gradient diffusive term");
   params.set<unsigned short>("ghost_layers") = 2;
   return params;
 }
 
 PINSFVMomentumDiffusion::PINSFVMomentumDiffusion(const InputParameters & params)
   : FVFluxKernel(params),
-  _mu_elem(getADMaterialProperty<Real>("mu")),
-  _mu_neighbor(getNeighborADMaterialProperty<Real>("mu")),
+  _mu_elem(isParamValid("mu") ? &getADMaterialProperty<Real>("mu") : nullptr),
+  _mu_neighbor(isParamValid("mu") ? &getNeighborADMaterialProperty<Real>("mu") : nullptr),
+  _mu_eff_elem(isParamValid("mu_eff") ? &getADMaterialProperty<Real>("mu_eff") : nullptr),
+  _mu_eff_neighbor(isParamValid("mu_eff") ? &getNeighborADMaterialProperty<Real>("mu_eff") : nullptr),
+  _effective_viscosity(isParamValid("mu_eff")),
   _eps(coupledValue("porosity")),
   _eps_neighbor(coupledNeighborValue("porosity")),
   _index(getParam<MooseEnum>("momentum_component")),
   _vel_elem(getADMaterialProperty<RealVectorValue>(NS::velocity)),
   _vel_neighbor(getNeighborADMaterialProperty<RealVectorValue>(NS::velocity)),
   _eps_var(dynamic_cast<const MooseVariableFVReal *>(getFieldVar("porosity", 0))),
-  _smooth_porosity_gradient(getParam<bool>("smooth_porosity_gradient"))
+  _smooth_porosity(getParam<bool>("smooth_porosity"))
 {
 #ifndef MOOSE_GLOBAL_AD_INDEXING
   mooseError("PINSFV is not supported by local AD indexing. In order to use PINSFV, please run "
              "the configure script in the root MOOSE directory with the configure option "
              "'--with-ad-indexing-type=global'");
 #endif
-  if (_smooth_porosity_gradient && !params.isParamSetByUser("momentum_component"))
+  if (_smooth_porosity && !params.isParamSetByUser("momentum_component"))
     mooseError("The momentum component parameter is required for modeling the porosity "
                "gradient contribution in the momentum diffusion term.");
+  if (_smooth_porosity && isParamValid("mu_eff"))
+    paramError("mu_eff", "The porosity gradient term need not be included when using an "
+               "effective viscosity.");
+  if (isParamValid("mu") == isParamValid("mu_eff"))
+    mooseError("Either the viscosity or the effective viscosity must be specified.");
 }
 
 ADReal
@@ -65,12 +73,20 @@ PINSFVMomentumDiffusion::computeQpResidual()
   // Compute the diffusion driven by the velocity gradient
   // Interpolate viscosity divided by porosity on the face
   ADReal mu_eps_face;
-  interpolate(Moose::FV::InterpMethod::Average,
-              mu_eps_face,
-              _mu_elem[_qp] / _eps[_qp],
-              _mu_neighbor[_qp] / _eps_neighbor[_qp],
-              *_face_info,
-              true);
+  if (!_effective_viscosity)
+    interpolate(Moose::FV::InterpMethod::Average,
[…]
