### 2021-03-06 c8dc662
Convert to static validParams (Closes #34)
