+++ modules/electromagnetics/src/kernels/CoeffADVectorTimeDerivative.C
@@ -3,16 +3,18 @@
 
 registerADMooseObject("ElkApp", CoeffADVectorTimeDerivative);
 
-defineADValidParams(
-    CoeffADVectorTimeDerivative,
-    ADVectorTimeDerivative,
-    params.addClassDescription("The time derivative operator with the weak form of $(\\psi_i, "
-                               "a(\\vec{r}, t) \\frac{\\partial u_h}{\\partial t})$, where $a$ is "
-                               "a coefficient function.");
-    params.addRequiredParam<FunctionName>("coefficient", "Coefficient function."););
+InputParameters
+CoeffADVectorTimeDerivative::validParams()
+{
+  InputParameters params = ADVectorTimeDerivative::validParams();
+  params.addClassDescription("The time derivative operator with the weak form of $(\\psi_i, "
+                             "a(\\vec{r}, t) \\frac{\\partial u_h}{\\partial t})$, where $a$ is "
+                             "a coefficient function.");
+  params.addRequiredParam<FunctionName>("coefficient", "Coefficient function.");
+  return params;
+}
 
-CoeffADVectorTimeDerivative::CoeffADVectorTimeDerivative(
-    const InputParameters & parameters)
+CoeffADVectorTimeDerivative::CoeffADVectorTimeDerivative(const InputParameters & parameters)
   : ADVectorTimeDerivative(parameters), _coefficient(getFunction("coefficient"))
 {
 }
@@ -20,8 +22,5 @@ CoeffADVectorTimeDerivative::CoeffADVectorTimeDerivative(
 ADRealVectorValue
 CoeffADVectorTimeDerivative::precomputeQpResidual()
 {
-  return _coefficient.value(_t, _q_point[_qp]) *
-         ADVectorTimeDerivative::precomputeQpResidual();
+  return _coefficient.value(_t, _q_point[_qp]) * ADVectorTimeDerivative::precomputeQpResidual();
 }
-
-
