+++ modules/electromagnetics/test/include/functions/MMSTestFunc.h
@@ -3,17 +3,13 @@
 #include "Function.h"
 #include "FunctionInterface.h"
 
-class MMSTestFunc;
-
-template <>
-InputParameters validParams<MMSTestFunc>();
-
 /**
  *  Function of RHS for manufactured solution in spatial_constant_helmholtz test
  */
 class MMSTestFunc : public Function, public FunctionInterface
 {
 public:
+  static InputParameters validParams();
   MMSTestFunc(const InputParameters & parameters);
 
   virtual Real value(Real t, const Point & p) const override;
