+++ modules/electromagnetics/src/interfacekernels/ParallelElectricFieldInterface.C
@@ -2,11 +2,10 @@
 
 registerMooseObject("ElkApp", ParallelElectricFieldInterface);
 
-template <>
 InputParameters
-validParams<ParallelElectricFieldInterface>()
+ParallelElectricFieldInterface::validParams()
 {
-  InputParameters params = validParams<VectorInterfaceKernel>();
+  InputParameters params = VectorInterfaceKernel::validParams();
   params.addClassDescription("VectorInterfaceKernel that implements the condtion $\\vec{E}_{1}^{\\parallel} - \\vec{E}_{2}^{\\parallel} = 0$");
   return params;
 }
