+++ modules/electromagnetics/include/bcs/CoupledGradBC.h
@@ -10,8 +10,8 @@ public:
   CoupledGradBC(const InputParameters & parameters);
 
 protected:
-  virtual Real computeQpResidual();
-  virtual Real computeQpJacobian();
+  virtual Real computeQpResidual() override;
+  virtual Real computeQpJacobian() override;
 
 private:
   Real _sign;
