### 2021-03-06 f7e0ee4
Update documentation Express friction in terms of superficial velocity Remove mass time derivative kernel Remove option for effective viscosity in momentum diffusion kernel
