+++ modules/navier_stokes/include/fvkernels/PINSFVMassTimeDerivative.h
@@ -1,27 +1,0 @@
-//* This file is part of the MOOSE framework
-//* https://www.mooseframework.org
-//*
-//* All rights reserved, see COPYRIGHT for full restrictions
-//* https://github.com/idaholab/moose/blob/master/COPYRIGHT
-//*
-//* Licensed under LGPL 2.1, please see LICENSE for details
-//* https://www.gnu.org/licenses/lgpl-2.1.html
-
-#pragma once
-
-#include "FVTimeKernel.h"
-
-class PINSFVMassTimeDerivative : public FVTimeKernel
-{
-public:
-  static InputParameters validParams();
-  PINSFVMassTimeDerivative(const InputParameters & params);
-
-protected:
-  ADReal computeQpResidual() override;
-
-  /// the density
-  const Real & _rho;
-  /// the time derivative of the porosity
-  const VariableValue & _eps_dot;
-};
