+++ modules/navier_stokes/doc/content/source/fvkernels/PINSFVMomentumPressure.md
@@ -1,20 +1,11 @@
 # PINSFVMomentumPressure
 
-!alert construction title=Undocumented Class
-The PINSFVMomentumPressure has not been documented. The content listed below should be used as a starting point for
-documenting the class, which includes the typical automatic documentation associated with a
-MooseObject; however, what is contained is ultimately determined by what is necessary to make the
-documentation clear for users.
+This object adds the $\epsilon \nabla p$ term of the
+incompressible Navier Stokes momentum equation as a volumetric term.
+This formulation causes oscillation near porosity discontinuities and the
+[PINSFVMomentumPressureFlux.md] kernel should be preferred then.
 
-!syntax description /FVKernels/PINSFVMomentumPressure
-
-## Overview
-
-!! Replace these lines with information regarding the PINSFVMomentumPressure object.
-
-## Example Input File Syntax
-
-!! Describe and include an example of how to use the PINSFVMomentumPressure object.
+The pressure is a Lagrange Multiplier that ensures the incompressibility constraint.
 
 !syntax parameters /FVKernels/PINSFVMomentumPressure
 

[…]
