+++ modules/navier_stokes/doc/content/source/fvkernels/PINSFVMomentumBoussinesq.md
@@ -1,20 +1,11 @@
 # PINSFVMomentumBoussinesq
 
-!alert construction title=Undocumented Class
-The PINSFVMomentumBoussinesq has not been documented. The content listed below should be used as a starting point for
-documenting the class, which includes the typical automatic documentation associated with a
-MooseObject; however, what is contained is ultimately determined by what is necessary to make the
-documentation clear for users.
-
-!syntax description /FVKernels/PINSFVMomentumBoussinesq
-
-## Overview
-
-!! Replace these lines with information regarding the PINSFVMomentumBoussinesq object.
-
-## Example Input File Syntax
-
-!! Describe and include an example of how to use the PINSFVMomentumBoussinesq object.
+This object adds a $\epsilon\rho\alpha\vec{g}(T - T_{ref})$ term to the incompressible porous media
+Navier Stokes (PINS) equations where $\epsilon$ is the porosity, $\rho$ is the density, $\alpha$ is the thermal
+expansion coefficient, $\vec{g}$ is the gravity vector, $T$ is the temperature,
+and $T_{ref}$ is a reference temperature. The term above introduces the
+Boussinesq approximation into the PINS equations, which allows for modeling
+natural convection and the effect of buoyancy.
 
 !syntax parameters /FVKernels/PINSFVMomentumBoussinesq
 
