+++ modules/navier_stokes/doc/content/source/fvkernels/INSFVGravityForce.md
@@ -1,11 +1,0 @@
-# INSFVGravityForce
-
-This object adds a simple gravitational force term to the incompressible Navier
-Stokes equations of the form $-\rho\vec{g}$ where $\rho$ is the density and
-$\vec{g}$ is the gravity vector.
-
-!syntax parameters /FVKernels/INSFVGravityForce
-
-!syntax inputs /FVKernels/INSFVGravityForce
-
-!syntax children /FVKernels/INSFVGravityForce
