+++ modules/navier_stokes/doc/content/source/fvkernels/INSFVMassAdvection.md
@@ -1,7 +1,7 @@
 # INSFVMassAdvection
 
 This object computes the residual and Jacobian contribution of the
-incompressible version of the mass continuity equation, e.g. $\nabla\cdot\vec
+incompressible version of the mass continuity equation, e.g. $\nabla\cdot \rho \vec
 u$. We apply the divergence theorem and compute the advective flux of mass
 across cell/element faces.
 
