+++ modules/navier_stokes/doc/content/source/fvkernels/PINSFVMomentumGravity.md
@@ -1,20 +1,8 @@
 # PINSFVMomentumGravity
 
-!alert construction title=Undocumented Class
-The PINSFVMomentumGravity has not been documented. The content listed below should be used as a starting point for
-documenting the class, which includes the typical automatic documentation associated with a
-MooseObject; however, what is contained is ultimately determined by what is necessary to make the
-documentation clear for users.
-
-!syntax description /FVKernels/PINSFVMomentumGravity
-
-## Overview
-
-!! Replace these lines with information regarding the PINSFVMomentumGravity object.
-
-## Example Input File Syntax
-
-!! Describe and include an example of how to use the PINSFVMomentumGravity object.
+This object adds a simple gravitational force term to the incompressible porous media Navier
+Stokes equations of the form $-\epsilon \rho \vec{g}$ where $\epsilon$ is the porosity, $\rho$ the density and
+$\vec{g}$ the gravity vector.
 
 !syntax parameters /FVKernels/PINSFVMomentumGravity
 
