### 2021-03-07 6c99a1f
Clang formatting
