### 2021-03-07 3b43903
Clang formatting Fix eps bug in friction
