+++ framework/src/variables/MooseVariableFV.C
@@ -976,7 +976,8 @@ MooseVariableFV<OutputType>::adGradSln(const FaceInfo & fi) const
   const ADReal & neighbor_value = elem_is_elem_one ? elem_two_value : elem_one_value;
 
   // perform the correction. Note that direction is important here because we have a minus sign.
-  // Neighbor has to be neighbor, and elem has to be elem. Hence all the elem_is_elem_one logic above
+  // Neighbor has to be neighbor, and elem has to be elem. Hence all the elem_is_elem_one logic
+  // above
   face_grad += ((neighbor_value - elem_value) / fi.dCFMag() - face_grad * fi.eCF()) * fi.eCF();
 
   return face_grad;
