+++ modules/navier_stokes/include/fvkernels/PINSFVMomentumFriction.h
@@ -39,7 +39,7 @@ protected:
   const bool _use_Darcy_friction_model;
   const bool _use_Forchheimer_friction_model;
   /// Porosity to compute the intersitial velocity from the superficial velocity
-  const VariableValue _eps;
+  const VariableValue & _eps;
   /// Momentum as a material property
   const ADMaterialProperty<Real> * const _momentum;
   /// Constant density, use only with incompressible flow

[…]
