### 2021-03-07 e990aef
Clang formatting Fix eps bug in friction
