+++ modules/navier_stokes/src/fvkernels/PINSFVMomentumPressure.C
@@ -15,8 +15,8 @@ InputParameters
 PINSFVMomentumPressure::validParams()
 {
   InputParameters params = INSFVMomentumPressure::validParams();
-  params.addClassDescription(
-    "Introduces the coupled pressure term $eps \nabla P$ into the Navier-Stokes porous media momentum equation.");
+  params.addClassDescription("Introduces the coupled pressure term $eps \nabla P$ into the "
+                             "Navier-Stokes porous media momentum equation.");
   params.addRequiredCoupledVar("porosity", "Porosity auxiliary variable");
 
   return params;
