+++ modules/navier_stokes/src/fvkernels/PINSFVMomentumAdvection.C
@@ -26,14 +26,14 @@ PINSFVMomentumAdvection::validParams()
 
 PINSFVMomentumAdvection::PINSFVMomentumAdvection(const InputParameters & params)
   : INSFVMomentumAdvection(params),
-  _eps_var(dynamic_cast<const MooseVariableFVReal *>(getFieldVar("porosity", 0))),
-  _eps(coupledValue("porosity")),
-  _eps_neighbor(coupledNeighborValue("porosity")),
-  _smooth_porosity(getParam<bool>("smooth_porosity"))
+    _eps_var(dynamic_cast<const MooseVariableFVReal *>(getFieldVar("porosity", 0))),
+    _eps(coupledValue("porosity")),
+    _eps_neighbor(coupledNeighborValue("porosity")),
+    _smooth_porosity(getParam<bool>("smooth_porosity"))
 {
   if (!dynamic_cast<const PINSFVVelocityVariable *>(_u_var))
     mooseError("PINSFVMomentumAdvection may only be used with a superficial advective velocity, "
-        "of variable type PINSFVVelocityVariable.");
+               "of variable type PINSFVVelocityVariable.");
 }
 
 const VectorValue<ADReal> &
@@ -172,14 +172,17 @@ PINSFVMomentumAdvection::coeffCalculator(const Elem & elem, const ADReal & mu) c
               Moose::FV::interpCoeffs(_advected_interp_method, *fi, elem_has_info, face_velocity);
           ADReal temp_coeff = _rho * face_velocity * surface_vector * advection_coeffs.first;
 
-          if (_fully_developed_flow_boundaries.find(bc_id) == _fully_developed_flow_boundaries.end())
-            // We are not on a fully developed flow boundary, so we have a viscous term contribution.
-            // This term is slightly modified relative to the internal face term. Instead of the
-            // distance between elem and neighbor centroid, we just have the distance between the elem
-            // and face centroid. Specifically, the term below is the result of Moukalled 8.80, 8.82,
-            // and the orthogonal correction approach equation for E_f, equation 8.89. So relative to
-            // the internal face viscous term, we have substituted eqn. 8.82 for 8.78
-            temp_coeff += mu / eps_face * surface_vector.norm() / (fi->faceCentroid() - rc_centroid).norm();
+          if (_fully_developed_flow_boundaries.find(bc_id) ==
+              _fully_developed_flow_boundaries.end())
+            // We are not on a fully developed flow boundary, so we have a viscous term
+            // contribution. This term is slightly modified relative to the internal face term.
+            // Instead of the distance between elem and neighbor centroid, we just have the distance
+            // between the elem and face centroid. Specifically, the term below is the result of
+            // Moukalled 8.80, 8.82, and the orthogonal correction approach equation for E_f,
+            // equation 8.89. So relative to the internal face viscous term, we have substituted
+            // eqn. 8.82 for 8.78
+            temp_coeff +=
+                mu / eps_face * surface_vector.norm() / (fi->faceCentroid() - rc_centroid).norm();
 
           // For flow boundaries, the coefficient addition is the same for every velocity component
           for (const auto i : make_range(_dim))
@@ -193,7 +196,8 @@ PINSFVMomentumAdvection::coeffCalculator(const Elem & elem, const ADReal & mu) c
           // Moukalled eqns. 15.154 - 15.156
           for (const auto i : make_range(_dim))
             coeff(i) += 2. * mu / eps_face * surface_vector.norm() /
[…]
