### 2021-03-07 24d2061
Clang formatting Fix eps bug in friction
