+++ framework/doc/content/sqa/inl_records.md
@@ -1,9 +1,9 @@
-# INL Software Quality Records
+# [!ac](INL) Software Quality Records
 
-Several software quality-related documents and records listed below are available by those with access
-to the [!ac](INL) network. Instructions on how to access these documents are provided below. These are
-stored in the same way for each MOOSE-based application, so where appropriate, use the name of the
-actual application of interest when searching.
+Several software quality-related documents and records listed below are available by those with
+access to the [!ac](INL) network. Instructions on how to access these documents are
+provided below. These are stored in the same way for each MOOSE-based application, so where
+appropriate, use the name of the actual application of interest when searching.
 
 !alert warning title=[!ac](INL) Network Required
 To access the following documents you must be on the [!ac](INL) network.
@@ -12,18 +12,20 @@ To access the following documents you must be on the [!ac](INL) network.
 
 !acronym list
 
-## Software Quality Plan id=software_quality_plan
+## [!ac](SQAP) id=software_quality_assurance_plan
 
-The [!ac](SQP) use is titled the "Software Quality Assurance Plan for MOOSE and MOOSE-Based
-Applications" (PLN-4005). It includes the [!ac](PMP), [!ac](CMP), and [!ac](VVP). It can be accessed
-from the [!ac](INL) network as follows.
+The [!ac](SQAP) used is titled the "Software Quality Assurance Plan for MOOSE and
+MOOSE-Based Applications" (PLN-4005). This plan lays out the specifics of how MOOSE and MOOSE-based
+applications ensure that life-cycle activities are planned and performed in a traceable and orderly
+manner.  It includes the requirements for [!ac](CMP), [!ac](VVP), and [!ac](SQP). PLN-4005 can be
+accessed from the [!ac](INL) network as follows.
 
 1. From the +Nucleus+ Home page select +A-Z+
 1. From the alphabetical list select +EDMS+
 1. Enter +PLN-4005+ in the "Document Id" field and select +Submit+.
 1. Select the +Doc Viewer down arrow+ and then select the +Record File pdf+.
 
-## Enterprise Architecture Record id=enterprise_architecture_entry
+## [!ac](EA) Record id=enterprise_architecture_entry
 
 The [!ac](EA) record can be accessed from the [!ac](INL) network as follows.
 
@@ -33,7 +35,7 @@ The [!ac](EA) record can be accessed from the [!ac](INL) network as follows.
 1. When the search is complete, click on the ellipses (three dots) to the left of the application
    name and then select +View Details+
 
-## Safety Software Determination id=safety_software+determination
+## [!ac](SSD) id=safety_software+determination
 
 The [!ac](SSD) can be accessed from the [!ac](INL) network as follows.
 
@@ -42,7 +44,7 @@ The [!ac](SSD) can be accessed from the [!ac](INL) network as follows.
 1. Scroll to the bottom of the screen and in the +Related Artifacts+ section select +Safety
    Software Determination+
 
-## Quality Level Determination id=quality_level_determination
+## [!ac](QLD) id=quality_level_determination
 
 The [!ac](QLD) can be accessed from the [!ac](INL) network as follows.
 
@@ -51,19 +53,21 @@ The [!ac](QLD) can be accessed from the [!ac](INL) network as follows.
 1. Scroll to the bottom of the screen and in the +Related Artifacts+ section select +Quality
    Level Determination+
 
-## Project Management Plan id=project_management_plan
+## Configuration Management Plan (CMP) id=configuration_management_plan
 
-The [!ac](PMP) is included in the [!ac](SQP), see [#software_quality_plan].
[…]
