### 2021-03-08 fa3285f
Allow user specification of drag quantity in NSFVMomentumFriction
