### 2021-03-08 3f02c16
Minor change to lang. in inl records page
(refs #17191)
