### 2021-03-08 21154f1
Add min_gap parameter to GapConductanceConstraint
