+++ modules/heat_conduction/src/constraints/GapConductanceConstraint.C
@@ -20,11 +20,15 @@ GapConductanceConstraint::validParams()
       "implementation of the thermal contact problem. For more information, see the "
       "detailed description here: http://tinyurl.com/gmmhbe9");
   params.addRequiredParam<Real>("k", "Gap conductance");
+  params.addParam<Real>("min_gap",
+                        1e-6,
+                        "The minimum gap distance allowed. This helps with preventing the heat "
+                        "flux from going to infinity as the gap approaches zero.");
   return params;
 }
 
 GapConductanceConstraint::GapConductanceConstraint(const InputParameters & parameters)
-  : ADMortarConstraint(parameters), _k(getParam<Real>("k"))
+  : ADMortarConstraint(parameters), _k(getParam<Real>("k")), _min_gap(getParam<Real>("min_gap"))
 {
 }
 
@@ -39,7 +43,8 @@ GapConductanceConstraint::computeQpResidual(Moose::MortarType mortar_type)
       return -_lambda[_qp] * _test_secondary[_i][_qp];
     case Moose::MortarType::Lower:
     {
-      auto l = (_phys_points_primary[_qp] - _phys_points_secondary[_qp]).norm();
+      auto l = std::max((_phys_points_primary[_qp] - _phys_points_secondary[_qp]) * _normals[_qp],
+                        _min_gap);
       return (_lambda[_qp] - _k * (_u_primary[_qp] - _u_secondary[_qp]) / l) * _test[_i][_qp];
     }
     default:
