+++ modules/navier_stokes/doc/content/modules/navier_stokes/pinsfv.md
@@ -1,0 +1,144 @@
+# Finite Volume Incompressible Porous media Navier Stokes
+
+## Equations
+
+This module implements the porous media Navier Stokes equations. They are expressed in terms of the superficial
+viscosity $\vec{v}_d = \eps \vec{V}$ where $\eps$ is the porosity and $\vec{V}$ the interstitial velocity. The
+superficial viscosity is also known as the extrinsic or Darcy velocity. The other non-linear variables used are
+pressure and temperature. This is known as the primitive superficial set of variables.
+
+Mass equation:
+\begin{equation}
+\div (\rho \vec{v}_d = 0)
+\end{equation}
+
+Momentum equation, with friction and gravity force for example:
+\begin{equation}
+\dfrac{\partial \rho \mathbf{v}_D}{\partial t} + \nabla \cdot (\dfrac{\rho}{\epsilon} \mathbf{v}_D \otimes \mathbf{v}_D) &= \nabla \cdot (\mu \nabla \dfrac{\mathbf{v}_D}{\epsilon}) - \epsilon \nabla p + \epsilon (\mathbf{F}_g + \mathbf{F}_f)
+\end{equation}
+
+Fluid phase energy equation, with a convective heat transfer term:
+\begin{equation}
+\dfrac{\partial \epsilon \rho c_p T_f}{\partial t} + \nabla \cdot (\dfrac{\rho}{\epsilon} \mathbf{v}_D \rho c_{pf} T_f) &= \nabla \cdot (\kappa_f \nabla T_f) - \epsilon \alpha (T_f - T_s)
+\end{equation}
+
+Solid phase energy equation, with convective heat transfer and an energy source $\dot{Q}$:
+\begin{equation}
+\dfrac{\partial (1-\epsilon) \rho c_{ps} T_s}{\partial t} &= \nabla \cdot (\kappa_s \nabla T_s) + (1-\epsilon) \alpha (T_f - T_s) + (1-\epsilon) \dot{Q}
+\end{equation}
+
+where $\rho$ is the fluid density, $\mu$ the viscosity, $c_p$ the specific heat capacity $\alpha$ the convective heat transfer coefficient.
+
+## Implementation for a straight channel
+
+We define a straight channel using a simple Cartesian mesh.
+
+!listing modules/navier_stokes/test/tests/finite_volume/pins/channel-flow/2d-rc.i block=Mesh
+
+The non linear variables are specified among the `INSFV` and `PINSFV` sets of variables as their evaluation
+and the computation of their gradients on faces is specific to the treatment of the Navier Stokes equations.
+
+!listing modules/navier_stokes/test/tests/finite_volume/pins/channel-flow/2d-rc.i block=FVKernels/Variables
+
+The porous media Navier Stokes equation are implemented by using `FVKernels` which correspond to
+each term of the equations. In the following example, the first kernel is the mass advection kernel.
+This kernel corresponds to the conservation of mass. It acts on the pressure non-linear variable which
+appears in the mass equation through the Rhie Chow interpolation of the mass fluxes on the element faces.
+
+!listing modules/navier_stokes/test/tests/finite_volume/pins/channel-flow/2d-rc.i block=FVKernels/mass
+
+Then we add the kernels acting on the 'x' component of the momentum, so essentially transcribing the first
+component of the momentum equation into MOOSE.
+
+The momentum advection term is defined by a `PINSFVMomentumAdvection` kernel. Here the porosity is constant,
+so the porosity gradient term, defined by a `PINSFVMomentumAdvectionPorosityGradient` kernel is not included.
+The momentum equation requires a velocity and an advected quantity interpolation method. This is because this
[…]
