+++ modules/navier_stokes/doc/content/source/fv/kernels/PINSFVHeatConvection.md
@@ -1,11 +1,0 @@
-# PINSFVHeatConvection
-
-This object adds the convective heat transfer term of the
-porous media incompressible Navier Stokes energy equation.
-It should be used for both the fluid and solid temperature variables.
-
-!syntax parameters /FVKernels/PINSFVHeatConvection
-
-!syntax inputs /FVKernels/PINSFVHeatConvection
-
-!syntax children /FVKernels/PINSFVHeatConvection
