+++ modules/navier_stokes/doc/content/source/fv/kernels/PINSFVHeatDiffusion.md
@@ -1,11 +1,0 @@
-# PINSFVHeatDiffusion
-
-This object adds the diffusive heat transfer term of the
-porous media incompressible Navier Stokes energy equation.
-It should be used for both the fluid and solid temperature variables.
-
-!syntax parameters /FVKernels/PINSFVHeatDiffusion
-
-!syntax inputs /FVKernels/PINSFVHeatDiffusion
-
-!syntax children /FVKernels/PINSFVHeatDiffusion
