+++ modules/navier_stokes/doc/content/source/fvkernels/PINSFVMomentumBoussinesq.md
@@ -1,8 +1,8 @@
 # PINSFVMomentumBoussinesq
 
 This object adds a $\epsilon\rho\alpha\vec{g}(T - T_{ref})$ term to the incompressible porous media
-Navier Stokes (PINS) equations where $\epsilon$ is the porosity, $\rho$ is the density, $\alpha$ is the thermal
-expansion coefficient, $\vec{g}$ is the gravity vector, $T$ is the temperature,
+Navier Stokes (PINS) momentum equations where $\epsilon$ is the porosity, $\rho$ is the density, $\alpha$
+is the thermal expansion coefficient, $\vec{g}$ is the gravity vector, $T$ is the temperature,
 and $T_{ref}$ is a reference temperature. The term above introduces the
 Boussinesq approximation into the PINS equations, which allows for modeling
 natural convection and the effect of buoyancy.
