+++ modules/navier_stokes/doc/content/source/fvkernels/PINSFVEnergyAdvection.md
@@ -4,7 +4,7 @@ This object adds a $\rho \epsilon c_p \vec u \cdot \nabla T$ term to a finite vo
 formulation of a heat transport equation. The user can control what (material) quantity is
 advected through the `advected_quantity` parameter. The default value is the
 name `rho_cp_temp` which corresponds to a material property name declared by
-[PINSFVMaterial.md].
+[INSFVMaterial.md].
 
 !syntax parameters /FVKernels/PINSFVEnergyAdvection
 
