### 2021-03-08 e423627
fix channel count calcs
