### 2021-03-08 cc8920e
Add test for target_boundary for CONSTANT monomial. Refs #17214
