+++ modules/navier_stokes/doc/content/source/variables/PINSFVVelocityVariable.md
@@ -1,23 +1,0 @@
-# PINSFVVelocityVariable
-
-!alert construction title=Undocumented Class
-The PINSFVVelocityVariable has not been documented. The content listed below should be used as a starting point for
-documenting the class, which includes the typical automatic documentation associated with a
-MooseObject; however, what is contained is ultimately determined by what is necessary to make the
-documentation clear for users.
-
-!syntax description /Adaptivity/Markers/PINSFVVelocityVariable
-
-## Overview
-
-!! Replace these lines with information regarding the PINSFVVelocityVariable object.
-
-## Example Input File Syntax
-
-!! Describe and include an example of how to use the PINSFVVelocityVariable object.
-
-!syntax parameters /Adaptivity/Markers/PINSFVVelocityVariable
-
-!syntax inputs /Adaptivity/Markers/PINSFVVelocityVariable
-
-!syntax children /Adaptivity/Markers/PINSFVVelocityVariable
