+++ framework/doc/content/source/mesh/GeneratedMesh.md
@@ -16,7 +16,7 @@ independently.
 
 ## Example Syntax
 
-!listing moose/test/tests/kernels/simple_diffusion/simple_diffusion.i
+!listing test/tests/kernels/simple_diffusion/simple_diffusion.i
          block=Mesh
 
 !syntax parameters /Mesh/GeneratedMesh
