### 2021-03-09 f4cd919
Protect against deriv insertion for node without active displacements
