+++ framework/src/base/Assembly.C
@@ -922,9 +922,7 @@ Assembly::computeAffineMapAD(const Elem * elem,
 {
   computeSinglePointMapAD(elem, qw, 0, fe);
 
-#ifdef MOOSE_GLOBAL_AD_INDEXING
   const auto sys_num = _sys.number();
-#endif
 
   for (unsigned int p = 1; p < n_qp; p++)
   {
@@ -944,14 +942,15 @@ Assembly::computeAffineMapAD(const Elem * elem,
         unsigned dimension = 0;
         if (_computing_jacobian)
           for (const auto & disp_num : _displacements)
-            Moose::derivInsert(elem_point(dimension++).derivatives(),
+            if (node.n_dofs(sys_num, disp_num))
+              Moose::derivInsert(elem_point(dimension++).derivatives(),
 #ifdef MOOSE_GLOBAL_AD_INDEXING
-                               node.dof_number(sys_num, disp_num, 0)
+                                 node.dof_number(sys_num, disp_num, 0)
 #else
-                               disp_num * _sys.getMaxVarNDofsPerElem() + i
+                                 disp_num * _sys.getMaxVarNDofsPerElem() + i
 #endif
-                                   ,
-                               1.);
+                                     ,
+                                 1.);
 
         _ad_q_points[p].add_scaled(elem_point, phi_map[i][p]);
       }
@@ -1034,9 +1033,7 @@ Assembly::computeSinglePointMapAD(const Elem * elem,
   const auto & dphidxi_map = fe->get_fe_map().get_dphidxi_map();
   const auto & dphideta_map = fe->get_fe_map().get_dphideta_map();
   const auto & dphidzeta_map = fe->get_fe_map().get_dphidzeta_map();
-#ifdef MOOSE_GLOBAL_AD_INDEXING
   const auto sys_num = _sys.number();
-#endif
 
   switch (dim)
   {
@@ -1064,14 +1061,15 @@ Assembly::computeSinglePointMapAD(const Elem * elem,
         unsigned dimension = 0;
         if (_computing_jacobian)
           for (const auto & disp_num : _displacements)
-            Moose::derivInsert(elem_point(dimension++).derivatives(),
+            if (node.n_dofs(sys_num, disp_num))
+              Moose::derivInsert(elem_point(dimension++).derivatives(),
 #ifdef MOOSE_GLOBAL_AD_INDEXING
-                               node.dof_number(sys_num, disp_num, 0)
+                                 node.dof_number(sys_num, disp_num, 0)
 #else
-                               disp_num * _sys.getMaxVarNDofsPerElem() + i
+                                 disp_num * _sys.getMaxVarNDofsPerElem() + i
 #endif
-                                   ,
-                               1.);
+                                     ,
+                                 1.);
 
         _ad_dxyzdxi_map[p].add_scaled(elem_point, dphidxi_map[i][p]);
 
