### 2021-04-18 #3031 @WilkAndy
Keeping this JOSS paper on @jedbrown 's radar.
