### 2021-05-26 #3314 @whedon
[REVIEW]: The MOOSE geochemistry module
**Submitting author:** @WilkAndy (<a href="http://orcid.org/0000-0001-6472-9560">Andy Wilkins</a>)
**Repository:** <a href="https://github.com/idaholab/moose" target ="_blank">https://github.com/idaholab/moose</a>
**Version:** Snapshop 21-November-14
