### 2021-06-17 #3314 @jedbrown
@volpatto :wave: Welcome to JOSS and thanks for agreeing to review!

The comments from @whedon up top :point_up: outline the review process, which takes place in this thread (possibly with issues filed in the rsudp repository). I'll be watching this thread if you have any questions.

We aim for reviews to be completed within a month or so. Please let me know if you require some more time. We can also use Whedon (our bot) to set automatic reminders if you know you'll be away for a known period of time.
