### 2021-08-09 #3314 @volpatto
Firstly, I would like to say that MOOSE's geochemistry module is quite interesting and provides an easy way to simulate coupled phenomena with geochemical reactions. For the authors, my congratulations.

Following the JOSS Reviewer's guideline, I will draw some comments below:

## General checks

> Repository: Is the source code for this software available at the repository URL?

Not clear to me if it should be the MOOSE's URL, despite it makes sense since the piece of software in this submission is a MOOSE's module. Very minor suggestion, but I think it would be more appropriate to point to the [geochemistry module's URL](https://github.com/idaholab/moose/tree/next/modules/geochemistry), where the `paper.md` is hosted.

## Software paper

Well written and clear.

> References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?

The .bib file is there, but references are not rendered in `paper.md` (only hyperlinks), and the bibliography file is not configured in the [References](https://github.com/idaholab/moose/blob/next/modules/geochemistry/paper.md#references). After fixing that, I think it would be OK after generating the PDF file with `pandoc`.
