### 2021-08-09 #3314 @jedbrown
Thanks, @volpatto. The repository URL is meant to be clonable, which a sub-tree is not. We understand that some projects take a monorepo approach while others use separate repositories, and JOSS is okay working with both so long as the scope of the module meant for review is clear.

And regarding references, those are automatically picked up by Whedon, as you saw in the proof.
