### 2021-08-10 #3314 @smolins
I started with the software paper. There are minor and major issues noted below (I refer to the lines in the pdf) :

line 13: flow --> transport

line 17: it is not clear just by the abstract how chemistry deforms the subsurface. It is a confusing sentence overall

line 35: complicated --> complex?

Section Existing software:
This list is missing probably the most active open-source RT code around, PFLOTRAN
Also, I would strongly suggest using this reference for CrunchFlow: (Steefel et al 2015)
https://link.springer.com/article/10.1007/s10596-014-9443-x
That's what Carl Steefel is using.
Generally, this section should make a distinction between codes with their own geochemical capabilities (like MOOSE) and the ones that tap into external geochemical capabilities (unlike MOOSE). This is the theme of this paper so it must be there. Please read the intro to section "4 Code descriptions" of the reference above (Steefel et al 2015) to see what I mean. MOOSE is not the first to offer native geochemical capabilities. Also, there are RT codes that were born out of geochemical codes, while there are RT codes that were first Flow and Transport codes and then added geochemical capabilities. So lots to discuss here.

line 68: beautifully-designed --> this is a subjective statement. Remove, replace

lines 79-82: This is a very general statement that is not justified. It goes with my comment for the "Existing software" section. There are many codes around and comparisons are hard to make without going into detail. It is in unfair to make such general statements without discussion or appropriate references.

line 87: it's ---> its

line 104: so mineral equilibrium is enforced, and the precipitation/dissolution rate calculated after that from mass balance? Is this is a time-dependent calculation?

Section "Reactive Transport":
In this section it is not clear whether one can or cannot use a global implicit approach to solve reactive transport with MOOSE. It is recommended to use operator splitting but what if I want to use global implicit. This is actually important. OS uses a linear transport solver while GI requires a non-linear solver for reactive transport. It appears that only the non-linear Newton solution approach for the geochemical problem is described in the Mathematical solution strategy section, not the global implicit.
It is also not clear whether the canonical approach, where the number of species to be transported is reduced using the equilibrium assumption (see Steefel et al 2015), is part of MOOSE or not. Does transport solve for primary aqueous species concentration or does it solve for total component concentrations  Or is it flexible?

line 201: a rudimentary

line 211: is only the finite-element-method supported?

Figure 1: Is that obtained with an operator splitting approach? What are the time step contraints here?
