### 2021-08-10 #3314 @smolins
Could you please help me to find the instructions to build the code to run the examples in the geochemical module included in the paper? What MOOSE component do I need? One is a batch calculation and the other is a reactive transport run, so I am assuming different modules from MOOSE are needed but I can't find the description anywhere.
