### 2021-08-10 #3314 @jedbrown
Thanks, @smolins.

@WilkAndy Please advise on the question about running and revise/respond to the other review points. In case the project documentation is unclear about how new users (who may not be experienced MOOSE users) can start running demos, it would be appropriate to revise that documentation so that it benefits other users, not just @smolins here. Thanks.
