### 2021-08-16 #3314 @jedbrown
You can use the syntax `@whedon generate pdf from branch joss-paper-edits` to get an updated draft for review before the branch merges.
