### 2021-08-22 #3314 @smolins
Both examples in the manuscript seem to run. There are 2 issues

1.When I plot the results of the reactive transport example, the mesh (and resolution) seems to be different from the figure in the manuscript. Is that expected?

2. Why were these two examples highlighted in the manuscript? The code has many documented examples (https://mooseframework.inl.gov/modules/geochemistry/tests_and_examples/index.html). Are these specially interesting?

Overall, the main issue with the examples is the reliance almost exclusively on 1 source (Bethke 2007) and comparison to a single proprietary code, GWB. The examples are generally described sufficiently in the documentation but I feel that I would
have benefited from having the book and GWB to fully follow everything.
