### 2021-08-24 idaholab/moose#18610 @WilkAndy
Also, need to address the following comments (from https://github.com/openjournals/joss-reviews/issues/3314#issuecomment-903193062)

1. When I plot the results of the reactive transport example, the mesh (and resolution) seems to be different from the figure in the manuscript. Is that expected?

2. Why were these two examples highlighted in the manuscript? The code has many documented examples (https://mooseframework.inl.gov/modules/geochemistry/tests_and_examples/index.html). Are these specially interesting?

3. Overall, the main issue with the examples is the reliance almost exclusively on 1 source (Bethke 2007) and comparison to a single proprietary code, GWB. The examples are generally described sufficiently in the documentation but I feel that I would have benefited from having the book and GWB to fully follow everything.

I plan to add the following:

```
# Tests and examples

The geochemistry module's [code coverage](https://mooseframework.inl.gov/docs/coverage/geochemistry/index.html) currently exceeds 96%.  Many of the more complicated benchmark tests and examples are [comprehensively documented](https://mooseframework.inl.gov/modules/geochemistry/tests_and_examples/index.html) and the sections below present one benchmark study and one example by way of illustration.
```

```
(These results are derived using a higher resolution than the example included in the test suite, since the latter must be small enough to rapidly pass MOOSE's automatic regression testing.)
```
