### 2021-08-24 cd377d5
build and usage instructions added to paper.md.  Also comments about test coverage, other tests and examples, and the mesh resolution in Weber-Tensleep example.  Refs #18610
