### 2021-08-24 #3314 @WilkAndy
Thank you @smolins .   I have addressed your comments immediately above in https://github.com/idaholab/moose/pull/18689 .  I have **not** yet addressed your comments of 14 days ago, but will do soon.   If you like, inspect the diffs at that PullRequest.  Or, you can generate a new PDF by going to https://whedon.theoj.org/ and using repo `https://github.com/WilkAndy/moose` and branch ` joss_geochem_18610`.   Or, you can wait for a while until the PullRequest makes it through to the main branch of MOOSE.

1. You are correct that the mesh resolution is different from the figure.  I'm going to add the following remark:

> These results are derived using a higher resolution than the example included in the test suite, since the latter must be small enough to rapidly pass MOOSE's automatic regression testing.

2. You are correct that there are other examples I could have included.  To explain, I'm going to add:

# Tests and examples

The geochemistry module's [code coverage](https://mooseframework.inl.gov/docs/coverage/geochemistry/index.html) currently exceeds 96%.  Many of the more complicated benchmark tests and examples are [comprehensively documented](https://mooseframework.inl.gov/modules/geochemistry/tests_and_examples/index.html) and the sections below present one benchmark study and one example by way of illustration.

3. Yes, we did a lot of comparisons against GWB.  I'm not sure whether it's appropriate to explain this in the paper, but the reason we did a lot of GWB comparisons is that it's considered a "gold standard", and it has excellent documentation in the form of Bethke's book, and we believe that many potential users would be familiar with GWB, so hopefully they'd find the transition to MOOSE easier.  I totally agree that not having Bethke's book makes reading the current paper and the documentation more difficult, but I'm also aware that we don't want to reproduce a lot of theory and explanation in this hopefully-brief paper.
