### 2021-08-25 #3314 @WilkAndy
Hi @smolins .  I believe that i've addressed all your concerns - the software ones from 3 days ago, and the article ones from 14 days ago.   The changes may be seen at https://github.com/idaholab/moose/pull/18689/files  and a new PDF may be made by going to https://whedon.theoj.org/ and using repo https://github.com/WilkAndy/moose and branch joss_geochem_18610 .
