### 2021-08-28 #3314 @smolins
>  **Please start on your review when you are able, and be sure to complete your review in the next six weeks, at the very latest**
>
> ## Review checklist for @smolins
> ### Conflict of interest
> * [x]  I confirm that I have read the [JOSS conflict of interest (COI) policy](https://github.com/openjournals/joss/blob/master/COI.md) and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.
>
> ### Code of Conduct
> * [x]  I confirm that I read and will adhere to the [JOSS code of conduct](https://joss.theoj.org/about#code_of_conduct).
>
> ### General checks
> * [x]  **Repository:** Is the source code for this software available at the [repository url](https://github.com/idaholab/moose)?
> * [x]  **License:** Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?
> * [x]  **Contribution and authorship:** Has the submitting author (@WilkAndy) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
> * [ ]  **Substantial scholarly effort:** Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)
>
> ### Functionality
> * [x]  **Installation:** Does installation proceed as outlined in the documentation?
> * [x]  **Functionality:** Have the functional claims of the software been confirmed?
> * [x]  **Performance:** If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
>
> ### Documentation
> * [x]  **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?
> * [x]  **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
> * [x]  **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
> * [x]  **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
> * [x]  **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?
> * [x]  **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
>
> ### Software paper
> * [x]  **Summary:** Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
> * [x]  **A statement of need:** Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?
> * [ ]  **State of the field:** Do the authors describe how this software compares to other commonly-used packages?
> * [ ]  **Quality of writing:** Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
> * [x]  **References:** Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax)?

As noted above the submission meets most of the criteria above. My outstanding issues are

1. relating to the "Substantial scholarly effort". From manuscript:

33 In the past, researchers have typically answered such questions using separate models for
34 subsets of the physical processes, or loosely-coupled models solved using multiple software
35 packages. The MOOSE geochemistry module introduced here allows researchers to perform
36 stand-alone geochemical modelling, but also to draw upon the power of other MOOSE modules
37 to solve complicated coupled transport, geomechanical and geochemical models using a single
38 code

While I see that integrating all these capabilities in the Moose framework is new. It is not new in the reactive transport community. This needs to be explicitly stated. Plus, the example presented shows some of these coupled capabilities for the resulting code but it is not clear whether these are unique to moose, or which ones are unique to moose.

2. I think you should state as well that why you chose to use GWB in the manuscript. It is clearly at odds with the open source angle of this journal and the open source licence for moose. Also, there are community benchmarking efforts for reactive transport that offer solutions that have been obtained and compared by a range of reactive transport codes each

Computational Geosciences Volume 19, issue 3
https://link.springer.com/journal/10596/volumes-and-issues/19-3

Computational Geosciences Volume 25, issue 4,
https://link.springer.com/journal/10596/volumes-and-issues/25-4

3. It is not clear what "nevertheless" means here.

215: Nevertheless, a rudimentary transport capability has been included in the geochemistry module,
216 which models advective-diffusive transport of the mobile concentrations, with hydrodynamic
217 dispersion

I believe that the test problems use this simple transport solver rather that PorousFlow, so you should mention this too.

4. "Reactive transport" section, lines 192, 195: what does the word "usually" mean here? Is it needed? It is not clear what other unusual options are there...
