### 2021-08-29 #3314 @jedbrown
Thanks @smolins. To clarify regarding "substantial scholarly effort", this verbiage exists mainly to screen "small" products that shouldn't be "counted" as paper-like units. JOSS does not have a "novelty" criteria like some journals -- quality software that fills a research need is in-scope even if for each specific problem instance that can be solved with the new software, there is existing software that can also solve it similarly well. It still needs an honest statement of need that does justice to alternatives. For instance, a new package might be preferred because of how it facilitates some type of integration for which there is a research need. That could be reaching a different language/library community, though the software should be well integrated in that environment.

JOSS papers should _not_ claim novel methods or new scientific results -- there are existing venues for that sort of work. (JOSS software, and any sufficiently mature software for that matter, often contains some unpublished work of that sort, it just isn't the scholarship that JOSS intends to evaluate.)

I'll leave it to @WilkAndy to revise and reply to specific points.
