### 2021-08-29 #3314 @smolins
Thanks @jedbrown  for the clarification. I would say then that the paragraph (lines 33-38)

33 In the past, researchers have typically answered such questions using separate models for
34 subsets of the physical processes, or loosely-coupled models solved using multiple software
35 packages. The MOOSE geochemistry module introduced here allows researchers to perform
36 stand-alone geochemical modelling, but also to draw upon the power of other MOOSE modules
37 to solve complicated coupled transport, geomechanical and geochemical models using a single
38 code.

should be more specific to moose. In other words, there are other codes that implement coupled models with their own capabilities. It is moose that did not have geochemistry and that was the need.
