### 2021-09-02 #3314 @WilkAndy
Thanks for more suggestions on how to improve the manuscript, @smolins .   I believe i have addressed all your points.   The changes may be seen at https://github.com/idaholab/moose/pull/18689/files and a new PDF may be made by going to https://whedon.theoj.org/ and using repo https://github.com/WilkAndy/moose and branch joss_geochem_18610 .

- Regarding use of `geochemistry`'s inbuilt transport capability, or `PorousFlow`'s version: about half the tests/examples use the former and half use the latter.  `PorousFlow` is always preferred in real-life situations so is used in the more complicated models.
- Regarding benchmarking against GWB (which I've highlighted in the new manuscript): the reason we benchmarked against GWB was to try to attract the attention of people who tend to dismiss open-source/research codes as impractical/buggy/idealised/unreliable.  Those people may also treat Bethke's (excellent) book as a bible, so we provided references to it throughout the online documentation.  Perhaps you and others wouldn't have taken this approach!
