### 2021-09-06 #3314 @volpatto
@smolins for me, I have clicked on the check-boxes in the first comment. Then the progress status was automatically updated.

@WilkAndy I will continue my review in the next few days. I think @smolins addressed a lot of relevant things already.
