### 2021-09-20 idaholab/moose#18856 @volpatto
[JOSS-Review] Broken contributing link
This issue is part of [JOSS' MOOSE geochemistry module review](https://github.com/openjournals/joss-reviews/issues/3314).

Very minor issue: The link pointing to an article with information on how to contribute to MOOSE with code changes is not working.
Please see the mentioned link [here](https://github.com/idaholab/moose#contributing).

Cheers.
