### 2021-09-20 #3314 @volpatto
> **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

2 and 3 are properly addressed on MOOSE website. However, I did not find that 1 is clear enough. I looked in MOOSE's README.md and find a link that is not working. Thus I opened [an Issue](https://github.com/idaholab/moose/issues/18856) in the MOOSE repo.
