### 2021-09-20 idaholab/moose#18856 @volpatto
[This link](https://mooseframework.inl.gov/framework/contributing.html) is the working one, I guess.
