### 2021-09-20 #3314 @volpatto
## Functionality

> Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

The authors performed (strong) scaling analyses and space complexity estimation. Results (in the [paper](https://github.com/idaholab/moose/blob/next/modules/geochemistry/paper.md#computational-aspects) and in the [MOOSE website](https://mooseframework.inl.gov/modules/geochemistry/theory/compute_efficiencies.html)) are interesting, nearly optimal for scaling. However, I should remark that there are no instructions on reproducibility. Moreover, the case "Reactive-transport using multiple processors" is only possible in HPC environments.

AFAIK, this is not a terrible problem, it is quite common to be unable to reproduce performance analyses, but I have to state that I can't confirm the results running them.
