### 2021-09-21 #3314 @WilkAndy
@volpatto - you'll see that the "README.md" issue you mentioned above has now been fixed.   I'm not sure whether you need me to address your "Functionality - Performance" issue above.  Could you please confirm one way or the other?
