### 2021-09-22 #3314 @volpatto
> @volpatto - you'll see that the "README.md" issue you mentioned above has now been fixed.

Great, thanks.

> I'm not sure whether you need me to address your "Functionality - Performance" issue above. Could you please confirm one way or the other?

Sure, sorry for not being clear. I think there is no need for changes, it was just a comment/feedback. Since measurements are provided, I understand that this is enough.
