### 2021-09-27 #3314 @volpatto
Okay, I think I almost finished my review. I will draw some comments.

## Minor/side comments

* Installations: Following the instructions to install MOOSE and everything would be fine. MOOSE uses `conda` to manage dependencies, which is great.

* I have compiled MOOSE in WSL2 (using Ubuntu 20.04), it worked fine. Compilation may take some time (around 10 min on my PC).

## Comments

### Automated tests

I executed the MOOSE test set and it runs properly on my PC. However, as I am not a user/dev of MOOSE (and MOOSE is huge, the tutorial is huge too, naturally), it was not immediately clear to me how to test the `geochemistry` module alone. It turns out to be very simple, and the `geochemistry` module has 126 tests when running `run_tests` in the geochemistry module's root dir. This is not the number of tests the manuscript claims to be implemented (350 tests), please see [the first paragraph here](https://github.com/idaholab/moose/blob/next/modules/geochemistry/paper.md#overview-and-comparisons-with-other-software). I understand that it probably included some benchmarks, but they are not formally implemented in the test set. Or the geochemistry module has another test script?

### Example cases

I tried to reproduce the [cooling with feldspars](https://github.com/idaholab/moose/blob/next/modules/geochemistry/paper.md#example-cooling-with-feldspars) example. Unfortunately, I had to look in the [test and examples](https://mooseframework.inl.gov/modules/geochemistry/tests_and_examples/index.html) page to find the example. The listed name is "Progressively changing the temperature", which is provided [here](https://mooseframework.inl.gov/modules/geochemistry/tests_and_examples/cooling_feldspar.html). I would like to suggest adding this link and the example name to the manuscript.

In the aforementioned example [docs](https://mooseframework.inl.gov/modules/geochemistry/tests_and_examples/cooling_feldspar.html), I found no direct instructions on how to run the case. I have check other examples, and I didn't find instructions as well. I know that this is a very basic MOOSE usage, but for some user that just has found MOOSE's geochemistry module and wants to give it a try (as standalone), it can discourage the user to try the software. In MOOSE's examples and tutorial, sometimes a ["Running the Problem" section](https://mooseframework.inl.gov/getting_started/examples_and_tutorials/examples/ex03_coupling.html) is provided. I think it would be convenient to have something like that in the `geochemistry` module example, as it offers standalone capabilities. Moreover, the example has a Python script to plot the results.

### Mathematical solution strategy

The geochemistry module online docs are great. Why the [Theory page](https://mooseframework.inl.gov/modules/geochemistry/theory/index.html) is not mentioned in [this section](https://github.com/idaholab/moose/blob/next/modules/geochemistry/paper.md#mathematical-solution-strategy)?

### Coupling

The authors may face some queries about the coupling strategy (with transport solver) when presenting the geochemistry module to the community. [In the manuscript](https://github.com/idaholab/moose/blob/next/modules/geochemistry/paper.md#reactive-transport), it is suggested to use a Sequential Non-Iterative Approach (SNIA). Although great for implementation ease and computational performance, this approach has limitations. Remarkably, the time-step size can affect (and violate) the chemical species' mass conservation due to the delay in the spatial operator discretization. The manuscript briefly mentions that different types of coupling are available. I would like to suggest mentioning these couplings in terms of Steefel and co-workers' (see references below) nomenclatures (GIA, SIA, and SNIA). I understand that every approach is possible using MOOSE, is that correct?

1. [C. Steefel and K. MacQuarrie, “Approaches to modeling of reactive transport in porous media, ”Reviews in Mineralogy and Geochemistry, vol. 34, pp. 85–129, 1996.](https://doi.org/10.1515/9781501509797-005)
2. [C. Steefel, C. Appelo, B. Arora, D. Jacques, T. Kalbacher, O. Kolditz, V. Lagneau, P. Lichtner, K. U. Mayer, J. Meeussen, et al., “Reactive transport codes for subsurface environmental simulation, ”Computational Geosciences, vol. 19, no. 3, pp. 445–478, 2015.](https://doi.org/10.1007/s10596-014-9443-x)
