### 2021-09-27 #3314 @volpatto
Since the geochemistry module has "standalone" capabilities, I think it would be interesting for new users the instructions cited below. Maybe it could be documented in the online documentations?
>
>
> Thanks @smolins . I need to check that the following instructions actually work before including them in the paper, but here's something to get you started
>
> To build `geochemistry` and run the examples:
>
>     1. Download and install the entire MOOSE package.  Instructions are at https://mooseframework.inl.gov/getting_started/installation/index.html .  Even if you're interested only in geochemistry (without transport, solid mechanics, etc), MOOSE comes as a complete package, so needs to be installed in its entirety.  Depending on your computer setup, this can be straightfoward (on a personal Mac computer) or complicated (on an administered supercomputer).
>
>     2. During the installation of MOOSE in part 1, only the "framework" will have been compiled.  To compile all the physics modules, including the `geochemistry` module, use the following instructions run from the command line:
>
>
> ```
> cd ~/projects/moose/modules
> make -j 4
> cd ~/projects/moose/modules/geochemistry
> make -j 4
> ```
>
>     1. Check that the geochemistry module is correctly compiled using the following instructions:
>
>
> ```
> cd ~/projects/moose/modules/geochemistry
> ./run_tests
> ```
>
> Virtually all the tests should run and pass. Some may be "skipped" due to your computer setup (for instance, not enough threads).
> 4. Now you're ready to use the geochemistry module. For example, to run the Weber-Tensleep GeoTES example from the command line:
>
> ```
> cd ~/projects/moose/modules/combined/examples/geochem-porous_flow/geotes_weber_tensleep
> ../../../combined-opt -i exchanger.i
> ```
