### 2021-10-07 #3314 @WilkAndy
Hi @volpatto , The revised article may be seen by going to https://whedon.theoj.org/ and using repo https://github.com/WilkAndy/moose and branch joss_volpatto_18610 .  The revisions will take a while to merge into MOOSE: the PR is at https://github.com/idaholab/moose/pull/19030 .   To answer all your points above:

1. _"... the geochemistry module has 126 tests when running run_tests in the geochemistry module's root dir. This is not the number of tests the manuscript claims to be implemented (350 tests)..."_   I apologize.  There are also 252 unit tests.  These may be run using
```
cd ~/projects/moose/modules/geochemistry/unit
make
./run_tests
```
I have added those instructions to the JOSS article.

2. _"...I would like to suggest adding this link and the example name to the manuscript..."_ (cooling feldspars example) I've added the appropriate link.
3.  _"... I found no direct instructions on how to run the case..."_   The JOSS article does contain the paragraph:
The geochemistry executable is called `geochemistry-opt` and is found at `~/projects/moose/modules/geochemistry`.  This may be used to run pure geochemistry simulations.  For coupled reactive-transport simulations using the PorousFlow module, the `combined-opt` executable must be used.  For example, to run the Weber-Tensleep GeoTES example from the command line:
```
cd ~/projects/moose/modules/combined/examples/geochem-porous_flow/geotes_weber_tensleep
../../../combined-opt -i exchanger.i
```
But I agree that the "standalone" capabilities of the module mean it should have standalone instructions.  I have expanded the instructions in the article, and also added to the online MOOSE documentation.

4. _"...Why the Theory page is not mentioned in this section?..."  I've added the link to the article.
5. Coupling strategies: I've now mentioned SIA and SNIA explicitly, and that users have to be careful with timestep sizes, and i've cited the Steefel2015 article.
