### 2021-11-12 #3314 @kyleniemeyer
It looks like this submission is in good shape to proceed, so I'm going to step in and get this past the finish line.

@WilkAndy thanks for addressing the reviewer comments. At this stage, could you now archive the software repository (for example, on Zenodo), make sure the title and authors list match the paper, and let us know the DOI here?

Also, I noticed a minor typo in the first paragraph in "Statement of Need": a missing comma after "e.g.".
