### 2021-11-15 #3314 @WilkAndy
@kyleniemeyer

- i have published on zonodo: https://doi.org/10.5281/zenodo.5701422
- i have added the comma that you mentioned, and the change has been merged into the "next" branch.  However, it will take 1-2 days to merge into "master".
