### 2021-11-16 dda0653
geochemistry: copy edit and PETSc website
The petsc-user-ref is somewhat old and perhaps should also be updated.

Closes #19405
