### 2021-11-16 #3314 @jedbrown
Thanks for your patience, @WilkAndy, and thanks to @kyleniemeyer for backstopping. Please consider the PR (https://github.com/idaholab/moose/pull/19404) with minor copy edits and fixing the duplicate link for the PETSc website (also updated to the new URL). We can proceed as soon as that is in a branch (any branch) of the main repo.
