### 2021-11-16 idaholab/moose#19405 @WilkAndy
geochemistry JOSS paper edits
## Reason

"Typo" edits in the geochemistry JOSS paper

## Design

See https://github.com/openjournals/joss-reviews/issues/3314

## Impact

Ability to publish paper
