### 2021-11-29 175b554
geochemistry: prefer doi in paper.bib
Fixes #19499
