+++ modules/geochemistry/paper.bib
@@ -261,7 +261,7 @@ title={{A compilation of rate parameters of water-mineral interaction kinetics f
 institution={U.S. Geological Survey},
 number={OF 2004-1068},
 year={2004},
-url={https://pubs.usgs.gov/of/2004/1068/}
+doi={10.3133/ofr20041068}
 }
 
 @book{prausnitz,
