### 2021-11-29 #3314 @jedbrown
Returning from the holiday, I'm sorry I missed this the first time around. This fixes the missing DOI. I guess it needs an issue and a PR referencing that issue.
https://github.com/jedbrown/moose/commit/836b5f6867dbba0d897ddf403b09de9a1207260c
