### 2021-11-29 idaholab/moose#19499 @WilkAndy
add DOI to JOSS geochemistry paper
## Reason

As in https://github.com/openjournals/joss-reviews/issues/3314#issuecomment-981329120 , a DOI needs to be added to geochemistry/paper.bib

## Design

This: https://github.com/jedbrown/moose/commit/836b5f6867dbba0d897ddf403b09de9a1207260c

## Impact

Ability to publish the paper!
